% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{add_cpts}
\alias{add_cpts}
\title{Add center points to an experimental design.}
\usage{
add_cpts(input, cpts = 3)
}
\arguments{
\item{input}{A \code{data.frame} containing the \strong{coded} input factors and their levels. Columns (factor names) must be named alphabetically.}

\item{cpts}{The number of center points to add (defaults to 3).}
}
\value{
The experimental design \code{data.frame} including the newly added center points.
}
\description{
Part of the \strong{experimental design generation} workflow.
}
\examples{
#Load data
doe_base <- load_internal("doe_base")

#Get experimental design data.frame
doe_df <- doe_base[["FD_3"]]

#Add center points
res <- add_cpts(doe_df, 3)

}
\seealso{
\code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
