% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{add_stars}
\alias{add_stars}
\title{Add star points to an experimental design}
\usage{
add_stars(input, factors, design = "CCC")
}
\arguments{
\item{input}{A \code{data.frame} containing the \strong{coded} input factors and their levels. Columns (factor names) must be named alphabetically.}

\item{factors}{The number of factors (\strong{between 2 and 4}).}

\item{design}{The type of Central Composite experimental design.
Currently supported designs are Circumscribed (\code{"CCC"}, default) or face-centered (\code{"CCF"}).}
}
\value{
The experimental design \code{data.frame} including the newly added star points.
}
\description{
Part of the \strong{experimental design generation} workflow.
}
\examples{
#Load data
doe_base <- load_internal("doe_base")

#Get experimental design data.frame
doe_df <- doe_base[["FD_3"]]

#Add center points
res <- add_stars(doe_df, 3, "CCC")

}
\seealso{
\code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
