% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{aug_sum}
\alias{aug_sum}
\title{Generate an augmented summary of linear/quadratic models}
\usage{
aug_sum(linmod, thres_coeff = 0.1, codenames = NA)
}
\arguments{
\item{linmod}{A linear model of class \code{lm}.}

\item{thres_coeff}{The fraction (between 0.01 and 1) of maximum eigenvalue to use for canonical analysis (see \code{\link{canon}}).}

\item{codenames}{An \strong{optional} \code{data.frame} containing disambiguation of factor names.
Generated from function \code{\link{twoway}}.}
}
\value{
The augmented summary of the input model \code{linmod} (class \code{"summary.lm"}).
}
\description{
Returns an augmented summary of a linear or quadratic model (i.e. model of order 1, 1.5, or 2).
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
