% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{bendens}
\alias{bendens}
\title{Calculate CO2 density using Bender EoS}
\usage{
bendens(pres, temp, units = "g/L")
}
\arguments{
\item{pres, temp}{Pressure (in bar) and temperature (in degrees Celsius) for which to return the CO2 density.}

\item{units}{Units required for output density as a \code{character} value. One of: \code{"g/L"} or \code{"g/mL"}.}
}
\value{
A named \code{numeric} vector containing the density (\code{["rho"]}) in chosen \code{units} and the specific enthalpy (\code{"ent"}).
}
\description{
Uses the Bender Equation of State to calculate CO2 density (g/L or g/mL) and specific enthalpy (kJ/kg)
between 1-1000 bar and -50 to 300 degress Celsius.
Part of the \code{bicmod} workflow.
}
\details{
For the complete Bender Equation of State used to calculate density and enthalpy herein, see \strong{Eqs. 1.1-1.4} in Rizza (2014).
}
\examples{
bendens(300, 45)
bendens(800, 70, "g/mL")

}
\references{
Bender, E. (1975), 'Equations of state for ethylene and propylene', \emph{Cryogenics} \strong{15} (11), pp. 667-673, DOI: \url{https://doi.org/10.1016/0011-2275(75)90100-9}.

Ghazouani, J., Chouaieb, O., Bellagi, A. (2005), 'Evaluation of the parameters of the Bender equation of state for low acentric factor fluids and carbon dioxide', \emph{Thermochimica Acta} \strong{432} (1), pp. 10-19, DOI: \url{https://doi.org/10.1016/j.tca.2004.11.008}.

Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sievers, U. (1984), 'Die thermodynamischen Eigenschaften von Kohlendioxid', \emph{Forsch Ing-Wes} \strong{50}, p. 192, DOI: \url{https://doi.org/10.1007/BF02560841}.
}
\seealso{
\code{\link{bicmod}}
}
