% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{bic_ct}
\alias{bic_ct}
\title{Derive the simplified BIC model based on characteristic times}
\usage{
bic_ct(
  yield,
  t,
  n,
  qaver,
  cu,
  N,
  Nm,
  ys,
  modpts = 100,
  thetaf0 = 1,
  ti0 = 30,
  maxt = round(max(t) * 1.2, -1)
)
}
\arguments{
\item{yield}{Fractional yield of extract (g/g total dry solid).}

\item{t}{A \code{numeric} vector of times at which \code{yield} was recorded (s).}

\item{n}{Period corresponding to the end of the CER (in \strong{number of experimental points}).}

\item{qaver}{Average solvent flow rate (kg/s).}

\item{cu}{Solute content in the untreated solid (kg/kg), i.e. maximum possible yield as a \strong{fraction}.}

\item{N}{Total dry mass (\strong{kg}).}

\item{Nm}{Mass of insoluble material (g).}

\item{ys}{Extract solubility in CO2 (g/g).}

\item{modpts}{A single \code{numeric} value specifying how many points should be modelled throughout the OEC curve range (\code{100} by default).}

\item{thetaf0}{Initial estimate of the external mass transfer resistance. Defaults to \code{1}.}

\item{ti0}{Initial estimate of the characteristic time of the solid phase mass transfer (s). Defaults to \code{30} s.}

\item{maxt}{Maximum x-axis value (time, s) to use for model predictions.
Defaults to 120\% of the maximum \code{t} value.}
}
\value{
A named \code{list} containing the following elements:
\enumerate{
\item \strong{$ordt}: A \code{data.frame} containing the original input data including the time variable \code{$x},
the \strong{actual} fractional yield (\code{$y}; g/g \strong{total dry solid}), as well as the \strong{modeled} yield (\code{$mod_y}.
\item \strong{$mdt}: A \code{data.frame} of modeled data including the model type (\code{$model}),
the region of the extraction curve (\code{$period}), as well as the time (\code{$t}) up to \code{maxt} and fractional yield \code{$y}.
\item \strong{$mod_pars}: The model parameters including the external mass transfer resistance (\eqn{\theta_e}; \code{"thetaf"}),
extraction time/duration of the FER (\eqn{t_i}; \code{"ti"}), extraction time (\eqn{t'}) and yield (\eqn{e'}; g/g total dry solid)
at the end of CER (\code{"tprime"} and \eqn{"eprime"}), and the initial fraction of solute in broken cells (\eqn{G}; \code{"G"}).
(\eqn{k_sa_s}; \code{"ksas"}). See \code{\link{show_pars}} for detailed description of these parameters.
\item \strong{$fit_pars}: A \code{character} vector of which parameters in \code{$mod_pars} were iteratively fit to the model.
\item \strong{$resid}: A named \code{numeric} vector of various error values for the model, including Average Absolute Relative Deviation
AARD (\code{"aard"}), round mean squared error RMSE (\code{"rmse"}), and the R2 value (\code{"r2"}).
}
}
\description{
Derives, summarises, and visualises the results of the characteristic times Broken-and-Intact Cells (BIC)
model of Sovova (2005).
Part of the \code{\link{bicmod}} workflow.
}
\details{
For a detailed description of this variant of the Broken and Intact Cells (BIC) model, refer to \code{\link{bicmod}}.
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}, \code{\link{moderr}}
}
