% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{bic_sm}
\alias{bic_sm}
\title{Build a simplified BIC model}
\usage{
bic_sm(
  yield,
  q,
  n,
  xu = NA,
  qaver,
  Nm = NA,
  porosity,
  ys = NA,
  c3 = NA,
  Ng = NA,
  est_cu = FALSE,
  maxq = round(max(q) * 1.2, -1),
  r0 = 0.4,
  ksas0 = 1e-04,
  qc0 = 400,
  aggreg = "aard",
  modpts = 100,
  nlsm = c("MM", "tau", "CM", "mtl"),
  const_flow = TRUE
)
}
\arguments{
\item{yield}{A \code{numeric} vector of yields (\eqn{e}; g/g insoluble solid).}

\item{q}{A \code{numeric} vector of relative amount of expended solvent (\eqn{q}; kg/kg insoluble solid).}

\item{n}{Period corresponding to the end of the CER (\eqn{n}; in \strong{number of experimental points}).}

\item{xu}{Weight fraction (concentration) in the untreated solid \eqn{x_u}. Defaults to \code{NA}.
Must be provided only when \code{est_cu} is \code{FALSE}.}

\item{qaver}{Average solvent flow rate (\eqn{\dot{Q}}; kg/s).}

\item{Nm}{Mass of insoluble material (\eqn{N_m}; g). Must be provided only when \code{est_cu} is \code{FALSE}.}

\item{porosity}{Bed porosity (\eqn{\varepsilon}; dimensionless).}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g). May be estimated using \code{\link{oec_bp}}. Defaults to \code{NA}.
Must be provided only when \code{est_cu} is \code{FALSE}.}

\item{c3}{An initial estimate of the asymptotic extraction yield at infinite time (\eqn{c_u}; fraction) to be provided
\strong{only} when \code{est_cu} is \code{TRUE}.}

\item{Ng}{Total dry mass \eqn{N} (solute + insoluble material; g). Defaults to \code{NA}. Must be provided when \code{est_cu} is \code{TRUE}.}

\item{est_cu}{A \code{logical} indicating whether solute content in the untreated solid (kg/kg), i.e. maximum possible yield,
should be estimated (\code{FALSE} by default).}

\item{maxq}{Maximum x-axis value (solvent expended, kg/kg insoluble solid) to use for model predictions.
Defaults to 120\% of the maximum \code{q} value.}

\item{r0}{Initial estimate of the grinding efficiency \eqn{r} (fraction of broken cells).}

\item{ksas0}{Initial estimate(s) of the product of the solid phase mass transfer coefficient (\eqn{k_s}; 1/s)
and the specific area between intact and broken cells (\eqn{a_s}; 1/m), in 1/m/s.
Can be a single value or a \code{numeric} vector for multi-start optimization (max 6 values recommended).}

\item{qc0}{Initial estimate of solvent expended at the end of the CER (\eqn{q_m}; kg/kg CO2).}

\item{aggreg}{A string specifying how the "best" results of non-linear optimization (\code{nlsm}) should be chosen.
One of \code{"aard"} (default; value with minimum Average Absolute Relative Deviation is chosen) or \code{"mean"}
(the arithmetic mean of all results is taken).}

\item{modpts}{A single \code{numeric} value specifying how many points should be modelled throughout the OEC curve range (\code{100} by default).}

\item{nlsm}{A \code{character} vector of non-linear optimization methods to use for convergence of the model.
One or more of: \code{"MM"}, \code{"tau"}, \code{"CM"}, and/or \code{"mtl"} (\strong{see \code{\link[robustbase]{nlrob}} for details}).}

\item{const_flow}{A \code{logical} specifying whether the flow rate is constant throughout the extraction.
If set to \code{TRUE}, the solvent-material ratio (S/M) is \strong{not} converted to extraction time (min).}
}
\value{
A named \code{list} containing the following elements:
\enumerate{
\item \strong{$ordt}: A \code{data.frame} containing the original input data including the temporal variable \code{$t},
the Solvent-Solute (S/M) ratio \code{$x}, the \strong{actual} fractional yield (\code{$y}; g/g insoluble solid), as well as
the \strong{modeled} yield (\code{$mod_y}).
\item \strong{$mdt}: A \code{data.frame} of modeled data including the model type (\code{$model}),
the region of the extraction curve (\code{$period}), as well as S/M ratio \code{$x}, fractional yield \code{$y}, and time (\code{$t})
up to \code{maxq}.
\item \strong{$mod_pars}: The model parameters including the constants \eqn{C_1} and \eqn{C_2} (\code{"c1"} and \code{"c2"}),
the relative amount of expended solvent (kg/kg insoluble solid) at the end of the CER (\eqn{q_m}; \code{"qm"}),
the grinding efficiency (\eqn{r}; \code{"r"}), the product of the solid phase mass transfer coefficient \eqn{k_s}
and the specific area between intact and broken cells \eqn{a_s} (\eqn{k_sa_s}; \code{"ksas"}), and the initial
fraction of solute in broken cells (\eqn{G}; \code{"G"}).
See \code{\link{show_pars}} for detailed description of these parameters.
\item \strong{$fit_pars}: A \code{character} vector of which parameters in \code{$mod_pars} were iteratively fit to the model.
\item \strong{$resid}: A named \code{numeric} vector of various error values for the model, including Average Absolute Relative Deviation
AARD (\code{"aard"}), round mean squared error RMSE (\code{"rmse"}), and the R2 value (\code{"r2"}).
}
}
\description{
Derives, summarises, and visualises the results of the simplified Broken-and-Intact Cells (BIC) model of Sovova (2005).
Part of the \code{\link{bicmod}} workflow.
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}, \code{\link{oec_bp}}, \code{\link{moderr}}
}
