% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{can_numeric}
\alias{can_numeric}
\title{Check which elements of vector can or cannot be numeric}
\usage{
can_numeric(vec)
}
\arguments{
\item{vec}{Character vector to evaluate}
}
\value{
A \code{logical} vector indicating whether numeric conversion is suitable
}
\description{
Finds elements which can (or cannot) be coerced to \code{numeric} format.
}
