% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{canon}
\alias{canon}
\title{Canonical Analysis (CA) of linear/quadratic models}
\usage{
canon(object, augs, threshold = 0.1 * max.eigen)
}
\arguments{
\item{object}{A model of class \code{lm}.}

\item{augs}{A \code{list} containing the matrices of first- and second-order coefficients necessary to carry out CA.}

\item{threshold}{The canonical analysis threshold. Defaults to 10\% of the maximum Eigen value.}
}
\value{
A named \code{list} of length 2 containing the \strong{coded} factor value estimates (\code{$xs})
and the \strong{eigen decomposition} (\code{$eigen}).
}
\description{
Carries out canonical analysis of RSM results.
Part of the \code{\link{doe_analyze}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
