% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{cer_cmp_cu}
\alias{cer_cmp_cu}
\title{Equation: First extraction period (CER) of the complete BIC model ('cu' estimation)}
\usage{
cer_cmp_cu(iota, q1, y1, a0, gam, porosity, qaver, Ng, ys, xup)
}
\arguments{
\item{iota}{Fluid phase mass transfer coefficient (\eqn{k_f}; 1/s).}

\item{q1}{Relative amount of passed solvent at the end of the CER (\eqn{q \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{y1}{Yield (\eqn{e}; g/g insoluble solid) at the end of the CER.}

\item{a0}{Specific surface area per unit volume of extraction bed (\eqn{a_0}; 1/m).}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{\gamma \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{porosity}{Bed porosity (\eqn{\varepsilon}; dimensionless).}

\item{qaver}{Average solvent flow rate (\eqn{\dot{Q}}; kg/s).}

\item{Ng}{Total dry mass of solute + insoluble material (\eqn{N}; g).}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{xup}{Weight fraction (concentration) in the untreated solid \eqn{x_u}.}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
