% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{cer_cmp_eq}
\alias{cer_cmp_eq}
\title{Equation: First extraction period (CER) of the complete BIC model}
\usage{
cer_cmp_eq(iota, q1, a0, gam, porosity, qaver, Nm, ys)
}
\arguments{
\item{iota}{Fluid phase mass transfer coefficient (\eqn{k_f}; 1/s).}

\item{q1}{Relative amount of passed solvent at the end of the CER (\eqn{q \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{a0}{Specific surface area per unit volume of extraction bed (\eqn{a_0}; 1/m).}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{\gamma \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{porosity}{Bed porosity (\eqn{\varepsilon}; dimensionless).}

\item{qaver}{Average solvent flow rate (\eqn{\dot{Q}}; kg/s).}

\item{Nm}{Mass of insoluble material (\eqn{N_m}; g).}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield \eqn{e} (g/g insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The Constant Extraction Rate (CER) period of the complete Sovova (2005) model is calculated as follows (see also \strong{Eq. 3.27} in
Rizza, 2014):
\deqn{e = q \times y_s \times (1 - \exp{(-1/\theta_e)}) \text{ for } 0 \leq q \leq q_m}

In the above equation, \eqn{q_m} is the relative amount of passed solvent at the end of the CER
and the external mass transfer resistance \eqn{\theta_e} is defined as (see also \strong{Eq 3.23} in Rizza, 2014):
\deqn{\theta_e = (\varepsilon \times \dot{Q})/\gamma \times k_f \times a_0 \times N_m}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
