% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{cer_ct_eq}
\alias{cer_ct_eq}
\title{Equation: First extraction period (CER) of the characteristic times BIC model}
\usage{
cer_ct_eq(delta, t1, ys, qspec)
}
\arguments{
\item{delta}{External mass transfer resistance (\eqn{\theta_e}; dimeonsionless).}

\item{t1}{Extraction time \eqn{t_1} at the end of the CER period (s).}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{qspec}{Specific solvent flow rate (\eqn{q' \text{ in } kg_{solvent}/kg_{solid}/s}).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield (g/g insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The equation estimates the extraction yield \eqn{e'} at the end of the CER for the Characteristic Times BIC model
(see also \strong{Eqn. 3.32} in Rizza, 2014):
\deqn{e' = y_s \times q' \times (1 - \exp{(-1/\theta_e)}) \text{ for } t \leq t_1}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
