% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{cer_ct_err}
\alias{cer_ct_err}
\title{RMSE: First extraction period (CER) of the characteristic times BIC model}
\usage{
cer_ct_err(delta, t1, y1, ys, qspec)
}
\arguments{
\item{delta}{External mass transfer resistance \eqn{\theta_e} (dimeonsionless).}

\item{t1}{Extraction time at the end of the CER period \eqn{t_1} (s).}

\item{y1}{A \code{numeric} vector of \strong{observed} yields \eqn{e} (g/g insoluble solid) up to the end of the CER period.}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{qspec}{Specific solvent flow rate (\eqn{q' \text{ in } kg_{solvent}/kg_{solid}/s}).}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\seealso{
\code{\link{cer_ct_eq}}, \code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
