% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_export.R
\name{com_export}
\alias{com_export}
\title{Export COM calculation results as .CSV and graphics}
\usage{
com_export(
  comres,
  expath = getwd(),
  plotpars = "default",
  plot_format = "png",
  silent = FALSE
)
}
\arguments{
\item{comres}{The output of function \code{\link{calcom}}.}

\item{expath}{The \strong{existing} export folder path.}

\item{plotpars}{A \strong{named} \code{numeric} vector of plotting parameters, including:
width (\code{["w"]}, in inches), height (\code{["h"]}), point size (\code{["psize"]}), and dots per inch or DPI (\code{["dpi"]}).
Initially set to \code{"default"}, i.e. \code{c(w = 10, h = 12, psize = 12, dpi = 300)}.}

\item{plot_format}{Format of plots to export. One of \code{"png"} (default) or \code{"pdf"}.}

\item{silent}{Should console output be silenced? Defaults to \code{FALSE}).}
}
\value{
A .CSV file and any accompanying graphics (where present in \code{comres}) are expoted to \code{expath}.
}
\description{
Exports the results and (if any) plots generated by \code{\link{calcom}}.
}
\seealso{
\code{\link{calcom}}
}
