% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp_main.R
\name{compare_gcm}
\alias{compare_gcm}
\title{Compare estimated parameters from various GCM methods}
\usage{
compare_gcm(solute, gorder = 0, simplicity = "auto")
}
\arguments{
\item{solute}{A (optionally named) \code{character} vector of solute information as provided to \code{\link{mol_find}}.}

\item{gorder}{A \code{numeric} value or vector denoting the maximum order of groups to be considered.
Defaults to \code{0}, which considers all group orders. Must either be of \strong{length 1} or a \strong{named vector}
with names corresponding to available GCM methods (see \code{\link{est_gcm}} for a full list).}

\item{simplicity}{A \code{character} string indicating whether overlapping sub-structures should be allowed (\code{"auto"} by default).
See \code{\link{sub_smarts}}.}
}
\value{
A \code{data.frame} containing the type of estimated \code{parameter}, the \code{method} used,
the level of \code{simplicity} of matching GCM SMARTS substructures, and various estimated parameters including
boiling points (\code{"Tb"}, \code{"Tb_corr"}), critical temperature (\code{"Tc"}), pressure (\code{"Pc"}),
and volume (\code{"Vc"}), as well as dispersion (\code{"dD"}), polarity (\code{"dP"}, \code{"dP_low"}),
and hydrogen bonding (\code{"dHB"}, \code{"dHB_low"}) components of HSPs.
}
\description{
Compares the boiling points, critical parameters (temperature, pressure, volume),
and/or Hansen Solubility Parameters (HSPs) estimated for a solute by various Group Contribution Methods (GCMs).
}
\examples{
#Limonene
mol <- c("CC1=CCC(CC1)C(=C)C", "5989-54-8", "Limonene")
res <- compare_gcm(mol)

#Linoleic acid
mol2 <- c("CCCCCC=CCC=CCCCCCCCC(=O)O", "60-33-3", "Linoleic Acid")
res <- compare_gcm(mol2)

}
\seealso{
\code{\link{mol_find}}, \code{\link{sfe_mod}}, \code{\link{sub_smarts}}
}
