% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{confound}
\alias{confound}
\title{Get the confounding pattern of a design}
\usage{
confound(input, def_rel)
}
\arguments{
\item{input}{A \code{data.frame} containing the \strong{coded} input factors and their levels. Columns (factor names) must be named alphabetically.}

\item{def_rel}{A \code{character} string specifying the \strong{defining relation} of a design.}
}
\value{
A two-column \code{data.frame} containing a list of main, interaction, and/or quadratic effects and their aliasing pattern (i.e. confounding pattern) generated from the defining relation.
}
\description{
Part of the \strong{experimental design generation} workflow.
\strong{This function manual is hidden from the Index.}
}
\references{
NIST/SEMATECH (2013), 'Engineering Statistics Handbook', section 5.3.3.4.3, available at: \url{https://www.itl.nist.gov/div898/handbook/index.htm} (accessed 29.09.2024).
}
\seealso{
\code{\link{defrel}}, \code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
\keyword{internal}
