% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{define_grps}
\alias{define_grps}
\title{Modify/Process SMARTS substrings matches to remove conflicts}
\usage{
define_grps(occr)
}
\arguments{
\item{occr}{The output of function \code{\link{sub_smarts}}.}
}
\value{
A \code{list} of all GCM groups and their contributions (\code{$contrib_data}),
corresponding SMARTS strings (\code{$smarts_data}), a \strong{corrected} vector of \code{$occurrences},
group \code{$interactions} (only relevant to certain GCMs), and the chosen \code{$method}.
}
\description{
Part of the \code{\link{sfe_mod}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{sub_smarts}}
}
\keyword{internal}
