% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{defrel}
\alias{defrel}
\title{Get the defining relation of a Fractional Factorial design}
\usage{
defrel(input, factors, aliasing)
}
\arguments{
\item{input}{A \code{data.frame} containing the \strong{coded} input factors and their levels. Columns (factor names) must be named alphabetically.
Only Fractional Factorial designs are supported at present.}

\item{factors}{The number of factors included in the design.}

\item{aliasing}{A \code{character} string defining the aliasing pattern of the design.}
}
\value{
A \code{list} containing the original input \code{data.frame} (\code{$data}) and the defining relation character string (\code{$def_rel}).
}
\description{
Part of the \strong{experimental design generation} workflow.
\strong{This function manual is hidden from the Index.}
}
\seealso{
\code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
\keyword{internal}
