% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_calc}
\alias{desir_calc}
\title{Calculate desirability function based on goal}
\usage{
desir_calc(input, goal = "min", lim = range(input), wt = 1)
}
\arguments{
\item{input}{A \code{numeric} vector of \strong{response} values from which to calculate desirability.}

\item{goal}{The desirability function objective. One of \code{"min"} (minimize; \strong{default}), \code{"max"} (maximize),
or \code{"trg"} (target).}

\item{lim}{A \code{numeric} vector of length 2 containing the minimum and maximum values of response in \code{input}.}

\item{wt}{A \code{numeric} value or vector of length 2 containing the desirability weight(s) to use.
Possible values range from 0.1 to 10 (defaults to \code{1}).
Two weights must be given if \code{goal} is set to \code{"trg"}.}
}
\value{
A \code{numeric} vector of desirability values derived from \code{input}.
}
\description{
A wrapper for the \code{\link[desirability2]{d_min}}, \code{\link[desirability2]{d_min}}, and
\code{\link[desirability2]{d_target}} for generating desirability functions based on a specific goal of
maximization, minimization, or specific target for 2 or more response variables.
Part of the \code{\link{doe_desir}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_desir}}
}
\keyword{internal}
