% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desir_export.R
\name{desir_export}
\alias{desir_export}
\title{Export desirability optimization results}
\usage{
desir_export(input, expath = getwd(), detailed = FALSE, silent = FALSE)
}
\arguments{
\item{input}{The output of function \code{\link{doe_desir}}.}

\item{expath}{The \strong{existing} export folder path.}

\item{detailed}{Should detailed results be exported (\code{FALSE} by default).}

\item{silent}{Should updates about function progress be suppressed in the console? (\code{FALSE} by default).}
}
\value{
A .CSV file exported into \code{expath} containing desirability function analysis results.
}
\description{
Exports the results of desirability function analysis generated by \code{\link{doe_desir}} into a summary .CSV file.
}
\seealso{
\code{\link{doe_desir}}
}
