% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_opt}
\alias{desir_opt}
\title{Maximization of overall desirability}
\usage{
desir_opt(
  pars,
  eqs,
  dsrng,
  wts,
  obj,
  eqtype = "coded",
  opt = "min",
  optype = "single"
)
}
\arguments{
\item{pars}{A \code{numeric} vector of factor levels.}

\item{eqs}{A named \code{character} vector of \strong{coded} equations (see \code{\link{get_meq}}) for each response.}

\item{dsrng}{A named \code{list} with each element named after a response and containing minimum
and maximum limits for that response.}

\item{wts}{A \code{list} of desirability function weights with length identical to the number of responses.}

\item{obj}{A \code{character} vector of desirability function objectives of length identical to the number of responses.}

\item{eqtype}{Which type of equation to retrieve from the model of type \code{mtype}? One of \code{"coded"} (default) or
\code{"raw"}.}

\item{opt}{The objective of optimization (determines the nature of function output).
One of \code{"min"}, \code{"max"}, or \code{"none"} (default).}

\item{optype}{A \strong{deprecated} character value (either the default \code{"single"} or \code{"multi"})}
}
\value{
If \code{opt != "none"}, the final desirability is returned as a single \code{numeric} value.
Otherwise, a \code{list} containing optimized responses (\code{$resp}), corresponding desirability function
values (\code{$di}), and the overall desirability calculated via a geometric mean (\code{$do}) is returned.
}
\description{
Part of the \code{\link{doe_desir}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_desir}}, \code{\link{get_meq}}
}
\keyword{internal}
