% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_prep}
\alias{desir_prep}
\title{Extract key data from a list of screening or RSM models}
\usage{
desir_prep(mods, modbase = "final", eqbase = "final")
}
\arguments{
\item{mods}{A \code{list} of \strong{2 or more} outputs from function \code{\link{doe_analyze}}.}

\item{modbase}{Which model to retrieve data from for desirability functions? One of \code{"initial"} or \code{"final"} (default), which
represent the full model and that with insignificant effects removed, respectively.}

\item{eqbase}{Which model to retrieve the equation from? One of \code{"initial"} or \code{"final"} (default).}
}
\value{
A named \code{list} containing the following elements:
\enumerate{
\item \strong{doe_df}: A \code{data.frame} containing the coded and uncoded factor values as well as responses used to build \strong{all} models in \code{mods}.
\item \strong{resps}: A \code{character} vector of response variable names used to build \code{mods}.
\item \strong{c_tot}: A \code{character} vector of \strong{total} coded factors.
\item \strong{uc_tot}: A \code{character} vector of \strong{total un}coded factors.
\item \strong{c_facs}: A \code{character} vector of coded factors utilized in \code{mods}.
\item \strong{uc_facs}: A \code{character} vector of \strong{un}coded factors utilized in \code{mods}.
\item \strong{cf_lst}: A named \code{list} of coded factor names used for \strong{each} model contained in \code{mods}.
\item \strong{ucf_lst}: The \strong{un}coded factor equivalent to \strong{cf_lst}.
\item \strong{eqs}: A named \code{list} containing both \code{$raw} (for labeling) and \code{$coded} (for calculations) model equations
for the models in \code{mods}.
}
}
\description{
Part of the \code{\link{doe_desir}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_desir}}, \code{\link{get_meq}}
}
\keyword{internal}
