% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_sample}
\alias{desir_sample}
\title{Generation/sampling of starting points for desirability optimization}
\usage{
desir_sample(orig, c_facs, uc_facs, resps, spts, frng)
}
\arguments{
\item{orig}{A \code{data.frame} containing coded and uncoded factors as well as all responses.}

\item{c_facs, uc_facs}{Both are \code{character} vectors of coded and uncoded factor names, respectively.}

\item{resps}{A \code{character} vector of response names.}

\item{spts}{A \code{numeric} vector of length 2 providing the number of desirability optimization starting points
to be randomly generated (\strong{element 1}), and those to be retrieved from the original model \code{data.frame}
(\strong{element 2}).}

\item{frng}{A \strong{named} \code{list} of \code{numeric} vectors of length 2 containing lower and upper limits of
all factors. Names must correspond to those of \strong{coded} factors (i.e. capital letters).}
}
\value{
A \code{data.frame} containing the coded and uncoded starting points to use for desirability optimization.
}
\description{
Randomly generates and/or samples starting points for desirability function optimization.
Part of the \code{\link{doe_desir}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_desir}}, \code{\link{desir_opt}}
}
\keyword{internal}
