% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_sumlims}
\alias{desir_sumlims}
\title{Summarize desirability optimization factor and response limits}
\usage{
desir_sumlims(doe_df, c_facs, uc_facs, resps, obj, frng, dsrng, wts)
}
\arguments{
\item{doe_df}{A \code{data.frame} containing coded and uncoded factors as well as all responses.}

\item{c_facs, uc_facs}{Both are \code{character} vectors of coded and uncoded factor names, respectively.}

\item{resps}{A \code{character} vector of response names.}

\item{obj}{A \code{character} vector of desirability function objectives corresponding to \code{resps}.
For possible values, see \code{\link{doe_desir}}.}

\item{frng}{A named \code{list} of numeric vectors (length 2-3) with names corresponding to those of \strong{coded} factors.
Each element must contain minimum, maximum, and (optionally) target values of factors (see \code{\link{doe_desir}}).}

\item{dsrng}{A named \code{list} with each element named after a response in \code{resps} and containing minimum
and maximum limits for that response.}

\item{wts}{A \code{list} of desirability function weights with length identical to that of \code{resps}.}
}
\value{
A \code{list} of length 2 containing \code{data.frame} objects summarizing the lower and upper limits of
factors (\code{[[1]]}) and responses (\code{[[2]]}).
}
\description{
Summarizes desirability-related metrics and response limits.
Part of the \code{\link{doe_desir}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_desir}}
}
\keyword{internal}
