% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{desir_sumods}
\alias{desir_sumods}
\title{Summarize key metrics of several DOE models}
\usage{
desir_sumods(mods, rm_eqs = FALSE)
}
\arguments{
\item{mods}{A \code{list} of \strong{2 or more} models output from function \code{\link{doe_analyze}}.}

\item{rm_eqs}{Should model equations be removed from the final output? Defaults to \code{FALSE} (equations are kept).}
}
\value{
A \code{data.frame} with each row containing key information about a single model, including the \code{$response}
name, the \code{$model_type} and \code{$order}, performance metrics including the R2 (\code{$r2}), adjusted R2 (\code{$adj_r2}),
residual standard error (\code{$stnd_error}), the F statistic (\code{$F}), degrees of freedom (\code{$DOF_1} and \code{$DOF_2}), and
the Lack-of-Fit test p-value (\code{$lof_pvalue}). When \code{rm_eqs} is \code{FALSE}, model equations (\code{model_equation})
are also included.
}
\description{
Summarizes performance metrics of 2 or more models output from function \code{\link{doe_analyze}}.
Part of the \code{\link{doe_desir}} workflow.
}
\seealso{
\code{\link{doe_desir}}, \code{\link{doe_analyze}}
}
