% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_analyze.R
\name{doe_analyze}
\alias{doe_analyze}
\title{Build, prune, analyze, and summarize RSM models from DoE}
\usage{
doe_analyze(
  doe,
  uc_facs = NA,
  cent_id = NA,
  resp_var,
  time_var,
  mod_order = 1.5,
  canon_thres = "auto",
  p_cutoff = 0.1,
  trim_method = "stepwise",
  which_facs = "coded",
  export = "none",
  asprat = "default",
  verbose = TRUE
)
}
\arguments{
\item{doe}{A \code{data.frame} containing DoE factors (\strong{coded and uncoded}), response (\code{resp_var}),
and temporal variable (e.g. run order, \code{time_var}). \strong{Coded factor names are assumed to be capital letters ordered alphabetically}.}

\item{uc_facs}{A \code{character} vector of \strong{uncoded} factor names, in order of appearance in \code{doe}.
If not provided (\code{NA}, default), uncoded factors are not included in the analysis.}

\item{cent_id}{A unique \code{character} identifier for center points to be highlighted in Exploratory Data Analysis (EDA) plots.
Defaults to \code{NA} (i.e. no center points are highlighted).}

\item{resp_var, time_var}{Single \code{character} strings specifying the response and run order (temporal) variables as they appear in \code{doe}.}

\item{mod_order}{The order of model to build. One of: \code{1} (first-order, linear), \code{1.5} (first order with interaction terms),
or \code{2} (quadratic polynomial, suitable for RSM).}

\item{canon_thres}{A threshold specific to Canonical Analysis (CA). When set to \code{"auto"}, defaults to 10\% of the maximum eigen value.}

\item{p_cutoff}{The p-value cutoff to use when pruning the initial model. Only used when \code{trim_method} includes \code{"p_cutoff"}.}

\item{trim_method}{A \code{character} string specifying which method(s) to use for removing \strong{in}significant terms
from the initial model. One of \code{"stepwise"} (Stepwise Regression, default), \code{"p_cutoff"} (p-value cutoff),
or \code{"both"}.}

\item{which_facs}{A \code{character} string specifying which factors to use for model building. One of \code{"coded"} (default)
or \code{"uncoded"}.}

\item{export}{An \strong{existing} export folder path where plots and results are exported as .PDF (or .PNG) and .CSV files
via function \code{\link{doeopt_export}}.}

\item{asprat}{Aspect ratio of generated plots. When set to \code{"default"}, a ratio of \code{1} is applied.}

\item{verbose}{Should detailed updates be provided during processing? Defaults to \code{TRUE}.}
}
\value{
A complex \code{list} including the following elements:
\enumerate{
\item \strong{$models}: The results of \strong{initial and final} models of class \code{"lm"}.
\item \strong{$results}: Summarized key results of \strong{initial and final} models, including a \code{data.frame} of
model coefficients, aliasing, standard errors, t- and p-values, significance level, ANOVA sum of squares and F-values (\code{$Model_Results}).
Other elements include a \code{data.frame} of main, interaction, and/or quadratic effects and residuals for each observation (\code{$Model_Data}),
the original input \code{data.frame} (\code{$Orig_Data}), a \code{list} of performance metrics such as R2, F-statistic, LoF test results,
and Canonical Analysis (\code{$Model_Metrics}). Finally, a summary of key input data such as the number of observations and both
\strong{coded and uncoded} model equations are also included.
\item \strong{$plots}: A list of \code{ggplot}-class plots describing the results and including Box, EDA, Pareto, Cook's Distance, and other
plots.
\item \strong{$statements}: A vector of descriptive statements about the results.
\item \strong{$call}: The function call.
}
}
\description{
A complete workflow for modeling and analysing DoE design responses that includes the following elements:
\enumerate{
\item Building of the \strong{initial} linear or quadratic (RSM) model.
\item Carry out Lack-of-Fit (LoF) testing.
\item Removing insignificant effects via Stepwise Regression and/or p-value cutoff.
\item Building the \strong{final} simplified model.
\item Minimizing and/or maximizing the response using \code{\link[stats]{optim}} and Canonical Analysis.
\item Compiling summaries, descriptive statements, and plots for all of the above.
For further information, see \strong{Details} and \strong{References}.
}
}
\details{
The key goal of SFE experimental design is process optimization, commonly achieved via Response Surface Methodology (RSM) which
calculates both the interaction and quadratic terms and approximates the shape of the local response surface  (Sharif et al., 2014; Yolmeh et al., 2017).
Because a natural product SFE rate almost always decreases along a single smooth curvature seen in a typical OEC,
RSM is especially suitable as it predicts a local response surface by a single 2nd-order (i.e. quadratic) polynomial function,
thus failing to describe piecewise smooth or jagged response trajectories in a single experiment. An example model containing linear,
two-way, and quadratic interaction terms for 3 factors is shown below.
\deqn{y = \beta_0 + \beta_2x_2 + \beta_3x_3 + \beta_{12}x_1x_2 + \beta_{13}x_1x_3 + \beta_{23}x_2x_3 + \beta_{11}x_1^2 + \beta_{22}x_2^2 + \beta_{33}x_3^2 + \epsilon}
At least a 3-level design is required to estimate quadratic interactions. This requirement is not met by 2-level designs with
center points, which are positioned within the experimental space such that all quadratic effects are aliased (NIST/Sematech, 2013).
Three-level FFDs are useful but are not rotatable and become prohibitive due to the large number of runs required for more than 3 factors.
Instead, rotatable Box-Wilson Central Composite (CCD) or Box-Behnken (BBD) designs are used, where response is dependent only on the
distance of factor levels from the center of the experimental space (and not on direction); thus, points at an identical distance from
the center of the experimental space exhibit the same prediction error/variance  (Sharif et al., 2014).
Cubic terms are seldom included when interpreting RSM. An ANOVA table should then be examined for goodness of fit
(consistently high R2 and adjusted R2) and insignificant terms of low-magnitude exhibiting a high p value (e.g. >0.05),
Probability plots can be used as a complementary technique. Insignificant effects may then be separated systematically
using stepwise regression and/or Pareto charts (Kiratu, 2015), and the simplified ANOVA model re-examined iteratively.
Once the simplest model that explains most of the response variability is attained, the residual distribution should be examined to
validate the normality and uniformity of variance assumptions (Yolmeh et al., 2017). Transformation of response data (e.g. via a
natural logarithm or a Box-Cox approach) is useful for mitigating violations of these assumptions. The residuals can also be checked
for curvature via if a 2nd-order model was not built. Once the residual distribution is satisfactory, the
optimal factor settings can be determined by plotting the model coefficients/terms against factor levels. In case of RSM designs,
a contour plot(s) for each response is used to both optimize each response separately and/or locate a compromise between multiple responses.
The current workflow incorporates linear and quadratic model building, lack-of-fit testing, systematic removal of insignificant effects,
visualisation of important factors and optimization of factor values to maximize or minimize the response via Canonical Analysis (CA).
}
\examples{
#Maximizing extraction yield of spearmint (SFE)
doe_optres <- doe_analyze(doe = doex[["ccd3"]][["data"]],
uc_facs = c("P_bar", "T_degC", "EtOH_gmin"),
cent_id = NA,
resp_var = "ExtYield",
time_var = "Actual_Order",
mod_order = 2,
canon_thres = "auto",
p_cutoff = 0.10,
trim_method = "both",
which_facs = "coded",
export = "none",
verbose = TRUE)

}
\references{
NIST/SEMATECH (2013), 'Engineering Statistics Handbook', available at: \url{https://www.itl.nist.gov/div898/handbook/index.htm} (accessed 29.09.2024).

Kiratu, J., Raynie, D.E. (2015), 'Aiding the Development of Extraction Procedures with Response Surface Methodology', \emph{LCGC North America} \strong{33} (7), pp. 104-111.

Sharif, K.M., Rahman, M.M., Azmir, J., Mohamed, A., Jahurul, M.H.A., Sahena, F., Zaidul, I.S.M. (2014), 'Experimental design of supercritical fluid extraction – A review', \emph{Journal of Food Engineering} \strong{124}, pp. 105-116, DOI: \url{https://doi.org/10.1016/j.jfoodeng.2013.10.003}.

Yolmeh, M., Jafari, S.M. (2017), 'Applications of Response Surface Methodology in the Food Industry Processes', \emph{Food Bioprocess. Technol.} \strong{10}, pp. 413-433, DOI: \url{https://www.doi.org/10.1007/s11947-016-1855-2}.
}
\seealso{
\code{\link{doeopt_export}}
}
