% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_main.R
\name{doe_bbd}
\alias{doe_bbd}
\title{Generate a Box-Behnken experimental design}
\usage{
doe_bbd(
  factors = 3,
  cpts = 0,
  fnames,
  flims,
  randomize = TRUE,
  export = "none"
)
}
\arguments{
\item{factors}{The number of factors to use (between 3-4). The value must equal to the length of \code{fnames} and \code{flims}.}

\item{cpts}{The number of \strong{additional} center points to add to the design (0 by default). BBD designs always have 3 center points by default.}

\item{fnames}{A \code{character} vector of factor names of length 3-4.}

\item{flims}{A \code{list} of numeric vectors containing the lower and upper factor level limits \strong{in the same order as given in \code{fnames}}.}

\item{randomize}{A \code{logical} indicating whether or not the run order of the design is randomized (\code{TRUE} by default).}

\item{export}{Either \code{"none"} (default) or a path to export the generated design .TAB file to.}
}
\value{
A \code{list} containing a \code{data.frame} with the generated design as well as a short description (\code{$description}).
}
\description{
Generates a Box-Behnken experimental design for 3-4 factors and 3 factor levels.
}
\examples{
#Box-Behnken Design for 3 factors with 2 additional center points
doe_res <- doe_bbd(factors = 3,
                   cpts = 2,
                   fnames = c("Pressure", "Temperature", "Flow"),
                   flims = list(c(100, 300), c(35, 65), c(2, 4)))

}
\seealso{
\code{\link{doe_export}}, \code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_tm}}
}
