% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_box}
\alias{doe_box}
\title{Create Box plots from DoE analysis results}
\usage{
doe_box(
  mod_df,
  resids,
  resp_var,
  facs,
  cols = "default",
  asprat = 1,
  modlab = ""
)
}
\arguments{
\item{mod_df}{The \code{data.frame} used as DoE model input.}

\item{resids}{A \code{numeric} vector of model residuals.}

\item{resp_var}{A \code{character} string specifying the response variable name.}

\item{facs}{A named \code{list} of length 2 containing the \code{$coded} and \code{$uncoded} factor names.}

\item{cols}{Plot colours. Either \code{"default"}, or a \strong{named} \code{numeric} vector of any of the following colours:
outlier outline (\code{["outcol"]}), outlier fill (\code{["outfill"]}), box fill (\code{["outcol"]}).}

\item{asprat}{The plot aspect ratio. Defaults to 1.}

\item{modlab}{The model name/label as a \code{character} string (none by default).}
}
\value{
A \code{list} of \code{ggplot} box plot objects.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
