% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_build}
\alias{doe_build}
\title{Build a linear/quadratic model from DoE data}
\usage{
doe_build(input, resp_var, effs, orig_data = NA, uc_facs = NA)
}
\arguments{
\item{input}{A \code{data.frame} of DoE input data. Must contain coded factors, as well as uncoded factors and the response variable as
specified in \code{uc_facs} and \code{resp_var}, respectively.}

\item{resp_var}{A character string specifying the name of the response variable included in \code{input}.}

\item{effs}{Main, interaction, and/or quadratic model effects to evaluate (as output by \code{\link{doe_prep}}).}

\item{orig_data}{An \strong{optional} \code{data.frame} containing the original input data to be appended to the model output
(mainly for reference). Defaults to \code{NA}.}

\item{uc_facs}{A character vector of \strong{uncoded} factor names.}
}
\value{
A linear or quadratic model of class \code{lm}.
}
\description{
Builds a linear or quadratic (i.e. RSM) model from input data.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{doe_prep}}
}
\keyword{internal}
