% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_main.R
\name{doe_ccd}
\alias{doe_ccd}
\title{Generate a Central Composite experimental design}
\usage{
doe_ccd(
  design = "CCC",
  levels = 3,
  factors = 3,
  cpts = 0,
  fnames,
  flims,
  randomize = TRUE,
  export = "none"
)
}
\arguments{
\item{design}{Type of design to generate. One of \code{"CCC"} (default) or \code{"CCF"}.}

\item{levels}{The number of factor levels to use (5 and 3 when \code{design} is \code{"CCC"} and \code{"CCF"}, respectively).}

\item{factors}{The number of factors to use (between 2-4). The value must equal to the length of \code{fnames} and \code{flims}.}

\item{cpts}{The number of center points to add to the design (0 by default).}

\item{fnames}{A \code{character} vector of factor names of length 2-4.}

\item{flims}{A \code{list} of numeric vectors containing the lower and upper factor level limits \strong{in the same order as given in \code{fnames}}.}

\item{randomize}{A \code{logical} indicating whether or not the run order of the design is randomized (\code{TRUE} by default).}

\item{export}{Either \code{"none"} (default) or a path to export the generated design .TAB file to.}
}
\value{
A \code{list} containing a \code{data.frame} with the generated design as well as a short description (\code{$description}).
}
\description{
Generates either a Central Composite Circumscribed (CCC) or Central Composite Face-Centered (CCF) experimental design.
Currently available for 2-4 factors.
}
\details{
CCC' and 'CCF' designs require 5 and 3 factor levels, respectively
}
\examples{
#Central Composite Circumscribed for 4 factors (standard CCD)
doe_res <- doe_ccd(design = "CCC",
levels = 5,
factors = 3,
cpts = 6,
fnames = c("Pressure", "Temperature", "Co-Solvent"),
flims = list("hard"=c(100, 345), "hard"=c(35, 75), c(2, 8)))

}
\seealso{
\code{\link{doe_export}}, \code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
