% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_cooks}
\alias{doe_cooks}
\title{Create a Cook's Distance plot from DoE analysis results}
\usage{
doe_cooks(
  model,
  type = 1,
  plt_title = NA,
  cols = "default",
  asprat = 1,
  modlab = ""
)
}
\arguments{
\item{model}{Model results obtained from \code{\link{doe_build}}.}

\item{type}{A \code{numeric} identifier \strong{between 1-4} specifying the equation used to calculate the distance limit
above which significance is noted. One of \strong{1:} 4/n (default), \strong{2:} 4/(n-p-1),
\strong{3:} 1/(n-p-1), \strong{4:} 3 * average Cook's Distance;
\strong{p and n are the number of model coefficients and DoE data points, respectively}.}

\item{plt_title}{Plot title as a \code{character} string. Defaults to \code{NA} (no title is displayed).}

\item{cols}{Plot colours. Either \code{"default"}, or a \strong{named} \code{numeric} vector of any of the following colours:}

\item{asprat}{The plot aspect ratio. Defaults to 1.}

\item{modlab}{The model name/label as a \code{character} string (none by default).}
}
\value{
A \code{list} of \code{ggplot} box plot objects.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
