% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_decode}
\alias{doe_decode}
\title{(De)code one or more model results}
\usage{
doe_decode(input, modf, c_facs, uc_facs, task = "decode", reassemble = FALSE)
}
\arguments{
\item{input}{A named \code{numeric} vector or \code{data.frame} containing values to (de)code.
Names must match either \code{c_facs} or \code{uc_facs} based on \code{task}.}

\item{modf}{A \code{data.frame} of model input data containing \strong{both} \code{c_facs} and \code{uc_facs}.}

\item{c_facs, uc_facs}{Character vectors of coded and uncoded factor names included in \code{modf}, respectively.}

\item{task}{String specifying what should be done with \code{input} values. One of: \code{"code"} or \code{"decode"} (default).}

\item{reassemble}{Should the \strong{new} (coded or decoded) values be converted into a list format? Defaults to \code{FALSE}.}
}
\value{
A named \code{list} containing the input minima (\code{$olmins}, either coded or decoded as appropriate for the given \code{task}),
input maxima (\code{$olmaxes}), corresponding \strong{output} minima and maxima (\code{$mins} and \code{$maxes}),
and the output factor values (\code{$newvals}).
}
\description{
Uses linear interpolation to (de)code any number of provided factor values.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
