% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_eda}
\alias{doe_eda}
\title{Create Exploratory Data Analysis (EDA) plots from DoE analysis results}
\usage{
doe_eda(
  mod_df,
  mod_res,
  time_var,
  resp_var,
  cent_id = NA,
  cols = "default",
  asprat = 1,
  modlab = ""
)
}
\arguments{
\item{mod_df}{The input \code{data.frame} used to build the model via \code{\link{doe_build}}.}

\item{mod_res}{Model summary results obtained from \code{\link{doe_summary}}.}

\item{time_var, resp_var}{Strings specifying the run order and response variable names, respectively.}

\item{cent_id}{A \code{character} identifier for center point highlighting. Alternatively, if set to \code{NA} (default),
center points are \strong{not} highlighted in the plots.}

\item{cols}{Plot colours. Either \code{"default"}, or a \strong{named} \code{numeric} vector of any of the following colours:
outlier outline (\code{"outcol"}), outlier fill (\code{"outfill"}), point/box fill (\code{"fill"}), center point fill (\code{"cp"}).}

\item{asprat}{The plot aspect ratio. Defaults to 1.}

\item{modlab}{The model name/label as a \code{character} string (none by default).}
}
\value{
A \code{list} of \code{ggplot} EDA plot objects.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
