% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{doe_export}
\alias{doe_export}
\title{Export generated experimental design}
\usage{
doe_export(input, export_name = NA, expath = getwd(), silent = FALSE)
}
\arguments{
\item{input}{Output from one of the experimental design generator functions: \code{\link{doe_ffd}}, \code{\link{doe_frfd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, or \code{\link{doe_tm}}.}

\item{export_name}{An \strong{optional} title to paste as the first line of the exported .TAB file.}

\item{expath}{The directory path into which to export the output .TAB file.}

\item{silent}{A \code{logical} specifying whether console output should be suppressed (\code{FALSE} by default).}
}
\value{
A .TAB file containing the experimental design \code{data.frame} is exported to the path specified in \code{expath}.
}
\description{
Part of the \strong{experimental design generation} workflow.
}
\seealso{
\code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
