% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_main.R
\name{doe_ffd}
\alias{doe_ffd}
\title{Generate a Full Factorial experimental design}
\usage{
doe_ffd(
  levels = 2,
  factors = 3,
  cpts = 0,
  fnames,
  flims,
  randomize = TRUE,
  export = "none"
)
}
\arguments{
\item{levels}{The number of factor levels to use (between 2-3).}

\item{factors}{The number of factors to use (between 2-5). The value must equal to the length of \code{fnames} and \code{flims}.}

\item{cpts}{The number of \strong{additional} (to the default 3) center points to add to the design (0 by default).}

\item{fnames}{A \code{character} vector of factor names of length 2-5.}

\item{flims}{A \code{list} of numeric vectors containing the lower and upper factor level limits \strong{in the same order as given in \code{fnames}}.}

\item{randomize}{A \code{logical} indicating whether or not the run order of the design is randomized (\code{TRUE} by default).}

\item{export}{Either \code{"none"} (default) or a path to export the generated design .TAB file to.}
}
\value{
A \code{list} containing a \code{data.frame} with the generated design as well as a short description (\code{$description}).
}
\description{
Generates a Full Factorial design with 2-3 levels and 2-5 factors.
Three-level designs include 3 center points by default.
}
\examples{
#Generate experimental design with 3 levels, 3 factors, and 3 additional center points (6 in total)
doe_res <- doe_ffd(levels = 3,
                   factors = 3,
                   cpts = 3,
                   fnames = c("Pressure", "Temperature", "Flow"),
                   flims = list(c(100, 300), c(35, 65), c(2, 4)))

}
\seealso{
\code{\link{doe_export}}, \code{\link{doe_frfd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
