% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_main.R
\name{doe_frfd}
\alias{doe_frfd}
\title{Generate a Fractional Factorial experimental design}
\usage{
doe_frfd(
  factors = 3,
  p = 1,
  cpts = 0,
  aliasing = "default",
  fnames,
  flims,
  randomize = TRUE,
  export = "none"
)
}
\arguments{
\item{factors}{The number of factors to use (between 2-5). The value must equal to the length of \code{fnames} and \code{flims}.}

\item{p}{A numeric value describing the fraction of the Full Factorial design used. Must be 1 when \code{factors} are between 2-4, and 1-2 for 5 \code{factors}.}

\item{cpts}{The number of center points to add to the design (0 by default).}

\item{aliasing}{Either \code{"default"} or a \code{character} vector containing the aliasing patterns to use.}

\item{fnames}{A \code{character} vector of factor names of length 2-5.}

\item{flims}{A \code{list} of numeric vectors containing the lower and upper factor level limits \strong{in the same order as given in \code{fnames}}.}

\item{randomize}{A \code{logical} indicating whether or not the run order of the design is randomized (\code{TRUE} by default).}

\item{export}{Either \code{"none"} (default) or a path to export the generated design .TAB file to.}
}
\value{
A \code{list} containing a \code{data.frame} with the generated design as well as a short description (\code{$description}).
}
\description{
Generates a 2-level Fractional Factorial experimental design.
}
\examples{
#Fractional Factorial (var. 3, 2^5-2)
doe_res <- doe_frfd(factors = 5,
                    p = 2,
                    cpts = 0,
                    aliasing = "default",
                    fnames = c("Pressure", "Temperature", "Flow", "Time", "PartSize"),
                    flims = list(c(100, 300), c(35, 65), c(2, 4), c(90, 240), c(0.2, 1.0)))

}
\seealso{
\code{\link{doe_export}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
