% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_hrp}
\alias{doe_hrp}
\title{Apply the Hierarchy Principle to a list of model effects}
\usage{
doe_hrp(effs)
}
\arguments{
\item{effs}{A \code{data.frame} of model effects, e.g. output as part of \code{\link{doe_prep}}.}
}
\value{
A named \code{list} containing the original effects (\code{$old_effs}), new updated effects (\code{$new_effs}),
and the difference between the two, i.e. added effects (\code{$added_effs}, if any).
}
\description{
Applies the Hierarchy Principle to add main effects which are currently present as part of
any interaction and/or quadratic effects.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
