% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_insig}
\alias{doe_insig}
\title{Determine insignificant model effects}
\usage{
doe_insig(model, facs, resp_var, trim_method = "stepwise", p_cutoff = NA)
}
\arguments{
\item{model}{A linear or quadratic model of class \code{lm}.}

\item{facs}{A \code{character} vector of factor names included in the \code{model}.}

\item{resp_var}{A string specifying the response variable name.}

\item{trim_method}{Which method(s) of pruning insignificant effects should be applied?
One or more of: \code{"stepwise"} (stepwise regression, default), \code{"p_cutoff"} (p-value cutoff), or \code{"both"}.}

\item{p_cutoff}{The p-value cutoff to use when \code{trim_method} includes \code{"p_cutoff"} (between 0 and 1).
Values of 0.01-0.05 are usually appropriate based on the desired level of confidence.}
}
\value{
A named \code{list} of length 3 containing \code{data.frame} objects with significant effects (\code{$signif}),
as well as those removed by either stepwise regression (\code{$removed_step}) and/or p-value cutoff (\code{$removed_p}).
}
\description{
Determines insignificant effects of linear or quadratic models using Stepwise Regression and/or p-value cutoff.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
