% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_lof}
\alias{doe_lof}
\title{Carry out a Lack-of-Fit test on a model}
\usage{
doe_lof(input, resp_var, model, effs)
}
\arguments{
\item{input}{A \code{data.frame} containing the response variable (\code{resp_var}), coded and uncoded factors.}

\item{resp_var}{The \code{character} name of the response variable. Must be included in both \code{input} and \code{model}.}

\item{model}{A linear or quadratic model of class \code{lm}.}

\item{effs}{A \code{data.frame} of main, interaction, and/or quadratic effects,
in the same format output as part of \code{\link{doe_prep}}.}
}
\value{
An ANOVA \code{list} object of class \code{c("anova","data.frame")}. The output is structured identically to that of
\code{\link[stats]{anova}}.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
