% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_optim}
\alias{doe_optim}
\title{Optimize the model for either minimum or maximum response}
\usage{
doe_optim(model, mod_df, resp_var, effs, c_facs, uc_facs = NA)
}
\arguments{
\item{model}{A linear or quadratic model of class \code{lm}.}

\item{mod_df}{A \code{data.frame} of model input data containing \strong{both} \code{c_facs} and \code{uc_facs}.}

\item{resp_var}{A string specifying the response variable name.}

\item{effs}{A \code{data.frame} of model effects, e.g. output as part of \code{\link{doe_prep}}.}

\item{c_facs, uc_facs}{Character vectors of coded and uncoded factor names included in \code{model} (and \code{mod_df}), respectively.}
}
\value{
The \code{model} summary object of class \code{"summary.lm"}, including a \code{$tradopt} list element
}
\description{
Uses \code{\link[stats]{optim}} to both minimize and maximize the response variable in a linear/quadratic model.
}
\seealso{
\code{\link{doe_analyze}}
}
