% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_pareto}
\alias{doe_pareto}
\title{Create a Pareto plot from DoE analysis results}
\usage{
doe_pareto(
  model,
  conf = 0.05,
  method = "zahn",
  plt_title = NA,
  cols = "default",
  asprat = 1,
  modlab = ""
)
}
\arguments{
\item{model}{Model results obtained from \code{\link{doe_build}}.}

\item{conf}{The confidence limit to use for function \code{\link{doe_pseme}}.}

\item{method}{Method to use for PSE, ME, and SME estimation as detailed in function \code{\link{doe_pseme}}.}

\item{plt_title}{Either set to \code{NA} or the plot title as a \code{character} string.}

\item{cols}{Plot colours. Either \code{"default"}, or a \strong{named} \code{numeric} vector of any of the following colours:}

\item{asprat}{The plot aspect ratio. Defaults to 1.}

\item{modlab}{The model name/label as a \code{character} string (none by default).}
}
\value{
A \code{list} of \code{ggplot} box plot objects.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{doe_pseme}}
}
\keyword{internal}
