% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_prep}
\alias{doe_prep}
\title{Prepare DoE data for modeling and analysis}
\usage{
doe_prep(
  doe,
  time_var,
  resp_var,
  uc_facs = NA,
  which_facs = "coded",
  mod_order = 2
)
}
\arguments{
\item{doe}{A \code{data.frame} containing all coded factors (named with \strong{capital letters}), their uncoded equivalents
with names as specified in \code{uc_facs}, the response variable as specified in \code{resp_var}, and the run order as provided
in \code{time_var}.}

\item{time_var, resp_var}{Character values specifying the column names in \code{doe}
corresponding to the run order and response variables, respectively.}

\item{uc_facs}{A character vector of column names in \code{doe} specifying \strong{uncoded} factors.}

\item{which_facs}{String specifying which type of factors to process in preparation for modeling.
One of \code{"coded"} (default) or \code{"uncoded"}.}

\item{mod_order}{The \code{numeric} model order, either one of 1, 1.5, or 2.}
}
\value{
A named \code{list} containing the original input \code{doe} (\code{$orig_df}), processed input ready for modeling
(\code{$minput_df}), a sub-list of coded and uncoded factor names (\code{$all_facs}), as well as a \code{data.frame} of
all main, interaction, and/or quadratic effects as appropriate for the set \code{mod_order} (\code{$all_effs}).
}
\description{
An auto-process function that performs necessary formatting and categorizing of an input DoE \code{data.frame}.
Part of the \code{\link{doe_analyze}} workflow.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{twoway}}
}
