% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_pseme}
\alias{doe_pseme}
\title{Calculate PSE, ME, and SME from model coefficients}
\usage{
doe_pseme(effs, method = "zahn", conf = 0.05)
}
\arguments{
\item{effs}{A \code{numeric} vector of model \strong{coefficients}.}

\item{method}{The method to use for calculating PSE, ME, and SME. One of \code{"zahn"}, \code{"daniel"}, \code{"lenth"},
or \code{"juanpena"}.}

\item{conf}{The \code{numeric} confidence level (defaults to 0.05).}
}
\value{
A named \code{list} containing the Pseudo-Standard Error (\code{$pse}), (Simultaneous) Margin of Error at \code{conf}
(\code{$me} and \code{$sme}), the \code{$method} used, and the chosen \code{conf} (\code{$alpha}).
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{doe_pareto}}
}
\keyword{internal}
