% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_resid}
\alias{doe_resid}
\title{Create a residuals plot from DoE analysis results}
\usage{
doe_resid(model, resp_var, resids, cols = "default", asprat = 1, modlab = "")
}
\arguments{
\item{model}{The linear or quadratic model built using \code{\link{doe_build}}.}

\item{resp_var}{A \code{character} string specifying the response variable name.}

\item{resids}{A \code{numeric} vector of model residuals.}

\item{cols}{Plot colours. Either \code{"default"}, or a \strong{named} \code{numeric} vector of any of the following colours:
colour (\code{"col"}), fill (\code{"fill"}), line colour (\code{"line"}), and center point fill (\code{"cp"}).}

\item{asprat}{The plot aspect ratio. Defaults to 1.}

\item{modlab}{The model name/label as a \code{character} string (none by default).}
}
\value{
A \code{list} of \code{ggplot} residual plot objects.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
