% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_state}
\alias{doe_state}
\title{Compile a list of statements about the DoE analysis}
\usage{
doe_state(
  initres = NA,
  finres = NA,
  signif = NA,
  hrp = NA,
  sums = NA,
  lofs = NA,
  misc = NA,
  labs = c("initial", "final"),
  print_res = TRUE
)
}
\arguments{
\item{initres, finres}{Results of the \strong{initial} and \strong{final} DoE models built as part of \code{\link{doe_analyze}}.}

\item{signif}{Results of \code{\link{doe_insig}}.}

\item{hrp}{Results of \code{\link{doe_hrp}}.}

\item{sums}{Results of \code{\link{aug_sum}}.}

\item{lofs}{Results of \code{\link{doe_lof}}.}

\item{misc}{Additional results retrieved from the output of \code{\link{doe_summary}}.}

\item{labs}{A \code{character} vector of length 2 containing labels for the initial and final models. Defaults to \code{c("initial", "final")}.}

\item{print_res}{A \code{logical} indicating whether the generated statements should be printed into the console. Defaults to \code{FALSE}.}
}
\value{
A named \code{character} vector of descriptive statements about DoE analysis.
}
\description{
Generates a set of descriptive statements about DoE analysis results.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
