% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{doe_summary}
\alias{doe_summary}
\title{Create a DOE model analysis summary}
\usage{
doe_summary(model, mod_sum, mod_anova, mod_lof, mod_df, code_names = TRUE)
}
\arguments{
\item{model}{A model built using \code{\link{doe_build}}.}

\item{mod_sum}{Summary of \code{model} obtained using \code{\link{aug_sum}}.}

\item{mod_anova}{The results of \code{anova(model)}.}

\item{mod_lof}{The results of Lack-of-Fit testing obtained via \code{\link{doe_lof}}.}

\item{mod_df}{The initial \code{data.frame} used as model input.}

\item{code_names}{A \code{logical} indicating whether coefficient names should be switched with codenames
contained in \code{mod_sum}.}
}
\value{
A named \code{list} containing a complete DoE model analysis. For further details, see \code{\link{doe_analyze}}.
}
\description{
Compiles a complete analysis of DoE model results.
Part of the \code{\link{doe_analyze}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{doe_build}}, \code{\link{aug_sum}}, \code{\link{doe_lof}}
}
\keyword{internal}
