% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_main.R
\name{doe_tm}
\alias{doe_tm}
\title{Generate a Taguchi experimental design}
\usage{
doe_tm(
  levels = 2,
  factors = 3,
  fnames,
  flims,
  randomize = TRUE,
  export = "none"
)
}
\arguments{
\item{levels}{The number of factor levels to use (between 2-4).}

\item{factors}{The number of factors to use (between 3-5). The value must equal to the length of \code{fnames} and \code{flims}.}

\item{fnames}{A \code{character} vector of factor names of length 3-5.}

\item{flims}{A \code{list} of numeric vectors containing the lower and upper factor level limits \strong{in the same order as given in \code{fnames}}.}

\item{randomize}{A \code{logical} indicating whether or not the run order of the design is randomized (\code{TRUE} by default).}

\item{export}{Either \code{"none"} (default) or a path to export the generated design .TAB file to.}
}
\value{
A \code{list} containing a \code{data.frame} with the generated design as well as a short description (\code{$description}).
}
\description{
Generates a Taguchi Method experimental design with 2-4 factor levels and 3-5 factors.
}
\examples{
#Taguchi Design for 3 levels and 3 factors
doe_res <- doe_tm(levels = 3,
                  factors =3,
                  fnames = c("Pressure", "Temperature", "Flow"),
                  flims = list(c(100, 300), c(35, 65), c(2, 4)))

}
\seealso{
\code{\link{doe_export}}, \code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}
}
