% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeopt_export.R
\name{doeopt_export}
\alias{doeopt_export}
\title{Export DOE analysis results}
\usage{
doeopt_export(
  input,
  expath = getwd(),
  plotpars = "default",
  plot_format = "png",
  silent = FALSE
)
}
\arguments{
\item{input}{The output of function \code{\link{doe_analyze}}.}

\item{expath}{The \strong{existing} export folder path.}

\item{plotpars}{A \strong{named} \code{numeric} vector of plotting parameters, including:
width (\code{["w"]}, in inches), height (\code{["h"]}), point size (\code{["psize"]}), and dots per inch or DPI (\code{["dpi"]}).
Initially set to \code{"default"}, i.e. \code{c(w = 10, h = 12, psize = 12, dpi = 300)}.}

\item{plot_format}{Format of plots to export. One of \code{"png"} (default) or \code{"pdf"}.}

\item{silent}{Should console output be silenced? Defaults to \code{FALSE}).}
}
\value{
A .CSV file and any accompanying graphics are expoted to \code{expath}.
}
\description{
Export the results and visualisations of \code{\link{doe_analyze}} workflow as .CSV and .PDF (or .PNG) files.
}
\seealso{
\code{\link{doe_analyze}}
}
