% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{est_gcm}
\alias{est_gcm}
\title{Estimate various parameters via GCMs}
\usage{
est_gcm(
  solute,
  tb = NA,
  crit = NA,
  hsp = NA,
  vdw = NA,
  hlight = TRUE,
  simplicity = "auto",
  gorder = 0,
  silent = FALSE
)
}
\arguments{
\item{solute}{Input solute data as provided for \code{\link{mol_find}}.}

\item{tb, crit, hsp, vdw}{The chosen GCM(s) for parameter estimation (see \strong{Details}). Any of:
\enumerate{
\item \strong{For boiling point} (\code{tb}): \code{"JR"} (Joback-Reid 1987), \code{"JR_corr"} (corrected Joback-Reid),
\code{"SB"} (Stein-Brown), \code{"SB_corr"} (corrected Stein-Brown 1994), or \code{"NL04"} (Nannoolal 2004). Units are K.
\item \strong{For critical parameters} (\code{crit}): \code{"JR"} (Joback-Reid 1987), \code{"NL07"}, or \code{"NL07_robust"} (Nannoolal 2007).
The units are K, MPa, and mL/mol for critical temperature, critical pressure, and critical volume, respectively.
\item \strong{For HSP} (\code{hsp}): \code{"SP08"}, \code{"SP12"} (Stefanis & Panayiotou 2008-2012). The units for these
parameters are MPa^(1/2).
\item \strong{For all of the above}: \code{"HKR_STW"}, \code{"HKR_SIM"} (Hukkerikar 2012).
\item \strong{For van der Waals volume} (\code{vdw}): \code{"ZHAO"} (Zhao 2003), \code{"BND"} (Bondi 1964), or \code{"SLON"} (Slonimskii 1970).
The units are L/mole (i.e. SI units).
}}

\item{hlight}{A \code{logical} indicating whether substructures are highlighted (see \code{\link{plot_gcm}}).}

\item{simplicity}{A \code{character} string indicating whether overlapping sub-structures should be allowed (\code{"auto"} by default).
See \code{\link{sub_smarts}} for details on available degrees of simplicity.}

\item{gorder}{A \code{numeric} value or \strong{named} vector denoting the maximum order of groups to be considered. Defaults to \code{0}, which
considers all group orders. When length > 1, names corresponding to those provided in \code{tb}, \code{crit},
\code{hsp}, and/or \code{vdw} must be provided.}

\item{silent}{A \code{logical}. When \code{FALSE} (default), additional information is printed in the console.
See \code{\link{sfe_mod}} for more details on the available methods.}
}
\value{
A \code{list} of estimated property/parameter(s) as named \code{numeric} vectors:
\describe{
\item{\code{solute_data}}{Solute data as output from function \code{\link{mol_find}}.}
\item{\code{pares}}{A \code{list} containing the boiling point and critical parameters (element \code{$Critical}),
HSPs (element \code{$HSP}), GCMs used for their estimation (element \code{$Methods}), as well as a \code{list} of
\code{data.frame} objects containing group contributions for each calculated value (element \code{$Contribs}).}
\item{\code{visres}}{Raster visualisations of the solute molecule with GCM-identified substructures highlighted (as output from function \code{\link{plot_gcm}}).}
}
}
\description{
Part of the \code{\link{sfe_mod}} workflow. Estimates boiling point,
critical parameters (temperature, volume, pressure), van der Waals volume, and/or HSPs via Group Contribution Methods (GCM).
See \strong{Details} and \strong{References}.
}
\details{
Group Contribution Methods (GCMs) predict thermodynamic and other properties from molecular structure using known
properties of functional groups. In addition to first-order groups (i.e. primary groups), second-order groups
(a.k.a. super-groups) may also be used and often encompass several first-order groups. Some newer methods utilize
group interactions to further enhance predictions.
Herein, several group contribution methods are utilized for prediction of boiling point, critical parameters
(temperature, pressure, and volume), and Hansen Solubility Parameters. The approach of Joback & Reid (1987)
remains the most popular method to estimate both boiling point \eqn{T_b} (K), critical temperature \eqn{T_c} (K),
critical pressure \eqn{P_c} (bar), and critical molar volume \eqn{V_c} (ml/mole) using the equations below,
where \eqn{\Sigma} is the sum-product of group contributions and the number of their occurrences in a given molecule:
\deqn{T_b = 198.2 + \Sigma}
\deqn{T_c = T_b\times(0.584 + 0.965\times\Sigma - \Sigma^2)^{-1}}
\deqn{P_c = (0.113 + 0.0032n_A - \Sigma)^{-2}}
\deqn{V_c = 17.5 + \Sigma}
The standard Joback-Reid approach tends to increasingly over-predict boiling points above 500 K. For this reason,
Stein and Brown (1994) introduced a correction either side of 700 K by fitting a polynomial to a larger set of
experimental data and minimizing the residuals:
\deqn{T_b\text{ }(corr.) = T_b - 94.84 + 0.5577\times T_b - 0.0007705\times T_b^2, \text{ where } \text{ } T_b \le 700 \text{ K}}
\deqn{T_b\text{ }(corr.) = T_b + 282.7 - 0.5209\times T_b, \text{ where } \text{ } T_b > 700 \text{ K}}
Both the Joback-Reid and Stein-Brown methods only utilize first-order groups, while second-order groups and possible
group interactions are not considered. The newer approach by Nannoolal et al. (2004, 2007) incorporates both of these aspects.
Same as above, \eqn{\Sigma} is the sum-product of group contributions, which \strong{include first-order and second-order}
groups as well as group interaction contributions. The \code{"NL07_robust"} variant of this method removes some group
\strong{interaction} contributions that were based on questionable or incomplete data according the authors. The letter
\eqn{n} denotes the number of atoms in the molecule \strong{\emph{sans}-hydrogen}, while \eqn{M} is the molar mass (g/mol).
\deqn{T_b = \Sigma/(n^{0.6583} + 1.6583) + 84.3395}
\deqn{T_c = T_b\times [0.6990 + 1/(0.9889 + \Sigma^{0.8607})]}
\deqn{P_c \text{ (bar)} = M^{-0.14041}/(0.00939 + \Sigma)^2\times 0.01}
\deqn{V_c \text{ (mL/mol)} = \Sigma/n^{-0.2266} + 86.1539}
In addition to boiling point and critical parameter estimation, HSP GCMs are also included as developed by
Stefanis & Panayiotou (2008, 2012). These methods incorporate both first-order and (\strong{optionally}) second-order
groups to estimate the dispersion (\eqn{\delta_d}), polarity (\eqn{\delta_p}), and hydrogen bonding (\eqn{\delta_{HB}}) HSPs.
The earlier version of the model developed in 2008 uses these equations:
\deqn{\delta_d\text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 17.3231)\text{ } MPa^{(1/2)}}
\deqn{\delta_p \text{ } (MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 7.3548)\text{ } MPa^{(1/2)}, \text{ where } \delta_p > 3}
\deqn{\delta_p \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 2.7467)\text{ } MPa^{(1/2)}, \text{ where } \delta_p < 3}
\deqn{\delta_{HB} \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 7.9793)\text{ } MPa^{(1/2)}, \text{ where } \delta_{HB} > 3}
\deqn{\delta_{HB} \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 1.3720)\text{ } MPa^{(1/2)}, \text{ where } \delta_{HB} < 3}
An update provided in 2012 updates the equations to the following:
\deqn{\delta_d \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 959.11)^{0.4126} \text{ } MPa^{(1/2)}}
\deqn{\delta_p \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 7.6134)\text{ } MPa^{(1/2)}, \text{ where } \delta_p > 3}
\deqn{\delta_p \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 2.6560)\text{ } MPa^{(1/2)}, \text{ where } \delta_p < 3}
\deqn{\delta_{HB} \text{ }(MPa^{0.5}) = (\sum_i N_iC_i + W\sum_j M_jD_j + 7.7003)\text{ } MPa^{(1/2)}, \text{ where } \delta_{HB} > 3}
Another most versatile method for estimating boiling point, critical parameters, Hansen Solubility Parameters,
and many other values is the one developed by Hukkerikar et al. (2012) using the Marrero & Gani (2001)
set of first-, second-, and third-order groups. The associated equations are also simple sums of contributions
from first- and higher-order groups as listed below, but also incorporate a \strong{universal constant}
\eqn{UNIV} unique to each parameter. This is either added, multiplied, or otherwise incorporated into the group
contribution sum. Since this operation varies based on parameter, the universal constant was not shown in the equation
below and the reader is instead referred to the Hukkerikar et al. (2012) publication for more information:
\deqn{Parameter = \sum_{i} N_{i}C_{i} + \sum_{j} M_{j}D_{j} + \sum_{k} E_{k}O_{k}}
Finally, van der Waals volume \eqn{V_{vdW}\text{ }(\r{A}^3/molecule)} may be estimated using various methods by Zhao et al. (2003),
based on a simplified version of group contributions by Slonimskii et al. (1970) and Bondi (1964). While estimates using the
latter are simply calculated by summing group contributions and converting to \eqn{m^3/mol},
Zhao et al. (2003) is based on atom contributions and those form the number of bonds \eqn{N_B}, as well as
aromatic and non-aromatic rings (\eqn{R_A} and \eqn{R_{NA}}, respectively):
\deqn{V_{vdW} = \sum atom\text{ }contributions - 5.92N_B - 14.7R_A - 3.8R_{NA}}
Slonimskii et al. (1970) adapts a similar but slightly more complex method incorporating fused structures
of aromatic and non-aromatic rings (e.g. \eqn{R_{NA+A}}), and fused sulphur (\eqn{R_{FS}}) ring structures:
\deqn{V_{vdW} = \sum atom\text{ }contributions - 5.92N_B - 14.7R_A - 3.8R_{NA} + 5R_{A+A} + 3R_{NA+A} + 1R_{NA+NA} - 5R_{FS}}

Note that in the above equations atomic symbols represent the number of their respective occurrences in a given molecule.
}
\examples{
#Define solute data
mol <- c("CC1=C(C(CCC1)(C)C)C=CC(=CC=CC(=CC=CC=C(C)C=CC=C(C)C=CC2=C(CCCC2(C)C)C)C)C",
"7235-40-7", "Beta-carotene")

#Estimate parameters by GCMs
estres <- est_gcm(mol, "JR", "JR", "SP08")

}
\references{
Bondi, A. (1964), 'Van der Walls Volumes and Radii', \emph{The Journal of Physical Chemistry} \strong{68} (3), pp. 441-451, DOI: \url{https://doi.org/10.1021/j100785a001}.

Hukkerikar, A.S., Sarup, B., Kate, A.T., Abildskov, J., Sin, G., Gani, R. (2012), 'Group-contribution+ (GC+) based estimation of properties of pure components: Improved property estimation and uncertainty analysis', \emph{Fluid Phase Equilibria} \strong{321}, pp. 25-43, DOI: \url{https://dx.doi.org/10.1016/j.fluid.2012.02.010}.

Joback, K.G., Reid, R.C. (1987), 'Estimation of Pure Components Properties from Group-contributions', \emph{Chemical Engineering Communications} \strong{57} (1-6), pp. 233-243, DOI: \url{https://www.doi.org/10.1080/00986448708960487}.

Marrero, J., Gani, R. (2001), 'Group-contribution based estimation of pure component properties', \emph{Fluid Phase Equilibria} \strong{183-184}, pp. 183-208, DOI: \url{https://doi.org/10.1016/S0378-3812(01)00431-9}.

Nannoolal, Y., Rarey, J., Ramjugernath, D., Cordes, W. (2004), 'Estimation of pure component properties: Part 1. Estimation of the normal boiling point of non-electrolyte organic compounds via group contributions and group interactions', \emph{Fluid Phase Equilibria} \strong{226}, pp. 45-63, DOI: \url{https://doi.org/10.1016/j.fluid.2004.09.001}.

Nannoolal, Y., Rarey, J., Ramjugernath, D. (2007), 'Estimation of pure component properties: Part 2. Estimation of critical property data by group contribution', \emph{Fluid Phase Equilibria} \strong{252} (1), pp. 1-27, DOI: \url{https://doi.org/10.1016/j.fluid.2006.11.014}.

Slonimskii, G.E., Askadskii, A.A., Kitaigorodskii, A.I. (1970), 'The Packing of Polymer Molecules', \emph{Polymer Science U.S.S.R.} \strong{12} (3), p. 556-577, DOI: \url{https://doi.org/10.1016/0032-3950(70)90345-X}.

Stefanis, E., Panayiotou, C. (2008), 'Prediction of Hansen Solubility Parameters with a New Group-Contribution Method', \emph{International Journal of Thermophysics} \strong{29} (2), pp. 568-585, DOI: \url{https://www.doi.org/10.1007/s10765-008-0415-z}.

Stefanis, E., Panayiotou, C. (2012), 'A new expanded solubility parameter approach', \emph{International Journal of Pharmaceutics} \strong{426} (1), pp. 29-43, DOI: \url{https://doi.org/10.1016/j.ijpharm.2012.01.001}.

Stein, S.E., Brown, R.L. (1994), 'Estimation of normal boiling points from group contributions', \emph{Journal of Chemical Information and Computer Sciences} \strong{34} (3), DOI: \url{https://www.doi.org/10.1021/ci00019a016}.

Zhao, Y.H., Abraham, M.H., Zissimos, A.M. (2003), 'Fast Calculation of van der Waals Volume as a Sum of Atomic and Bond Contributions and Its Application to Drug Compounds', \emph{The Journal of Organic Chemistry} \strong{68} (19), DOI: \url{https://doi.org/10.1021/jo034808o}.
}
\seealso{
\code{\link{mol_find}}, \code{\link{sfe_mod}}, \code{\link{plot_gcm}}, \code{\link{sub_smarts}}
}
