% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{etoh_dens}
\alias{etoh_dens}
\title{Calculate density of ethanol and its mixtures with CO2}
\usage{
etoh_dens(temp, co2_frac = 0, co2_rho = 0)
}
\arguments{
\item{temp}{Temperature for which to calculate density (in degrees Celsius).}

\item{co2_frac}{The volume fraction of CO2 (from 0 to 0.99).}

\item{co2_rho}{Density of CO2 (\strong{in g/L}).}
}
\value{
A named \code{numeric} vector containing densities of pure ethanol (\code{["etoh"]}) and its mixture with CO2 (\code{["co2_etoh"]}).
}
\description{
Calculates the density of pure ethanol at specific temperatures.
The density of CO2-ethanol mixtures may also be calculated via the Linear Blend Rule, but density of CO2 must be provided
(for example, from the \code{\link{bendens}} function).
Part of the \code{bicmod} workflow.
}
\details{
In this implementation ethanol is considered an incompressible fluid and the following equation (Poling et al., 2008) is
used to calculate density \eqn{\rho} in g/L:

\deqn{\rho = C_1/C_2^{1+(1-T/C_3)^{C_4}} \times RMM_{EtOH}}

In this equation, \eqn{C_1} to \eqn{C_4} are constants: \eqn{C_1 = 1.6288}, \eqn{C_2 = 0.27469}, \eqn{C_3 = 514},
\eqn{C_4 = 0.23178}, \eqn{T} is temperature in Kelvin, and \eqn{RMM_{EtOH}} is the molar mass of ethanol (46.068 g/mol).
The same expression is listed and described in \strong{Eq. 3.14} (Rizza, 2014).

Density adjustment for CO2-Ethanol mixtures is carried out using the Linear Blend Rule. For example, when 6\% ethanol is used
as co-solvent (see also \strong{Eq. 3.13} in Rizza, 2014):

\eqn{1/\rho_f = 0.94/\rho_{CO2} + 0.06/\rho_{EtOH}}
}
\examples{
etoh_dens(55)
etoh_dens(55, 0.90, 880)

}
\references{
Poling, B.E., Thomson, G.H., Friend, D.G., Rowley, R.L., Wilding, W.V. (2008), 'Physical and Chemical Data', In: Perry, R.H., Green, D.W. (eds.), \emph{Perry's Chemical Engineers Handbook}, McGraw-Hill, chapter 2.

Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.
}
\seealso{
\code{\link{bicmod}}
}
