% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_cmp_cu}
\alias{fer_cmp_cu}
\title{Equation: Second (FER) and third (DC) periods of the complete BIC model ('cu' estimation)}
\usage{
fer_cmp_cu(r, ksas, xu, x, kf, ys, porosity, qaver, gam, a0, Ng)
}
\arguments{
\item{r}{Grinding efficiency \eqn{r} (fraction of broken cells).}

\item{ksas}{The product of the solid phase mass transfer coefficient (\eqn{k_s}; 1/s).
and the specific area between intact and broken cells (\eqn{a_s}; 1/m), in 1/m/s.}

\item{xu}{Weight fraction (concentration) in the untreated solid \eqn{x_u}.}

\item{x}{Relative amount of expended solvent spanning the DC (\eqn{q}; kg/kg insoluble solid).}

\item{kf}{Fluid phase mass transfer coefficient (\eqn{k_f}; 1/s).}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{porosity}{Bed porosity (\eqn{\varepsilon}; dimensionless).}

\item{qaver}{Average solvent flow rate (\eqn{\dot{Q}}; kg/s).}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{\gamma \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{a0}{Specific surface area per unit volume of extraction bed (\eqn{a_0}; 1/m).}

\item{Ng}{Total dry mass \eqn{N} (solute + insoluble material; g).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield \eqn{e} (g/g insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
