% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_cmp_eq}
\alias{fer_cmp_eq}
\title{Equation: Second (FER) and third (DC) periods of the complete BIC model}
\usage{
fer_cmp_eq(
  r,
  ksas,
  x3,
  kf,
  xu,
  ys,
  porosity,
  qaver,
  gam,
  Nm,
  a0,
  out = "final",
  x1 = NA,
  x2 = NA
)
}
\arguments{
\item{r}{Grinding efficiency (\eqn{r}; fraction of broken cells).}

\item{ksas}{The product of the solid phase mass transfer coefficient (\eqn{k_s}; 1/s)
and the specific area between intact and broken cells (\eqn{a_s}; 1/m), in 1/m/s.}

\item{kf}{Fluid phase mass transfer coefficient (\eqn{k_f}; 1/s).}

\item{xu}{Weight fraction (concentration) in the untreated solid \eqn{x_u}.}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{porosity}{Bed porosity (\eqn{\varepsilon}; dimensionless).}

\item{qaver}{Average solvent flow rate (\eqn{\dot{Q}}; kg/s).}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{\gamma \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{Nm}{Mass of insoluble material (\eqn{N_m}; g).}

\item{a0}{Specific surface area per unit volume of extraction bed (\eqn{a_0}; 1/m).}

\item{out}{One of \code{"final"} (default) or \code{"all"}. Determines which results are reported in the output (see \strong{Value}).}

\item{x1, x2, x3}{Relative amounts of expended solvent spanning the CER, FER, and DC, respectively (\eqn{q_m}, \eqn{q_n}, \eqn{q},
respectively; kg/kg insoluble solid). The first two parameters are optional (\code{NA} by default).}
}
\value{
When \code{out} is set to \code{"final"}, returns a \code{numeric} value/vector of yields spanning the DC region.
If \code{out} is \code{"all"}, a \code{list} is returned containing the following named elements:
\enumerate{
\item \strong{$pars}: A named \code{numeric} vector of parameters calculated as part of the evaluation.
\item \strong{$y1; $y2; $y3}: Vectors of calculated yields (kg/kg insoluble solid) spanning the CER, FER, and DC regions of the
extraction curve, respectively.
}
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The Falling Extraction Rate (FER) and subsequent Diffusion-Controlled (DC) periods are defined by the following equations
(see also \strong{Eqs. 3.28} and \strong{3.29} in Rizza, 2014):
\deqn{e \text{ (FER)} = q \times y_s - r \times x_u \times \theta_e \times \exp{(\beta/\theta_e \times \ln{\{1 + 1/r \times (\exp{[\frac{q - q_m}{\gamma \times \theta_i}]}-1)\}}-1/\theta_e)} \text{ for } q_m \leq q< q_n}
\deqn{e \text{ (DC)} = x_u \times [1 - \beta \times \ln{\{1 + (1 - r) \times [\exp{(1/\beta)}-1] \times \exp{(\frac{q - q_m}{\gamma \times \theta_i})}\}}] \text{ for } q \geq q_n}

Herein, the coefficient \eqn{\beta} is used to simplify equation expression and is defined as
(see also \strong{Eq. 3.26} in Rizza, 2014):
\deqn{\beta = (\gamma \times \theta_i \times y_s)/x_u}

The internal and external mass transfer resistances (\eqn{\theta_i} and \eqn{\theta_e}, respectively) are defined as
(see also \strong{Eqs. 3.22} and \strong{3.23} in Rizza, 2014):
\deqn{\theta_i = ((1-\varepsilon) \times \dot{Q})/(\gamma \times k_sa_s \times N_m)}
\deqn{\theta_e = (\varepsilon \times \dot{Q})/\gamma \times k_f \times a_0 \times N_m}

Finally, the relative amounts of solvent expended at the end of the CER and FER (\eqn{q_m} and \eqn{q_n}, respectively) is calculated via
(see also \strong{Eqs. 3.24} and \strong{3.25} in Rizza, 2014):
\deqn{q_m = (r \times x_u \times \theta_e)/y_s}
\deqn{q_n = q_m + \gamma \times \theta_i \times \ln{[1 - r + r \times \exp{(1/\beta)}]}}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
