% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_ct_eq}
\alias{fer_ct_eq}
\title{Equation: Second extraction period (FER) of the characteristic times BIC model}
\usage{
fer_ct_eq(G, ti, t2, cu, t1)
}
\arguments{
\item{G}{Initial fraction of solute in open (broken) cells \eqn{G}.}

\item{ti}{Characteristic time of the solid phase mass transfer (\eqn{t_i}; s).}

\item{t2}{A \code{numeric} vector of extraction time (\eqn{t}; s).}

\item{cu}{Solute content in the untreated solid (\eqn{c_u}; kg/kg), i.e. maximum possible yield as a \strong{fraction}.}

\item{t1}{Extraction time at the end of the CER (\eqn{t_1}; s).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield (\eqn{e'}; g/g insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The Falling Extraction Rate (FER) period of the Characteristic Times BIC model is defined as (see also \strong{Eq. 3.33} in Rizza, 2014):
\deqn{e' = c_u \times (1 - (1 - G) \times \exp{(-(-t - t_1)/t_i)}) \text { for } t > t_1}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
