% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_ct_err}
\alias{fer_ct_err}
\title{RMSE: Second extraction period (FER) of the characteristic times BIC model}
\usage{
fer_ct_err(zeta, t2, y2, cu, t1)
}
\arguments{
\item{zeta}{A \code{numeric} vector of length 2 containing the characteristic time of solid phase mass transfer \eqn{t_i} (s) and
the initial fraction of solute in open (broken) cells \eqn{G}, \strong{in that order}.}

\item{t2}{Extraction time at the end of the second extraction period (\eqn{t}; s).}

\item{y2}{A \code{numeric} vector of \strong{observed} yields \eqn{e'} after the CER.}

\item{cu}{Solute content in the untreated solid (\eqn{c_u}; kg/kg), i.e. maximum possible yield as a \strong{fraction}.}

\item{t1}{Extraction time at the end of the CER (\eqn{t_1}; s).}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\seealso{
\code{\link{fer_ct_eq}}, \code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
