% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_ct_err_cu}
\alias{fer_ct_err_cu}
\title{Equation: Second extraction period (FER) of the characteristic times BIC model ('cu' estimation)}
\usage{
fer_ct_err_cu(zeta, t2, y2, t1)
}
\arguments{
\item{zeta}{A \code{numeric} vector of length 2 containing the characteristic time of solid phase mass transfer \eqn{t_i} (s) and
the initial fraction of solute in open (broken) cells (\eqn{G}), in that order.}

\item{t2}{Extraction time at the end of the second extraction period (\eqn{t}; s).}

\item{y2}{A \code{numeric} vector of \strong{observed} yields after the CER \eqn{e'}.}

\item{t1}{Extraction time at the end of the CER (\eqn{t_1}; s).}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{fer_ct_eq}}, \code{\link{fer_ct_err}}, \code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
