% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_sim_cu}
\alias{fer_sim_cu}
\title{Equation: Second extraction period of the simplified BIC model ('cu' estimation)}
\usage{
fer_sim_cu(c1, c2, c3, x, Ng)
}
\arguments{
\item{c1, c2, c3}{Constants \eqn{C_1}, \eqn{C_2}, and \eqn{C_3}.}

\item{x}{A \code{numeric} value or vector containing the input data (relative amount of solvent passed \eqn{q}; kg/kg insoluble solid).}

\item{Ng}{Total dry mass \eqn{N} (solute + insoluble material; \strong{g}).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield \eqn{e} (g/g insoluble solid).
}
\description{
A modification of the simplified BIC model equation in \code{\link{fer_sim_eq}} used to estimate
\eqn{c_u}, the solute content of the untreated solid (kg/kg).
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{fer_sim_eq}}, \code{\link{show_pars}}, \code{\link{bicmod}}
}
\keyword{internal}
