% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_sim_eq}
\alias{fer_sim_eq}
\title{Equation: Second extraction period of the simplified BIC model}
\usage{
fer_sim_eq(c1, c2, x, xu)
}
\arguments{
\item{c1, c2}{Constants \eqn{C_1} and \eqn{C_2}.}

\item{x}{A \code{numeric} value or vector containing the input data (relative amount of solvent passed \eqn{q}; kg/kg insoluble solid).}

\item{xu}{Weight fraction (concentration) \eqn{x_u} in the untreated solid.}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the extraction yield \eqn{e} (g/g insoluble solid).
}
\description{
One of the equations used for the \strong{Broken and Intact Cells (BIC)} model workflow in \code{\link{bicmod}}.
See also \strong{Details} for more information.
\strong{This function manual is hidden from the Index.}
}
\details{
The relevant equation describes the \strong{second period} of a \strong{simplified} BIC model (Rizza, 2014, \strong{Eq. 3.19}):
\deqn{e = x_u \times (1-C_1 \times exp(-C_2 \times q)) \text{ for } q > q_c}
Here, \eqn{e} is extraction yield, \eqn{x_u} is the fraction of solute in untreated solid, \eqn{q_c} is the relative
amount of solvent passed at the end of the Constant Extraction Period (CER; kg/kg of \strong{insoluble solid}),
while \eqn{C_1} and \eqn{C_2} are constants.

Note that the CER period is considered completely linear such that \eqn{e = qy_s} where \eqn{y_s} is extract solubility
(in g/g; slope of best-fit line).
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
