% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{fer_sim_err}
\alias{fer_sim_err}
\title{RMSE: Second extraction period of the simplified BIC model}
\usage{
fer_sim_err(alpha, x, xu, input)
}
\arguments{
\item{alpha}{A \code{numeric} vector of length 2 containing constants \eqn{C_1} and \eqn{C_2} from \code{\link{fer_sim_eq}}.}

\item{x}{A \code{numeric} value or vector containing the input data (relative amount of solvent passed \eqn{q}; kg/kg insoluble solid).}

\item{xu}{Weight fraction (concentration) \eqn{x_u} in the untreated solid.}

\item{input}{A vector of \strong{observed} extraction yields \eqn{e}.}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Calculates Round-Mean-Squared-Error (RMSE) for the BIC model equation in \code{\link{fer_sim_eq}}.
One of the equations used for the \strong{Broken and Intact Cells (BIC)} model workflow in \code{\link{bicmod}}.
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{fer_sim_eq}}, \code{\link{bicmod}}
}
\keyword{internal}
