% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flattenlist}
\alias{flattenlist}
\title{Flatten nested lists}
\usage{
flattenlist(x)
}
\arguments{
\item{x}{A nested \code{list} object.}
}
\value{
A one-level list created from the input list \code{x}.
}
\description{
Flattens nested lists into a one-level list.
}
