% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{gcm_contribs}
\alias{gcm_contribs}
\title{Compile GCM group contributions}
\usage{
gcm_contribs(solute, occr)
}
\arguments{
\item{solute}{Output from function \code{\link{mol_find}}.}

\item{occr}{Output from function \code{\link{define_grps}}.}
}
\value{
A \code{list} of the summarized group contributions for the input molecule (\code{$results}),
solute descriptors and MOL representation (\code{$solute}), SMARTS strings of all \strong{detected} groups
(\code{$smarts}), and the applied \code{$method}.
}
\description{
Part of the \code{\link{sfe_mod}} workflow.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{mol_find}}, \code{\link{sfe_mod}}
}
\keyword{internal}
