% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desir.R
\name{get_meq}
\alias{get_meq}
\title{Retrieve linear or quadratic model equation}
\usage{
get_meq(mod, mtype = "final", eqtype = "coded")
}
\arguments{
\item{mod}{The output of function \code{\link{doe_analyze}}.}

\item{mtype}{Which model to retrieve the equation from? One of \code{"initial"} or \code{"final"} (default).}

\item{eqtype}{Which type of equation to retrieve from the model of type \code{mtype}? One of \code{"coded"} (default) or
\code{"raw"}.}
}
\value{
A named string containing the \strong{raw} or \strong{coded} equation for \code{mod}.
}
\description{
Retrieves the coded or uncoded model equation from the output of function \code{\link{doe_analyze}}.
Part of the \code{\link{doe_desir}} workflow.
}
\examples{
doe_lst1 <- load_internal("doe_lst1")
get_meq(doe_lst1[[1]])

}
\seealso{
\code{\link{doe_desir}}, \code{\link{desir_prep}}, \code{\link{doe_analyze}}
}
