% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{h2o_dens}
\alias{h2o_dens}
\title{Calculate water density in the subcritical region}
\usage{
h2o_dens(temp = 25, pres = 1)
}
\arguments{
\item{temp}{Water temperature in degrees Celsius. Defaults to \code{25}.}

\item{pres}{Water pressure in bar. Defaults to \code{1}.}
}
\value{
The water density (in g/L) as a single \code{numeric} value.
}
\description{
Calculates the density of pure water in the subcritical region (<374 degC, <221 bar)
via the International Association for the Properties of Water and Steam formulation IAPWS-95 (Wagner & Pruß, 2002).
}
\examples{
h2o_dens(200, 40)

}
\references{
Wagner, W., Pruß, A. (2002), 'The IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use', \emph{J. Phys. Chem. Ref. Data} \strong{31}, pp. 387-535, DOI: \url{https://doi.org/10.1063/1.1461829}.
}
