% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp_export.R
\name{hsp_export}
\alias{hsp_export}
\title{Export results of HSP-based miscibility enhancement function}
\usage{
hsp_export(
  input,
  addres = NA,
  plotpars = "default",
  plot_format = "png",
  expath = getwd(),
  silent = FALSE
)
}
\arguments{
\item{input}{The output of function \code{\link{sfe_mod}}.}

\item{addres}{Additional output of class \code{"ggplot"}. Defaults to \code{NA}.}

\item{plotpars}{Either \code{"default"} or a \strong{named} vector containing one or more of the
following parameters: \code{"w"} (width, defaults to 10 inches),
\code{"h"} (height, defaults to 12 inches),
\code{"psize"} (point size, default to 12),
\code{"res"} (raster resolution, defaults to 800 px, only applies when \code{plot_format} is \code{"png"}),
\code{"dpi"} (vector resolution, defaults to 300 dpi, only applies when \code{plot_format} is \code{"pdf"}).}

\item{plot_format}{The plot format. One of: \code{"png"} (default) or \code{"pdf"}.}

\item{expath}{The export directory as a \code{character} value. The path must exist on the system.}

\item{silent}{A \code{logical}. When \code{FALSE} (default), additional information is printed in the console.}
}
\value{
A folder with a timestamp is created in the output directory \code{expath}, into which the output
\strong{.CSV} file is exported. Any plots are exported into an eponymous sub-folder.
}
\description{
Exports the output of function \code{\link{sfe_mod}} as a .CSV file along any additional data of class \code{"ggplot"}.
}
\seealso{
\code{\link{sfe_mod}}
}
