% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{hsp_optim}
\alias{hsp_optim}
\title{Optimize HSP-based solubility enhancement for SFE}
\usage{
hsp_optim(
  cp,
  hsp_vals,
  modif,
  modfracs = NA,
  pres = seq_last(80, 500, 20),
  temps = seq_last(32, 65, 3),
  vfrac = 0.1,
  silent = FALSE
)
}
\arguments{
\item{cp}{The \code{numeric} critical temperature of the solute (\strong{in Kelvin}).}

\item{hsp_vals}{A named \code{numeric} vector of length 3-5 containing HSPs for the solute.
Names must be: \code{"dD"}, \code{"dP"}, \code{"dHB"}, optionally also including \code{"dP_low"} and/or \code{"dHB_low"}.}

\item{modif}{A \code{character} vector of one or more modifiers (i.e. co-solvents) to evaluate.
May also be a \code{list} where each element of \code{length > 1} includes two or more solvents to be \strong{blended}.
For a list of modifiers, see \code{\link{sfe_mod}}.}

\item{modfracs}{Either \code{NA} (default) or a \code{list} of numeric vectors specifying the \strong{volume} percentages
for solvent blends included in \code{modif}. Overall length must match the number of blends provided.
Each volume percentage must be \code{>0} and \code{<100}, and the vector length must either equal the number of corresponding
blend elements (in which case the sum must equal to 100) or one less (in which case the fraction of the last element
is automatically determined to yield a total percentage of 100).}

\item{pres, temps}{A \code{numeric} vector of one or more pressure (75-700 bar) and temperature (31-70) values
at which to evaluate co-solvents.}

\item{vfrac}{The volume fraction of co-solvent(s) to use (between 0-1). Defaults to 0.1 (10\% co-solvent).}

\item{silent}{A \code{logical} indicating whether to show user updates when executing the function
(\code{FALSE} by default).}
}
\value{
A \code{list} containing the following elements:
\describe{
\item{\code{Modifiers}}{A \code{character} vector of modifier names used for the evaluation.}
\item{\code{Volume_Fraction}}{The \code{numeric} volume fraction (between 0 and 1) used for the evaluation.}
\item{\code{SoluteHSP_vs_Temp}}{A \code{list} of 3 \code{data.frame} objects showing the variation of dD, dP, and dHB HSP parameters with temperature (as set by the \code{temps} argument).}
\item{\code{SolventBlend_HSPs}}{A \strong{nested} \code{list} of \code{data.frame} objects showing the variation of HSP parameters with pressure and temperature \strong{for each chosen modifier}.}
\item{\code{Ra}}{A \code{list} of \code{data.frame} objects showing the calculated HSP distances (\eqn{R_a}) for selected pressure and temperature values (\code{pres} and \code{temps}).}
\item{\code{Miscib_Enhancement}}{A \code{list} of \code{data.frame} objects showing the \strong{percentage Miscibility Enhancement} for pure CO2 and CO2-modifier mixtures.}
\item{\code{Best_Modifier}}{A \code{data.frame} showing the modifier providing the highest Miscibility Enhancement for the solute at every combination of pressure and temperature.}
\item{\code{Modifier_Ranking}}{A \strong{named} vector showing the percentage occurrence of each modifier in the \code{Best_Modifier} output list element.}
}
}
\description{
Part of the \code{\link{sfe_mod}} workflow. Assesses requirement and optimizes
co-solvent choice for Supercritical CO2 Extraction (SFE) from Hansen Solubility Parameters (HSPs)
and critical temperature of any solute.
}
\examples{
#Using limonene as an example
lim_cp <- 628.4626
lim_hsp <- c(dD = 14.00539, dP = 0, dHB = 1.7012, dP_low = 2.0448, dHB_low = 1.2838)

#With simple pure co-solvents
modif <- c("Methanol", "Ethanol", "Hexane")
res <- hsp_optim(lim_cp, lim_hsp, modif, silent = TRUE)

#Including some co-solvent blends
modif <- list(c("Ethanol", "Water"), "Ethanol", "Methanol", c("Ethanol", "Methanol"))
modfracs <- list(c(50,50), 34)
res <- hsp_optim(lim_cp, lim_hsp, modif, modfracs, silent = TRUE)

}
\seealso{
\code{\link{sfe_mod}}
}
