% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hspex}
\alias{hspex}
\title{Example data for estimation of solubility and critical parameters}
\format{
An object of class \code{data.frame} with 33 rows and 7 columns.
}
\usage{
hspex
}
\description{
A data frame with 33 rows and 7 columns containing example data about chemically diverse compounds to be used for
Hansen solubility parameter (HSP), boiling point, and critical parameter estimation (pressure, temperature, volume). Data includes the following columns:
\describe{
\item{Name}{Name of the compound.}
\item{CAS}{The CAS number.}
\item{MVol}{The molar volume of the compound under standard temperature and pressure.}
\item{D}{The deltaD (dispersion, van der Waals forces) Hansen solubility parameter.}
\item{P}{The deltaP (dipole moment) Hansen solubility parameter.}
\item{HB}{The deltaH (hydrogen bonding) Hansen solubility parameter.}
\item{SMILES}{The \strong{canonical} SMILES notation.}
}
}
\references{
Hansen Solubility Parameters, available at: \url{https://www.hansen-solubility.com} (accessed 02.06.2024).

National Library of Medicine, available at: \url{https://pubchem.ncbi.nlm.nih.gov} (accessed 02.06.2024).
}
\keyword{datasets}
