% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{iscrit_etoh}
\alias{iscrit_etoh}
\alias{iscrit_demo}
\title{Determine critical parameters of binary CO2-Ethanol mixtures}
\usage{
iscrit_etoh(fracs, pres = NA, temp = NA, units = "mass", method = "redlich")

iscrit_demo()
}
\arguments{
\item{fracs}{A \code{numeric} value specifying the molar fraction of CO2 in the binary mixture.}

\item{pres, temp}{\strong{Optional} pressure and temperature (in bar and degrees Celsius, respectively), to
at which to determine whether the physical state of the binary mixture is supercritical or not.}

\item{units}{One of \code{"mass"} (default) or \code{"mol"}. Specifies the units of \code{fracs}.}

\item{method}{One of \code{"chueh"} (Chueh-Prausnitz method), \code{"redlich"} (Redlich-Kister method),
or \code{"both"}.}
}
\value{
A \code{list} including the \code{$results} of critical pressure and temperature estimations for each
specified method, and a \code{$statement} summarizing the physical state of the system at the optionally provided
pressure (\code{pres}) and temperature (\code{temp}).
}
\description{
A more accurate determination of critical parameters specific to
CO2-Ethanol binary mixtures using the Chueh-Prausnitz and Redlich-Kister methods.
These cover the entire CO2 molar fraction range of 0-1, while Redlich-Kister also reduces
overestimation of the Chueh-Prausnitz method at high molar fractions of >0.5 CO2.
}
\details{
The calculations required for the Chueh-Prausnitz and Redlich-Kister methods may be found both in the source
publications cited below. They are also summarized and optimized coefficients given in Sun et al. (2022), with
experimental data sourced from Gil et al. (2012).
}
\examples{
res <- iscrit_etoh(0.8, "mol", pres = 400, temp = 45, "redlich")
res[["statement"]]

#For visualization of both methods
iscrit_demo()

}
\references{
Chueh, P.L., Prausnitz, J.M. (1967), 'Vapor-Liquid Equilibria at High Pressures: Calculation of Critical Temperatures, Volumes, and Pressures of Nonpolar Mixtures', \emph{AIChE Journal} \strong{13} (6), pp. 1107-1113, DOI: \url{https://doi.org/10.1002/aic.690130613}.

Gil, L., Blanco, S.T., Rivas, C., Laga, E., Fernandez, J., Artal, M., Velasco, I. (2012), 'Experimental determination of the critical loci for {n-C6H14 or CO2 + alkan-1-ol} mixtures. Evaluation of their critical and subcritical behavior using PC-SAFT EoS', \emph{Journal of Supercritical Fluids} \strong{71}, pp. 26-44, DOI: \url{https://www.doi.org/10.1016/j.supflu.2012.07.008}.

Redlich, O., Kister, A.T. (1948), 'Algebraic Representation of Thermodynamic Properties and the Classification of Solutions', \emph{Industrial & Engineering Chemistry} \strong{40} (2), pp. 345-348, DOI: \url{https://doi.org/10.1021/ie50458a036}.

Sun, R., Tian, H., Wu, Z., Shi, L., Shu, G. (2022), 'Comparative Study on Critical Points of Carbon Dioxide-Based Binary Mixtures', \emph{International Journal of Thermophysics} \strong{43}, article 122, DOI: \url{https://doi.org/10.1007/s10765-022-03048-3}.
}
\seealso{
\code{\link{iscrit_gen}}
}
