% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{iscrit_gen}
\alias{iscrit_gen}
\title{Determine the critical parameters of a solvent mixture and its current state}
\usage{
iscrit_gen(
  solv,
  fracs,
  pres = NA,
  temp = NA,
  units = "mass",
  tc = "all",
  pc = "all",
  getdf = TRUE,
  silent = FALSE
)
}
\arguments{
\item{solv}{A \code{character} vector of two or three solvent names to calculate critical parameters for. For a list of solvents,
see \code{\link{show_solv}}.}

\item{fracs}{Fractions corresponding to solvents given in \code{solv}. Must sum to 1. Fraction units are specified in the
eponymous argument, \code{units}.}

\item{pres, temp}{\strong{Optional} pressure and temperature (in bar and degrees Celsius, respectively), to
at which to determine whether the physical state of the binary mixture is supercritical or not.}

\item{units}{The units of solvent fractions specified in \code{fracs}. One of \code{"mass"} (default) or \code{"mol"}.}

\item{tc, pc}{Methods to use for parameter estimation specified by \code{character} values. Currently implemented methods
include \code{"KAY"} (Kay, 1938), \code{"LI"} (Li, 1971),
\code{"FECP"} (First Extended Chueh-Prausnitz; Najafi et al., 2014, 2015) \strong{for critical temperature only},
\code{"HECP"} (He et al., 2017),
and variations of the Tang method \code{"TANG1"}, \code{"TANG2"}, \code{"TANG3"}, and \code{"TANG4"} (Tang et al., 2025).}

\item{getdf}{A \code{logical} specifying whether to compile the final predictions from all methods into a \code{data.frame}.}

\item{silent}{A \code{logical} that suppresses information printed to the console when \code{TRUE} (defaults to \code{FALSE}).}
}
\value{
A list with the following elements:
\describe{
\item{\code{$fractions}: }{Mole and mass fractions for all solvents provided in \code{solv}.}
\item{\code{$results}: }{A \code{list} of estimated parameters for each of the methods specified in \code{tc} and \code{pc}.}
\item{\code{$results_df}: }{A \code{data.frame} compilation of \code{$results} described above (only when \code{getdf} is \code{TRUE}).}
\item{\code{$global_means}: }{Global averages and standard deviations of parameters obtained from all estimation methods.}
\item{\code{$statement}: }{\strong{Optional} descriptive text which specifies whether the examined solvent mixture
is supercritical at given reference pressure and temperature.}
}
}
\description{
Uses various empirical approaches to estimate the critical tempearatures and pressures
of binary and ternary mixtures given their mole or mass fractions. See \strong{Details} for further information.
}
\details{
The function implements empirical methods of Kay (1938), Li (1971), as well as various extensions of the Chueh-Prausnitz method
introduced by Najafi et al. (2014, 2015), He et al. (2017), and Tang et al. (2025). Comprehensive descriptions of methods
can be found in the source publications.
}
\examples{
res <- iscrit_gen(solv = c("CO2", "Hexane", "Methanol"),
fracs = c(0.216, 0.196, 0.588),
pres = 400,
temp = 45,
units = "mol")

}
\references{
He, M., Liu, Y., Liu, X. (2017), 'Prediction of critical temperature and critical pressure of multicomponent mixtures', \emph{Fluid Phase Equilibria} \strong{441}, pp. 2-8, DOI: \url{http://dx.doi.org/10.1016/j.fluid.2016.11.017}.

Kay, W.B. (1938), 'Liquid-Vapor Phase Equilibrium Relations in the Ethane-n-Heptane System', \emph{Industrial & Engineering Chemistry} \strong{30} (4), pp. 459-465, DOI: \url{https://doi.org/10.1021/ie50340a023}.

Li, C.C. (1971), 'Critical temperature estimation for simple mixtures', \emph{The Canadian Journal of Chemical Engineering} \strong{49} (5), pp. 709-710, DOI: \url{https://doi.org/10.1002/cjce.5450490529}.

Najafi, H., Maghbooli, B., Sobati, M.A. (2014), 'Prediction of true critical temperature of multi-component mixtures: An extension to Chueh and Prausnitz method', \emph{Fluid Phase Equilibria} \strong{363}, pp. 1-17, DOI: \url{http://dx.doi.org/10.1016/j.fluid.2013.10.054}.

Najafi, H., Maghbooli, B., Sobati, M.A. (2015), 'Prediction of true critical temperature of multi-component mixtures: Extending fast estimation methods', \emph{Fluid Phase Equilibria} \strong{392}, pp. 104-126, DOI: \url{http://dx.doi.org/10.1016/j.fluid.2015.02.001}.

Tang, B., Dong, X., Zhao, Y., Gong, M. (2025), 'Prediction for Critical Temperature and Critical Pressure of Mixtures by Improved Empirical Correlations', \emph{International Journal of Thermophysics} \strong{46}, article 83, DOI: \url{https://doi.org/10.1007/s10765-025-03560-2}.
}
\seealso{
\code{\link{iscrit_etoh}}, \code{\link{iscrit_demo}}
}
