% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_export.R
\name{kin_export}
\alias{kin_export}
\title{Export the results of one or more kinetic models and their summaries}
\usage{
kin_export(
  modres,
  sumres = NA,
  expath = getwd(),
  which_mod = "all",
  plotpars = "default",
  plot_format = "png",
  silent = FALSE
)
}
\arguments{
\item{modres}{The output of \code{\link{bicmod}} or \code{\link{ktsmod}}.}

\item{sumres}{\strong{Optional} results summarizing \strong{several} kinetic models output from function \code{\link{kin_splot}}.}

\item{expath}{An \strong{existing} directory where the results should be exported. Defaults to the working directory.}

\item{which_mod}{A \code{character} vector specifying which model results should be exported. Either \code{"all"} (default)
or any of \code{"sim"} (simple BIC), \code{"ct"} (characteristic times BIC), \code{"cmp"} (complete BIC), and/or \code{"tws"}
(two-site kinetic desorption).}

\item{plotpars}{Either \code{"default"} or a \strong{named} vector containing one or more of the
following parameters: \code{"w"} (width, defaults to 10 inches), \code{"h"} (height, defaults to 12 inches),
\code{"psize"} (point size, default to 12), \code{"dpi"} (resolution, defaults to 300 dpi).}

\item{plot_format}{The plot format. One of: \code{"png"} (default) or \code{"pdf"}.}

\item{silent}{A \code{logical}. When \code{FALSE} (default), additional information is printed in the console.}
}
\value{
Results as a .CSV files and plots as either .PDF or .PNG files are output into a new directory in \code{expath}.
}
\description{
Compiles and exports the results and plots of Broken-and-Intact Cells (BIC) or Two-Site Kinetic Desorption
(TWS) model into a specified directory.
}
\seealso{
\code{\link{bicmod}}, \code{\link{ktsmod}}, \code{\link{kin_splot}}, \code{\link{show_pars}}
}
