% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{kin_plot}
\alias{kin_plot}
\title{Create kinetic model plots}
\usage{
kin_plot(
  pts,
  mod,
  ptvars = c("x", "y"),
  modvars = NA,
  grp = c(mod = "model", reg = "period"),
  cols = "default",
  pltlabs = c(x = ptvars[1], y = ptvars[2]),
  draw = FALSE
)
}
\arguments{
\item{pts}{A \code{data.frame} of original Overall Extraction Curve (OEC) points (including both temporal and signal variables).}

\item{mod}{A \code{data.frame} of the model curve to plot alongside \code{pts}.}

\item{ptvars, modvars}{A \code{character} vector of length 2 specifying column names of the temporal and signal variables
in \code{pts} and \code{mod}, respectively. If \code{modvars} is not provided, it is set to be equal to \code{ptvars}.}

\item{grp}{An \strong{optional} named \code{character} vector of length 2 specifying column names with grouping variables
for \strong{model type} (\code{"mod"}) and \strong{OEC region} (\code{"reg"}).}

\item{cols}{Either \code{"default"} or a named \code{character} vector of plot colours for points (\code{"id"}),
model curve without sub-division into regions (\code{"mod"}), as well as with sub-division into CER (\code{"cer"}),
FER (\code{"fer"}), and DC (\code{"dc"}) regions of the extraction curve.}

\item{pltlabs}{An \strong{optional} named \code{character} vector of labels for the temporal (\code{"x"}) and
signal (\code{"y"}) variables to display in plots. Defaults to \code{ptvars}.}

\item{draw}{A \code{logical} switch specifying whether the generated plot(s) should be printed? Defaults to \code{FALSE}.}
}
\value{
A \code{list} of one or more plots of class \code{"ggplot"}.
}
\description{
Creates a plot of BIC or other kinetic model showing extraction curve points and up to three regions of the
modeled curve including the Constant Extraction Rate (CER), Falling Extraction Rate (FER), and the Diffusion-Controlled (DC) regions.
Part of the \code{\link{bicmod}} workflow.
}
\seealso{
\code{\link{bicmod}}, \code{\link{ktsmod}}, \code{\link{show_pars}}
}
