% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_splot.R
\name{kin_splot}
\alias{kin_splot}
\title{Visually compare several kinetic extraction models}
\usage{
kin_splot(
  m,
  which_mod,
  mvars = c(x = "x", y = "y"),
  cols = "default",
  leglab = NA,
  axlab = c(title = NA, x = NA, y = NA)
)
}
\arguments{
\item{m}{A \code{list} of \strong{2-5} output objects from function \code{\link{bicmod}} \strong{OR} \code{\link{ktsmod}}.}

\item{which_mod}{A string specifying which type of models to summarize and plot. \strong{One} of \code{"sim"} (simplified BIC),
\code{"ct"}, \code{"cmp2"} or \code{"cmp3"} (complete BIC with 2 or 3 OEC regions, respectively), \code{"tws"} (two-site kinetic desorption).}

\item{mvars}{A \strong{named} \code{character} vector of column names specifying \code{"x"} and \code{"y"} values in all results in \code{m}.
Defaults to \code{c(x = "x", y = "y")}.}

\item{cols}{Either \code{"default"} or a \strong{named} \code{character} vector of point/line colours for models \code{"one"},
\code{"two"}, \code{"three"}, \code{"four"}, and/or \code{"five"}. Unspecified colours are set to the following defaults:
\code{c(one = "darkgreen", two = "darkorange", three = "darkred", four = "blue", five = "purple")}.}

\item{leglab}{A \code{character} vector of legend labels for models in the same order as in argument \code{m}.}

\item{axlab}{A \strong{named} character vector of plot labels for the x-axis (\code{"x"}), y-axis (\code{"y"}),
and the title (\code{"title"}).}
}
\value{
A \code{list} containing the following results:
\enumerate{
\item \strong{$plot}: A summary plot containing experimental points and modeled curves for all models of type \code{which_mod}
in argument \code{m}.
\item \strong{$summary}: A \code{data.frame} summarizing key input and calculated parameters of evaluated models, as well as three standard
error metrics (see \code{\link{moderr}}).
See/use \code{\link{bicmod}}, \code{\link{ktsmod}}, and \code{\link{show_pars}} for parameter descriptions.
\item \strong{$model_type}: A \code{character} string stating the value of \code{which_mod} (the summarized model type) for user information.
}
}
\description{
Summarizes the results of between two and five kinetic models both visually and in a table.
}
\examples{
#Prepare input data
bicdt <- list(sfex[["rizza1"]][["data"]],
sfex[["rizza2"]][["data"]],
sfex[["rizza3"]][["data"]])

#Estimates of extractable solute fraction
cuvals <- c(0.165, 0.21, 0.08)

#Generate several BIC models
biclst <- list()
for(i in seq_along(bicdt)) {
  biclst[[i]] <- bicmod(oec = bicdt[[i]],
                        oec_vars = c(x = "Time_min", y = "Yield_g", slv = "Solvent_mL"),
                        pars = c(pres = 300,
                                 cu = cuvals[i],
                                 temp = 45,
                                 flow = NA,
                                 mass_in = 0.5125,
                                 moisture = 8.6,
                                 D = 0.015,
                                 L = 0.015,
                                 etoh = 0.5,
                                 dr = 1554,
                                 dp = 0.0004,
                                 n = 2),
                        opt_est = "default",
                        flowpar = c(1.01325, 25),
                        etoh_frac = 0.06, #For when CO2 flow not provided but 'etoh' is >0
                        ro_co2 = NA,
                        tmax = NA,
                        qmax = NA,
                        cumulative = FALSE,
                        mass_flow = FALSE,
                        draw = FALSE,
                        units = c(flow = "none", resp = "g"),
                        modtype = "all") #"cu"
}

#Summarize the results
bic_summary <- kin_splot(m = biclst,
                         which_mod = "cmp3",
                         mvars = c(x = "x", y = "y"), #c("x", "y")
                         leglab = c("S. obliquus", "N. salina", "C. protothecoides"),
                         axlab = c(title = "BIC models", x = "q (kg/kg)", y = "e (kg/kg)"))

}
\seealso{
\code{\link{bicmod}}, \code{\link{ktsmod}}, \code{\link{show_pars}}, \code{\link{moderr}}
}
