% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktsmod.R
\name{ktsmod}
\alias{ktsmod}
\title{Construct two-site kinetic desorption models}
\usage{
ktsmod(
  oec,
  oec_vars,
  pars,
  units = "default",
  opt_est = "default",
  plot_units = c(x = "time", y = "cc0"),
  flowpar = rep(NA, 2),
  ro_h2o = NA,
  tmax = NA,
  qmax = NA,
  cumulative = TRUE,
  mass_flow = FALSE,
  draw = TRUE,
  optmet = "nlopt",
  silent = FALSE
)
}
\arguments{
\item{oec}{A \code{data.frame} of OEC data. Must include all \code{oec_vars}.}

\item{oec_vars}{A \strong{named} \code{character} vector of column names included in \code{oec}.
These \strong{must} include the extraction time (\code{"x"}) and response (\code{"y"}) with appropriate \code{units}.
A third parameters indicating the solvent usage (\code{"slv"}) is \strong{optional} and may be provided as an
\strong{alternative} to an explicit flow rate normally provided in \code{pars["flow"]} - this is useful when the
flow rate is uneven throughout the extraction process.}

\item{pars}{A named \code{numeric} vector of input parameters for the model. \strong{Mandatory} parameters
include the pressure (\code{"pres"}; bar), temperature (\code{"temp"}; degrees Celsius), mass of raw material
loaded (\code{"m_in"}; g), and the maximum possible yield (\code{"c0"}; units identical to those given in \code{units}).
\strong{Optional} parameters include the fraction of easily-desorbed solute (\code{"f"}) and the flow rate of water
(\code{"flow"}; units given in \code{units}).}

\item{units}{A \strong{named} \code{character} vector of length 1 or 2 specifying the units of \code{"flow"}
(one of \code{"mL/min"}, the \strong{default} \code{"g/min"}, \code{"kg/h"}, \code{"L/h"}, or \code{"none"}) and/or
the response \code{"resp"} (one of \code{"g"}, the \strong{default} \code{"percent"}, \code{"permille"}, \code{"ppm"},
or \code{"ppb"}). Where not provided, default values are used.}

\item{opt_est}{Either \code{"default"} or a \strong{named} \code{numeric} vector of initial parameter estimates
for iterative optimization. May include any of the first-order rate constants \eqn{k_1} and \eqn{k_2} (\code{"k1"}
and \code{"k2"}), and/or the fraction of easily-desorbed solute \eqn{F} (\code{"f"}, \strong{only required if
it is not explicitly provided in \code{pars}}). If not provided, the following default values are used:
\code{c(k1 = 0.1, k2 = 0.1, f = 0.5)}.}

\item{plot_units}{A \strong{named} \code{character} vector of length 2 specifying which units to use for the x- and y-axis
of the model plot(s). Possible values are \code{c("time", "q")} for element \code{["x"]} and
\code{c("abs", "cc0")} for element \code{["y"]}.}

\item{flowpar}{Either \code{NA} (default) or a \code{numeric} vector of length 2 providing temperature and pressure
at which flow rate of water is measured.}

\item{ro_h2o}{The subcritical water density (in g/L) at extraction conditions. If not provided (\code{NA}; default),
it is calculated via the IAPWS-95 formulation (see \code{\link{h2o_dens}}).}

\item{tmax, qmax}{Maximum x-axis value (time, min for \code{tmax} and S/M ratio for \code{qmax}) to use for model predictions.
Setting to \code{NA} defaults to 120\% of the maximum experimental value.}

\item{cumulative}{A \code{logical} switch specifying whether the \strong{response} and/or \strong{solvent consumption}
values provided in \code{oec} are cumulative or not (defaults to \code{FALSE}).}

\item{mass_flow}{A \code{logical} indicating whether the flow rate provided in \code{pars} is \strong{mass}
or \strong{volumetric} (\code{FALSE}; default).}

\item{draw}{A \code{logical} switch. Should generated plots be plotted? Defaults to \code{TRUE}.}

\item{optmet}{The method of iterative curve fitting to use for estimating \eqn{k_1}, \eqn{k_2}, and (optionally) \eqn{F}.
One of \code{"nlopt"} (Non-Linear Optimization via \code{\link[nloptr]{nloptr}}; default) or \code{"nlrob"} (Robust Fitting
via \code{\link[robustbase]{nlrob}}).}

\item{silent}{Should console output be silenced? Defaults to \code{FALSE}).}
}
\value{
A named \code{list} with the following elements:
\enumerate{
\item \strong{$data}: A \code{data.frame} including time, response, S/M ratio (\code{$q}), mass yield of extract (\code{"yield_g"}),
and fractional yield relative to the total amount of extractable solute (\code{"yield_cc0"}).
\item \strong{$tws}: The summarized TWS model output from function \code{\link{tws_cmp}}.
\item \strong{$plots}: The model visualisation (plot) of class \code{"ggplot"}.
\item \strong{$input}: A list of input model parameters including pressure (\code{"pres"}), temperature
(\code{"temp"}), flow rate (\code{"flow"} \strong{in kg/s}), the maximum possible yield (\code{"c0"}),
mass of loaded material (\code{"m_in"}), and the number of experimental observations (\code{"m"}).
\item \strong{$call}: The function call.
}
}
\description{
Derives, evaluates, and visualizes the results of two-site kinetic desorption (TWS) models.
Useful for subcritical water extraction (SWE) process evaluation. See \strong{Details} for further information.
}
\details{
The TWS model is generated given the vector of time data \eqn{t} (e.g. extraction time), first-order rate constants
\eqn{k_1} and \eqn{k_2} (in \eqn{min^{-1}}) describing the "fast" and "slow" component of kinetic desorption, and
the fraction \eqn{F} of easily-desorbed target compound(s). The equation describing fractional yield \eqn{e}
(ratio between extract collected at time \eqn{t} and the maximum amount of extractable material) is:
\deqn{e = 1 - [F \times e^{-k_1t}] - [(1 - F) \times e^{-k_2t}]}
See \strong{References} for example applications of the TWS model for subcritical water extraction (SWE) studies.
}
\examples{
twosite_res <- ktsmod(oec = swex[["duba1"]][["data"]],
oec_vars = c(x = "Time_min", y = "Yield_100C"),
pars = c(pres = 15, temp = 100, flow = 2, c0 = 77, m_in = 2, f = 0.24),
opt_est = "default",
units = c(flow = "mL/min", resp = "permille"),
plot_units = c(x = "q", y = "abs"), #"time" "q" "abs" "cc0"
cumulative = TRUE,
mass_flow = FALSE,
flowpar = rep(NA,2),
ro_h2o = NA,
tmax = NA,
qmax = NA,
optmet = "nlopt")

}
\references{
Abidin, Z.Z., Samadi, M., Biak, D.R.A., Yunus, R. (2024), 'Mathematical Modelling Of Extraction Of Oil From \emph{Aquilaria malacenssis} Wood Employing Subcritical Conditions', \emph{Journal of Applied Science and Engineering} \strong{27} (12), pp. 3725-3738, DOI: \url{http://dx.doi.org/10.6180/jase.202412_27(12).0012}.

Anepankul, T., Goto, M., Sasaki, M., Pavasant, P., Shotipruk, A. (2007), 'Extraction of anti-cancer damnacanthal from roots of \emph{Morinda citrifolia} by subcritical water', \emph{Separation and Purification Technology} \strong{55}, pp. 343-349, DOI: \url{http://dx.doi.org/10.1016/j.seppur.2007.01.004}.

Jamaludin, R., Kim, D.-S., Salleh, L.M., Lim, S.-B. (2021), 'Kinetic Study of SubcriticalWater Extraction of Scopoletin, Alizarin, and Rutin from \emph{Morinda citrifolia}', \emph{Foods} \strong{2021} (10), article 2260, DOI: \url{https://doi.org/10.3390/foods10102260}.

Pereira, D.T.V., Tarone, A.G., Cazarin, C.B.B., Barbero, G.F., Martinez, J. (2019), 'Pressurized liquid extraction of bioactive compounds from grape marc', \emph{Journal of Food Engineering} \strong{240}, pp. 105-113, DOI: \url{https://doi.org/10.1016/j.jfoodeng.2018.07.019}.
}
\seealso{
\code{\link{tws_cmp}}, \code{\link{kin_splot}}, \code{\link{show_pars}}, \code{\link{moderr}}, \code{\link{bicmod}}
}
