% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{load_internal}
\alias{load_internal}
\title{Load internal package data where necessary}
\usage{
load_internal(dtname)
}
\arguments{
\item{dtname}{Name of dataset. One of \code{c("doe_base", "doe_lst1", "doe_lst2", "gcm_cnt", "gcm_int",
"gcm_smarts", "solv_dmass", "solv_dmol", "solv_mv")}}
}
\value{
The selected internal data.
}
\description{
Load internal package data where necessary
}
\examples{
\dontrun{
load_internal("doe_lst1")
}
}
