% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp_main.R
\name{miscomp}
\alias{miscomp}
\title{Compare HSP-based miscibility enhancement for two or more solutes}
\usage{
miscomp(
  sols,
  tb = "SB_corr",
  crit = "NL07_robust",
  hsp = "SP12",
  modif = "Ethanol",
  modfracs = NA,
  pres = seq(100, 600, 100),
  pres_comp = pres,
  cols = "default",
  plt_title = TRUE,
  temp = 40,
  vfrac = seq(0.05, 0.4, 0.05),
  simplicity = "auto",
  draw = TRUE,
  silent = FALSE
)
}
\arguments{
\item{sols}{A \code{list} where each element provides information for a solute as required by \code{\link{mol_find}}.}

\item{tb, crit, hsp}{All \code{character} values of GCM methods to use. For available methods, see \code{\link{est_gcm}}.
The length of each element must either be 1 or equal to that of \code{sols}.}

\item{modif}{A \code{character} value or vector of the co-solvent (or solvent blend) to use.
For possible values, see \code{\link{sfe_mod}}.}

\item{modfracs}{A vector or \code{list} of co-solvent volume fractions to use when a \strong{solvent blend} is specified
in \code{modif}.}

\item{pres}{A \code{numeric} vector of \strong{up to 6} pressures (in bar) to evaluate miscibility enhancement at.}

\item{pres_comp}{A \code{numeric} pressure value to use for comparative plots between solutes. Must also be present in \code{pres}.}

\item{cols}{Either \code{"default"} or a \strong{named} vector of colours to use for up to six values of pressure (given in \code{pres}).
Possible names are \code{"one"}, \code{"two"}, \code{"three"}, \code{"four"}, \code{"five"}, and/or \code{"six"}.}

\item{plt_title}{A \code{logical} indicating whether titles should be added to plots.}

\item{temp}{The single \code{numeric} temperature value to use (between 40-70 Celsius).}

\item{vfrac}{A \code{numeric} vector of volume fractions of co-solvent to evaluate. Defaults to \code{seq(0.05, 0.40, 0.05)}.}

\item{simplicity}{A \code{character} string indicating whether overlapping sub-structures should be allowed (\code{"auto"} by default).
See \code{\link{sub_smarts}}.}

\item{draw}{A \code{logical} indicating whether generated plots should be printed automatically (\code{TRUE} by default).}

\item{silent}{A \code{logical}. When \code{FALSE} (default), additional information is printed in the console.}
}
\value{
A named \code{list} containing the output \code{data.frame} (\code{$data}) with miscibility enhancement
at all values of pressure and co-solvent volume fractions, and corresponding visualizations (\code{$plots}).
}
\description{
Uses the \code{\link{sfe_mod}} framework to visually compare the miscibility enhancement provided by a chosen co-solvent
for at least two solutes at various pressures (up to 6) and a set temperature.
}
\examples{
#Get information about solute 1 (beta-carotene)
mol1 <- c("CC1=C(C(CCC1)(C)C)C=CC(=CC=CC(=CC=CC=C(C)C=CC=C(C)C=CC2=C(CCCC2(C)C)C)C)C",
"7235-40-7", "Beta-carotene")

#Get information about solute 2 (xanthohumol)
mol2 <- c(hspex[24,c("SMILES", "CAS", "Name")])

#Compare miscibility enhancement with ethanol as a co-solvent
gcm_comp <- miscomp(sols = list(mol1, mol2))

}
\seealso{
\code{\link{sfe_mod}}
}
