% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{moderr}
\alias{moderr}
\title{Calculate kinetic model errors (AARD, RMSE, R2)}
\usage{
moderr(act, pred, m = length(act))
}
\arguments{
\item{act, pred}{Both \code{numeric} vectors of actual and model-predicted values \strong{of equal length}.}

\item{m}{\strong{Optional} number of data points (by default, equals to the length of \code{act}).}
}
\value{
A named \code{numeric} vector containing the calculated AARD (\code{"aard"}), RMSE (\code{"rmse"}), and R2 (\code{"r2"}) values.
}
\description{
Calculates the Average Absolute Relative Deviation (AARD), Round-Mean Squared Error (RMSE), and the R2 (R-Squared) value
of any model given predicted and actual values alongside the number of recorded observations (data points).
}
\details{
The Average Absolute Relative Deviation (\eqn{AARD}) is defined as:
\deqn{AARD = 100 \times \sum{\vert (x_{pred} - x_{act})/x_{act}\vert}/m}

The RMSE and R2 value are defined by:
\deqn{RMSE = \sum{\sqrt{(x_{pred}-x_{act})^2}/m}}
\deqn{R2 = 1 - [\sum{(x_{act} - x_{pred})^2}]/[\sum{(x_{act} - \bar{x}_{act})^2}]}

In all the above equations, \eqn{m} is the number of observations, while \eqn{x_{pred}} and \eqn{x_{act}} are predicted
and actual response values, respectively. An overline denotes the mean value.
}
\seealso{
\code{\link{bicmod}}, \code{\link{ktsmod}}, \code{\link{bic_sm}}, \code{\link{bic_ct}}, \code{\link{bic_cmp}}
}
