% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_find.R
\name{mol_find}
\alias{mol_find}
\title{Get MOL from SMILES and derive molecular descriptors}
\usage{
mol_find(mol)
}
\arguments{
\item{mol}{A \code{character} vector of \code{length} 1-3 containing the \strong{mandatory CAS number and name}
as well as \strong{optional SMILES string} of a given molecule.
The vector may \strong{optionally} be named with \code{"Name"}, \code{"SMILES"},
and/or \code{"CAS"}. Otherwise the function will attempt automatic assignment based on values.}
}
\value{
A \code{list} of length 3 containing the following elements:
\enumerate{
\item \strong{IDs}: A named vector of descriptors, including SMILES, CAS number, Name, InChI, InChIKey,
the number of atoms (\code{"Atom_Count"}), number of hydrogens (\code{"H_Count"}),
molecular weight (\code{"MW"}), molecular formula (\code{"MF"}), and a \code{logical}
specifying whether the molecule is aromatic (\code{"Aromaticity"}).
\item \strong{Molfile}: A representation of the molecule in MOL format.
\item \strong{Synonyms}: A character vector of alternative names found in PubChem (up to 5).
}
}
\description{
Converts a SMILES string and (optionally) other information from SMILES to MOL.
Part of the \code{\link{sfe_mod}} workflow.
}
\examples{
#Limonene input vector
mol <- c("CC1=CCC(CC1)C(=C)C", "5989-54-8", "Limonene")
#Get descriptors
res <- mol_find(mol)

}
