% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{oec_bp}
\alias{oec_bp}
\title{Run segmented regression to estimate CER endpoint}
\usage{
oec_bp(input, x, y, plt = TRUE, segmode = "step")
}
\arguments{
\item{input}{A \code{data.frame} containing the time \code{x} and \code{y} values of the extraction curve.}

\item{x, y}{Both are \code{character} column names identifying x- and y-values in the \code{input} data.}

\item{plt}{A \code{logical} switch specifying whether to plot segmented regression results (defaults to \code{TRUE}).}

\item{segmode}{Segmentation mode to use. One of: \code{"step"} (default) or \code{"seg"}.}
}
\value{
A single \code{numeric} value showing the observation number at which the CER likely ends.
}
\description{
Estimates the number of observation along an extraction curve
using \code{\link[segmented]{segreg}} or \code{\link[segmented]{stepreg}}.
May be used in preparation for the \code{\link{bicmod}}.
}
\examples{
oec_bp(input = sfex[[1]][["data"]], x = "time_min", y = "yield_g")

}
\seealso{
\code{\link{bicmod}}
}
