% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{plot_gcm}
\alias{plot_gcm}
\title{Visualize molecules and sub-structures}
\usage{
plot_gcm(solute, smarts, hlight = TRUE, draw_plot = TRUE, plot_title = TRUE)
}
\arguments{
\item{solute}{MOL file and molecule descriptors output from function \code{\link{mol_find}}.}

\item{smarts}{A \code{character} vector of SMARTS strings specifying substructures to be highlighted.
Only relevant if \code{hlight} is \code{TRUE}.}

\item{hlight}{A \code{logical} indicating whether substructures are highlighted.}

\item{draw_plot}{A \code{logical} determining whether the plot is drawn.}

\item{plot_title}{A \code{logical} determining whether the plot should be annotated.}
}
\value{
The molecule visualization as an \code{array}.
}
\description{
Parts of the \code{\link{sfe_mod}} workflow.
Visualizes a molecule from a MOL file and optionally highlights SMARTS substructures.
}
\examples{
smarts <- c("[CX4H3]", "[CX3H2]", "[$([!R;#6X3H0]);!$([!R;#6X3H0]=[#8])]", "[R;CX4H2]", "[R;CX4H]",
"[R;CX3H1,cX3H1]", "[$([R;#6X3H0]);!$([R;#6X3H0]=[#8])]")
mol <- mol_find(c("CC1=CCC(CC1)C(=C)C", "5989-54-8", "Limonene"))
res <- plot_gcm(mol, smarts, draw_plot = TRUE)

}
\seealso{
\code{\link{mol_find}}, \code{\link{sfe_mod}}
}
