% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{predict_bic}
\alias{predict_bic}
\title{Predict response using built BIC models}
\usage{
predict_bic(input, newdata, units = "sm", get_yields = TRUE)
}
\arguments{
\item{input}{A BIC model output from \code{\link{bicmod}}.}

\item{newdata}{A \code{numeric} value or vector of new data (extraction time or solvent-to-insoluble solid ratio)
to make predictions for.}

\item{units}{Specifies the units of the input data (\code{newdata}). One of \code{"time"} (extraction time in minutes)
or the default \code{"sm"} (solvent-to-insoluble solid material ratio, dimensionless).}

\item{get_yields}{A \code{logical} specifying whether mass and percentage yields should be calculated from the default
prediction format (g/g insoluble solid, i.e. \strong{fractional yield}).}
}
\value{
A \code{list} of length 3 containing the model \code{$predictions}, a helpful \code{$unit_chart} showing the units
pertaining to the predictions, and a \code{$description} providing some additional details about the predictions (such as the
process conditions which they are valid for).
}
\description{
Predicts responses from new temporal data using BIC (Broken-and-Intact Cells) models
built via \code{\link{bicmod}}.
}
\seealso{
\code{\link{bicmod}}
}
