% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{predict_doe}
\alias{predict_doe}
\title{Predict response for new factor values from built RSM models}
\usage{
predict_doe(input, newdata, coded = FALSE)
}
\arguments{
\item{input}{The models output from \code{\link{doe_analyze}}. Includes initial and final (i.e. simplified) models.}

\item{newdata}{New data to predict the response for. Either a \code{numeric} vector or a \code{data.frame}.}

\item{coded}{A \code{logical} indicating whether the provided \code{newdata} is coded or not (defaults to \code{FALSE}).}
}
\value{
A \code{list} containing \code{data.frame} objects with coded and uncoded factors provided in \code{newdata},
and their associated predicted response values. The predicted responses are also provided in a second list element named
\code{$predictions}.
}
\description{
A simple wrapper for the predictor function. Works with models output from \code{\link{doe_analyze}}.
}
\examples{
\dontrun{
#Build RSM model
doe_optres <- doe_analyze(doe = doex[["ccd3"]][["data"]],
uc_facs = c("P_bar", "T_degC", "EtOH_gmin"),
cent_id = NA,
resp_var = "ExtYield",
time_var = "Actual_Order",
mod_order = 2,
canon_thres = "auto",
p_cutoff = 0.10,
trim_method = "both",
which_facs = "coded",
export = "none",
verbose = TRUE)

#Predict response for new factor values
new_preds <- predict_doe(doe_optres, newdata = c(A = -0.67, B = 0, C = 0.67), coded = TRUE)

}
}
