% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{predict_kts}
\alias{predict_kts}
\title{Predict responses using built KTS desorption models}
\usage{
predict_kts(input, newdata, get_yields = TRUE, moisture = NA)
}
\arguments{
\item{input}{The output of \code{\link{ktsmod}} containing model information.}

\item{newdata}{A single \code{numeric} value or a vector of new input data to made predictions for
(extraction time, in minutes).}

\item{get_yields}{A \code{logical} value specifying whether the default yield should be converted into mass
and percentage yield. Defaults to \code{TRUE}.}

\item{moisture}{An optional \code{numeric} value showing the moisture content of the raw material
in percent fresh weight. Only used to correct percentage yield predictions when \code{get_yields == TRUE}. Defaults to \code{NA}.}
}
\value{
A \code{list} of length 3 containing the \code{$predictions} as a \code{data.frame}, the \code{$input} parameters,
and a \code{$description} outlining which process parameters the model is valid for along with other useful information.
}
\description{
Predicts the response for new input values using two-site kinetic desorption models
built via \code{\link{ktsmod}}.
}
\seealso{
\code{\link{ktsmod}}
}
