% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rowVars}
\alias{rowVars}
\title{Calculate row-wise variances}
\usage{
rowVars(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} for which to calculate row-wise variances.}

\item{na.rm}{A \code{logical} indicating whether to ignore \code{NA} values (\code{FALSE} by default).}
}
\value{
A \code{numeric} vector of variances (one value per row of \code{x}).
}
\description{
Calculate row-wise variances
}
