% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scap}
\alias{scap}
\title{Capitalize first letter of every word in string}
\usage{
scap(x)
}
\arguments{
\item{x}{A \code{character} string.}
}
\value{
The input character string with every word (separated by a space) capitalized.
}
\description{
Capitalize first letter of every word in string
}
\examples{
scap("every word of this sentence should now be capitalized")
}
