% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seq_last}
\alias{seq_last}
\title{Get sequence including the last number}
\usage{
seq_last(x, y, z)
}
\arguments{
\item{x, y}{Starting and ending points of the sequence.}

\item{z}{Increment of the sequence.}
}
\value{
A \code{numeric} vector containing the sequence inclusive of the last number.
}
\description{
Get sequence including the last number
}
\examples{
seq_last(80,310,20)
}
