% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp_main.R
\name{sfe_mod}
\alias{sfe_mod}
\title{Compare HSP-based utility of various co-solvents for SFE}
\usage{
sfe_mod(
  solute,
  tb,
  crit,
  hsp,
  gorder = 0,
  modif = "all",
  modfracs = NA,
  hlight = TRUE,
  simplicity = "auto",
  pres = seq_last(80, 400, 20),
  temps = seq_last(32, 70, 3),
  vfrac = 0.1,
  silent = FALSE
)
}
\arguments{
\item{solute}{Information about the solute as required by function \code{\link{mol_find}}.}

\item{tb, crit, hsp}{All are \code{character} values specifying the methods to use for GCMs (see \code{\link{est_gcm}}).}

\item{gorder}{A \code{numeric} value denoting the maximum order of groups to be considered. Defaults to \code{0}, which
considers all available group orders. May also be a \strong{named} vector - see \code{\link{est_gcm}} for details.}

\item{modif}{A \code{character} value, vector, or \code{list} indicating which co-solvent(s) to evaluate alongside pure CO2.
Each element of vector must be either a solvent \strong{name} or its \strong{abbreviation}. For possible values,
consult the output of \code{\link{show_solv}}. \strong{Mixtures of co-solvents} may
also be evaluated if a \code{list} is provided, where a mixture may be specified in each list element as a vector of length 2
or more containing the \strong{names of solvents}.}

\item{modfracs}{An \strong{optional} \code{list} of solvent mixture volume percentages to use for solvent mixtures specified in
\code{modif}. Percentages given in each list element must either be less than or add up to 100 and contain an equal number of
percentage values to the number of solvents in the corresponding mixture (or one less value, in which case the remainder is
made up to a total of 100). The total number of list elements must equal the number of mixtures specified in \code{modif}.}

\item{hlight}{A \code{logical} indicating whether substructures are highlighted (see \code{\link{plot_gcm}}).}

\item{simplicity}{A \code{character} string indicating whether overlapping sub-structures should be allowed (\code{"auto"} by default).
See \code{\link{sub_smarts}}.}

\item{pres, temps}{Both are \code{numeric} values \strong{or sequences} of pressures (75-1000 bar) and temperatures (32-200 Celsius)
at which to evaluate solute and solvent HSPs. Defaults are \code{seq(80,300,20)} and \code{seq(32,65,3)} for pressure and
temperature, respectively.}

\item{vfrac}{The volume fraction of co-solvent to use (defaults to \code{0.10}, or 10\%).}

\item{silent}{A \code{logical}. When \code{FALSE} (default), additional information is printed in the console.}
}
\value{
A named \code{list} with the following elements:
\describe{
\item{solute_ids}{Solute information as retrieved by \code{\link{mol_find}}.}
\item{parameters}{Solute parameters (boiling point, critical parameters, and HSPs)
estimated by GCM via function \code{\link{est_gcm}}.}
\item{sfe}{Results of miscibility enhancement assessment as output from \code{\link{hsp_optim}}.}
\item{gcm_vis}{A \code{list} of raster visualisation included as part of \code{\link{est_gcm}} output.}
\item{call}{The function \code{call}.}
}
}
\description{
Uses group contribution methods to estimate the boiling point, critical temperature, and/or Hansen Solubility Parameters
of a solute given its SMILES string and MOL file (for example, output from \code{\link{mol_find}}). The estimated parameters are then
used to calculate \strong{solubility parameter distance \eqn{R_a}} between the solute and both pure carbon dioxide and its volumetric
mixture with a given co-solvent. These distances may be compared for various co-solvents to determine the optimal solvent throughout
a range of supercritical CO2 pressures and temperatures.
}
\details{
The workflow is based on various group contribution methods (see \code{\link{est_gcm}} for relevant sources), the
Hansen Solubility Theory (Hansen, 2007), and the work of Tirado et al. (2018, 2019) and Diego & Calvo (2019). Given the SMILES string
and molecular geometry (MOL file) of a solute, the boiling point, critical temperature, and Hansen Solubility Parameters
(HSPs) are estimated via GCMs. For pure CO2 and co-solvents, these values as well as the molar volume (mL/mol)
were compiled from CoolProp (Bell et al., 2014). The influence of temperature on \strong{solute} HSPs is then calculated
from \strong{reduced temperatures} via the following equation where \eqn{T_{r2}} and \eqn{T_{r1}} are given by
\eqn{T_{actual}/T_{critical}} at the desired temperature and 298.15 K, respectively:
\deqn{\delta_2 = (\frac{1-T_{r2}}{1-T_{r1}})^{0.34}\times\delta_1}
The effects of temperature on dispersion (\eqn{\delta_d}), polarity (\eqn{\delta_p}), and hydrogen bonding (\eqn{\delta_{HB}})
HSPs are calculated via the following equations:
\deqn{\delta_{d\text{ }ref}/\delta_d = (V_{ref}/V)^{-1.25}}
\deqn{\delta_{p\text{ }ref}/\delta_p = (V_{ref}/V)^{-0.5}}
\deqn{\delta_{HB\text{ }ref}/\delta_{HB} = exp[-1.32\times10^{-3}\times(T_{ref}-T)-ln(V_{ref}/V)^{-0.5}]}
The HSP distance \eqn{R_a} between the solvent (pure CO2 and/or CO2+co-solvent mixture, subscript 1) and solute
(subscript 2) is then derived using:
\deqn{R_a = \sqrt{4\times(\delta_{d1} - \delta_{d2})^2 + (\delta_{p1} - \delta_{p2})^2 + (\delta_{HB1} - \delta_{HB2})^2}}
The above calculations are repeated for every combination of pressure and temperature. The miscibility enhancement (\eqn{ME}, in \%)
is finally derived as the percentage ratio between the solute \eqn{R_a} to that of CO2+co-solvent and pure CO2:
\deqn{ME\text{ }(\%) = (1-(R_{a\text{ }scCO_2\text{ }+\text{ }co-solvent}/R_{a\text{ }pure\text{ }scCO_2}))\times 100}
Marking the miscibility enhancement values at all combinations of pressure and temperature allows the determination of the best-suited
co-solvent for any given solute.
}
\examples{
#Retrieve molecule (beta-carotene)
mol <- c("CC1=C(C(CCC1)(C)C)C=CC(=CC=CC(=CC=CC=C(C)C=CC=C(C)C=CC2=C(CCCC2(C)C)C)C)C",
"7235-40-7", "Beta-carotene")

#Assess SFE miscibility enhancement with various co-solvents
optres <- sfe_mod(solute = mol,
tb = "SB_corr",
crit = "NL07_robust",
hsp = "SP12")

}
\references{
Bell, Ian H., Wronski, Jorrit, Quoilin, Sylvain, Lemort, Vincent (2014), 'Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and the Open-Source Thermophysical Property Library CoolProp', \emph{Industrial & Engineering Chemistry Research} \strong{53} (6), pp. 2498-2508, DOI: \url{https://doi.org/10.1021/ie4033999}.

Hansen, Charles M. (2007), \emph{Hansen Solubility Parameters: A User's Handbook (2nd edition)}, CRC Press, London, United Kingdom.

Tirado, Diego F., Tenorio, Maria Jose, Cabanas, Albertina, Calvo, Lourdes (2018), 'Prediction of the best cosolvents to solubilise fatty acids in supercritical CO2 using the Hansen solubility theory', \emph{Chemical Engineering Science} \strong{190}, pp. 14-20, DOI: \url{https://www.doi.org/10.1016/j.ces.2018.06.017}.

Tirado, Diego F., Rousset, Amandine, Calvo, Lourdes (2019), 'The Selective Supercritical Extraction of High-value Fatty Acids from Tetraselmis suecica using the Hansen Solubility Theory', \emph{Chemical Engineering Transactions} \strong{75}, pp. 133-138, DOI: \url{https://www.doi.org/10.3303/CET1975023}.

Tirado, Diego F., Calvo, Lourdes (2019), 'The Hansen theory to choose the best cosolvent for supercritical CO2 extraction of beta-carotene from \emph{Dunaliella salina}', \emph{The Journal of Supercritical Fluids} \strong{145}, pp. 211-218, DOI: \url{https://www.doi.org/10.1016/j.supflu.2018.12.013}.
}
\seealso{
\code{\link{mol_find}}, \code{\link{est_gcm}}, \code{\link{plot_gcm}}, \code{\link{hsp_optim}}, \code{\link{miscomp}}, \code{\link{compare_gcm}}, \code{\link{show_solv}}
}
