% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfex}
\alias{sfex}
\title{Example Supercritical Fluid Extraction (SFE) kinetic modeling data}
\format{
An object of class \code{list} of length 4.
}
\usage{
sfex
}
\description{
A named nested \code{list} containing extraction curve and Broken-and-Intact Cells (BIC) kinetic model data from scientific literature.
Each element is itself a \code{list} including the following elements:
\enumerate{
\item \code{$data}: A \code{data.frame} containing the OEC data (time, response and, optionally, amount of solvent expended).
\item \code{$add_pars}: Additional vector of \code{logical} parameters specifying whether the data is cumulative and whether the flow units are mass or volumetric.
\item \code{$units}: A named vector of units for the response (\code{["resp"]}) and flow (\code{["flow"]}).
\item \code{$ext_pars}: A vector of extraction parameters such as pressure and temperature.
\item \code{$input_pars}: Input model parameters (use \code{\link{show_pars}} for description and units of all parameters).
\item \code{$adj_pars}: Adjustable model parameters (use \code{\link{show_pars}} for description and units of all parameters).
\item \code{$proc_desc}: A short description of the modeled process.
\item \code{$citation}: A reference for the source publication.
}

Currently, the following Overall Extraction Curves (OEC) and BIC model results are included for SFE of:
\enumerate{
\item \code{$dimic}: Cherry seed oil (Dimic et al., 2021).
\item \code{$rizza}: Lipids from microalgae (Rizza, 2014).
}
}
\references{
Dimic et al. (2021), 'Supercritical Fluid Extraction Kinetics of Cherry Seed Oil: Kinetics Modeling and ANN Optimization',
\emph{Foods} \strong{10}, article 1513, DOI: \url{https://doi.org/10.3390/foods10071513}.

Rizza (2014), Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae, Universita Degli Studi di Padova,
Padua, Italy, MSc thesis.
}
\seealso{
\code{\link{show_pars}}
}
\keyword{datasets}
