% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{show_gcm}
\alias{show_gcm}
\title{GCM method selection chart}
\usage{
show_gcm()
}
\value{
A \code{data.frame} showing various Group Contribution Methods (GCMs) usable with \code{\link{est_gcm}} and parameters
which they are able to estimate (marked with \code{"Y"}).
}
\description{
GCM method selection chart
}
\seealso{
\code{\link{est_gcm}}
}
