% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{show_pars}
\alias{show_pars}
\title{Show kinetic model parameters and their description}
\usage{
show_pars(type)
}
\arguments{
\item{type}{The type of model to show parameters for. One or more of: \code{"bic"} (BIC model),
\code{"ts"} (two-site kinetic desorption model),
and/or \code{"com"} (Cost of Manufacturing).}
}
\value{
A \code{data.frame} with types, names, units, and descriptions of model parameters.
}
\description{
Displays names, units, and descriptions of kinetic model and Cost of Manufacturing
calculation parameters implemented in \pkg{supeRcrit}.
}
\examples{
#Show parameter list for the Broken and Intact Cells (BIC) model
show_pars("bic")
}
