% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{show_solv}
\alias{show_solv}
\title{HSP data for various solvents}
\usage{
show_solv()
}
\value{
A \code{data.frame} of solvent data to be used for HSP-based calculations of relative solubility improvement. The columns include
solvent name (\code{"Solvent"}), abbreviation for function arguments (\code{"Abbreviation"}), the \code{"CAS"} number, molecular weight (\code{"MW"}),
molar volume in mL/mol (\code{"MV"}), density in g/mL (\code{"RHO"}), Hansen Solubility Parameters (dispersion \code{"dD"}, dipole moment \code{"dP"}, and hydrogen bonding \code{"dH"}),
boiling point temperature (\code{"Tb"}), triple point temperature (\code{"Ttp"}), as well as the critical temperature (\code{"Tc"}), pressure (\code{"Pc"}, in MPa),
volume (\code{"Vc"}, in L/mol), density (\code{"RHOc"}, in mol/L), and acentric factor (\code{"Omega"}). All temperatures are in Kelvin.
}
\description{
Displays the Hansen Solubility Parameters (HSP), molecular weight, molar volume, boiling point, critical temperature and pressure,
as well as the molar density and triple point temperature of various solvents used in the \code{\link{sfe_mod}} workflow.
}
