% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{solv_cmp}
\alias{solv_cmp}
\title{Equation: Relative amount of expended solvent considering only the CER and DC of the complete BIC model}
\usage{
solv_cmp(x, ys, thetae, xu, betam, r, qm, gam, thetai)
}
\arguments{
\item{x}{A \code{numeric} vector of relative amounts of expended solvent (\eqn{q}; kg/kg insoluble solid) to evaluate.}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}

\item{thetae}{External material (mass) transport resistance \eqn{\theta_e}.}

\item{xu}{Weight fraction (concentration) in the untreated solid \eqn{x_u}.}

\item{betam}{Coefficient that simplifies the form of the equation (\eqn{\beta}; see \strong{Eq. 3.26 in Rizza (2014)} for definition).}

\item{r}{Grinding efficiency \eqn{r} (fraction of broken cells).}

\item{qm}{Relative amount of expended solvent (\eqn{q_m}; kg/kg insoluble solid) at the end of the CER.}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{\gamma \text{ in } kg_{solvent}/kg_{insoluble solid}}).}

\item{thetai}{Internal material (mass) transfer resistance \eqn{\theta_i}.}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, the relative amount of expended solvent
\eqn{q} (kg/kg insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The equation used to calculate the relative amount of expended solvent \eqn{q_{cmp}} considering only the Constant Extraction Rate (CER) and
Diffusion-Controlled (DC) periods of the overall extraction curve is (Rizza, 2014):

\deqn{q_{cmp} = x_u \times [1 - \beta \times \ln{\{1 + (1 - r) \times [\exp{(1/\beta)}-1] \times \exp{(\frac{q - q_m}{\gamma \times \theta_i})}\}}] - q \times y_s \times (1 - \exp{(-1/\theta_e)})}

This is a combination of \strong{Eqs. 3.27} and \strong{3.29} as presented by Rizza (2014).
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
