% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{solv_sim}
\alias{solv_sim}
\title{Equation: Relative amount of expended solvent at the end of the CER (simplified BIC model)}
\usage{
solv_sim(x, xu, c1, c2, ys)
}
\arguments{
\item{x}{A \code{numeric} vector of estimates of possible relative amounts of solvent expended at the end of the CER (\eqn{q}; kg/kg CO2).}

\item{xu}{Weight fraction (concentration) in the untreated solid \eqn{x_u}.}

\item{c1, c2}{Constants \eqn{C_1} and \eqn{C_2}.}

\item{ys}{Extract solubility in CO2 (\eqn{y_s}; g/g).}
}
\value{
The single \code{numeric} result of the evaluated equation - specifically, relative amount of expended solvent at the
end of the CER (\eqn{q_m}; kg/kg insoluble solid).
}
\description{
Part of the \code{\link{bicmod}} workflow. \strong{This function manual is hidden from the Index.}
}
\details{
The following equation calculates the relative amount of passed solvent (CO2) at the end of the Constant Extraction Rate (CER) period
(Rizza, 2014):
\deqn{q_m = x_u \times (1 - C_1 \times \exp{(-C_2 \times q)}) - q * y_s}
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}
}
\keyword{internal}
