% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitAt}
\alias{splitAt}
\title{Split vector at specific position(s)}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{An \code{atomic} vector to split.}

\item{pos}{A \code{numeric} position (or vector thereof) at which to split the input vector \code{x}.}
}
\value{
A \code{list} of output vectors resulting from splitting \code{x} at position(s) given in \code{pos}.
}
\description{
Splits an input atomic vector into two or more vectors at specific indices.
Function adapted from a StackOverflow answer (see \strong{References}).
}
\examples{
splitAt(seq(1,5,1), c(2,4))
}
\references{
\href{StackOverflow answer}{https://stackoverflow.com/questions/16357962/r-split-numeric-vector-at-position} by \strong{flodel}.
}
