% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{steepath}
\alias{steepath}
\title{Find path of steepest ascent}
\usage{
steepath(mod_in, uc_facs, c_facs, respname = NA)
}
\arguments{
\item{mod_in}{Model of class \code{"lm"}.}

\item{c_facs, uc_facs}{Coded and uncoded factors as \code{character} vectors.}

\item{respname}{Response name as a \code{character} string.}
}
\value{
A \code{data.frame} containing the path of steepest ascent.
}
\description{
Part of the \code{\link{doe_analyze}} workflow. Calculates the path of steepest ascent.
\strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{doe_analyze}}
}
\keyword{internal}
