% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hsp.R
\name{sub_smarts}
\alias{sub_smarts}
\title{Identify GCM groups in a molecule}
\usage{
sub_smarts(
  solute,
  method,
  gorder = 0,
  simplicity = "auto",
  tlim = 20,
  silent = FALSE
)
}
\arguments{
\item{solute}{Molecular descriptors and MOL file output from function \code{\link{mol_find}}.}

\item{method}{The GCM method to use as a character string. One of:
Joback & Reid (\code{"JR"}), Stein & Brown (\code{"SB"}), their corrected counterparts (\code{"JR_corr"} and \code{"SB_corr"}),
Nannoolal (2004; \code{"NL04"}), Nannoolal (2007; \code{"NL07"}) and its robust counterpart (\code{"NL07_robust"},
Stefanis & Panayiotou (2008 and 2012; \code{"SP08"} and \code{"SP12"}, respectively),
Hukkerikar (2012; \code{"HKR_STW"} and \code{"HKR_SIM"}),
Zhao (2003; \code{"ZHAO"}), Bondi (1964; \code{"BND"}), and Slonimskii (1970; \code{"SLON"}).
See \code{\link{est_gcm}} for details.}

\item{gorder}{A \code{numeric} value denoting the maximum order of groups to be considered. Defaults to \code{0}, which
considers all available group orders.}

\item{simplicity}{A \code{character} string indicating the level of fragmentation simplicity. One of:
\code{"simple"} (overlapping patterns are allowed), \code{"normal"} (overlaps not allowed with the first fragmentation accepted),
or \code{"complex"} (overlaps not allowed with every possible fragmentation pattern computed).
When set to \code{"auto"} (default), selects the most appropriate simplicity level based on \code{method}.}

\item{tlim}{A single \code{numeric} value setting the time limit (in seconds) of fragmentation iterations
when \code{simplicity} is set to \code{"complex"}. Exceeding the time limit without completing fragmentation
results in an error. Computing all possible fragmentation patterns for some molecules may take an impractically
long time.}

\item{silent}{Should console output be silenced? Defaults to \code{FALSE}).}
}
\value{
A \code{data.frame} of all GCM group contributions for the chosen \code{method} (\code{$contrib_data}),
corresponding SMARTS strings (\code{$smarts_data}), group occurrences in the input molecule (\code{$occurrences}),
and additional settings (\code{$method} and \code{$simplicity}). Also included is a \code{data.frame} of atom indices for
each group match (\code{$atoms}), and the function \code{call}.
}
\description{
Part of the \code{\link{sfe_mod}} workflow. Identifies GCM groups in a solute provided SMARTS substrings.
}
\examples{
mol <- mol_find(c("CC1=CCC(CC1)C(=C)C", "5989-54-8", "Limonene"))
res <- sub_smarts(mol, "JR")

}
\seealso{
\code{\link{mol_find}}, \code{\link{sfe_mod}}, \code{\link{est_gcm}}
}
