% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supeRcrit-package.R
\docType{package}
\name{supeRcrit-package}
\alias{supeRcrit}
\alias{supeRcrit-package}
\title{supeRcrit: Supercritical CO2 and Subcritical Water Process Development}
\description{
Provides tools for kinetic modeling, systematic co-solvent choice, process optimization via experimental design and response surface methodology (RSM), and Cost of Manufacturing (COM) analysis of supercritical CO2 and subcritical water processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Deniz-Koseoglu/supeRcrit}
  \item Report bugs at \url{https://github.com/Deniz-Koseoglu/supeRcrit/issues}
}

}
\author{
\strong{Maintainer}: Deniz Can Koseoglu \email{denizcan.koseoglu@gmail.com} (\href{https://orcid.org/0000-0002-1569-9788}{ORCID})

}
\keyword{internal}
