% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{supp_pars}
\alias{supp_pars}
\title{Supplement function arguments (named vectors) with default values}
\usage{
supp_pars(defpars, pars = "default", parlb = NA)
}
\arguments{
\item{defpars}{A \strong{named} vector of default parameters (numeric or character).}

\item{pars}{A \strong{named} vector of input parameters where all names must correspond to those of \code{defpars}.}

\item{parlb}{The label for parameters. If \code{NA} (default), defaults to \code{"pars"}.}
}
\value{
The input vector supplemented with default parameters (if necessary).
}
\description{
Compares an input \strong{named} vector of parameters with a default vector,
supplementing with default parameters where necessary.
\strong{This function manual is hidden from the Index}.
}
\keyword{internal}
