% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swex}
\alias{swex}
\title{Example Subcritical Water Extraction (SWE) kinetic modeling data}
\format{
An object of class \code{list} of length 7.
}
\usage{
swex
}
\description{
A named nested \code{list} containing extraction curve and two-site kinetic desorption model data from scientific literature.
Each element is itself a \code{list} including the following elements:
\enumerate{
\item \code{$data}: A \code{data.frame} containing the OEC data (time and response).
\item \code{$resp_unit}: The response value units.
\item \code{$ext_pars}: A vector of extraction parameters such as pressure and temperature.
\item \code{$input_pars}: Input model parameters (use \code{\link{show_pars}} for description and units of all parameters).
\item \code{$output_pars}: Output model parameters (use \code{\link{show_pars}} for description and units of all parameters).
\item \code{$proc_desc}: A short description of the modeled process.
\item \code{$citation}: A reference for the source publication.
}

Currently, the following Overall Extraction Curves (OEC) and kinetic desorption model results are included for SWE of:
\enumerate{
\item \code{$abidin}: Cherry seed oil (Dimic et al., 2021).
\item \code{$duba1}: Grape skins (Duba, 2015).
\item \code{$duba2}: Grape seeds (Duba, 2015).
\item \code{$jamaludin1}: Alizarin from \emph{Morinda citrifolia} (Jamaludin et al., 2021).
\item \code{$jamaludin2}: Scopoletin from \emph{Morinda citrifolia} (Jamaludin et al., 2021).
\item \code{$kim}: Citrus flavonoids (Kim et al., 2020).
\item \code{$vasquez}: Anthraquinones from \emph{Heterophyllaea pustulata} (Barrera Vasquez et al., 2015).
}
}
\references{
Abidin, Z.Z., Samadi, M., Biak, D.R., Awang Yunus, R. (2024), 'Mathematical Modelling of Subcritical Water Extraction of
Essential Oil From Aquilaria Malacenssis Wood', \emph{Journal of Applied Science and Engineering} \strong{27} (12), pp. 3725-3738, DOI:
\url{https://www.doi.org/10.6180/jase.202412_27\%2812\%29.0012}.

Barrera vazquez, M.F., Comini, L.R., Milanesio, J.M., Nunez Montoya, S.C., Cabrera, J.L., Bottini, S., Martini, R.E. (2015),
'Pressurized hot water extraction of anthraquinones from \emph{Heterophyllaea pustulata} Hook f. (Rubiaceae)',
\emph{Journal of Supercritical Fluids} \strong{101}, pp. 170-175, DOI: \url{https://dx.doi.org/10.1016/j.supflu.2015.02.029}.

Duba (2015), Supercritical Technologies for the Valorization of Wine Industry By Products, University of Trento, PhD thesis.

Jamaludin, R., Kim, D.-S., Salleh, L.M., Lim, S.-B. (2021), 'Kinetic Study of SubcriticalWater Extraction of Scopoletin, Alizarin,
and Rutin from \emph{Morinda citrifolia}', \emph{Foods} \strong{10}, article 2260, DOI: \url{https://doi.org/10.3390/foods10102260}.

Kim, D.-S. (2020), Subcritical Water Extraction and Hydrolysis of Citrus Flavonoids: Kinetics, Optimization, and Biological Activities,
Jeju National University, South Korea, PhD thesis.
}
\seealso{
\code{\link{show_pars}}
}
\keyword{datasets}
