% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe-analyze.R
\name{twoway}
\alias{twoway}
\title{Get all two-way interactions from a linear/RSM model}
\usage{
twoway(c_facs, mod_order = 2)
}
\arguments{
\item{c_facs}{A \code{character} vector of \strong{coded} factors. \strong{Elements must all be capital letters}.}

\item{mod_order}{The \code{numeric} model order, either one of 1, 1.5, or 2.}
}
\value{
A \code{data.frame} of 3 columns containing the main, interaction, and/or quadratic terms (depending on \code{mod_order}) in
3 different representations (data, formula, and label).
}
\description{
Obtains all interaction and quadratic terms of a linear or quadratic model given labels of coded factors
alongside model order. Part of the \code{\link{doe_analyze}} workflow.
}
\examples{
twoway(c("A","B","C"), 2)
}
