% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{tws_cmp}
\alias{tws_cmp}
\title{Build two-site kinetic desorption model}
\usage{
tws_cmp(
  x,
  y,
  c0,
  q = NA,
  f = NA,
  k1_0 = 0.5,
  k2_0 = 0.5,
  f0 = NA,
  est_f = TRUE,
  modpts = 100,
  aggreg = "aard",
  maxt = round(max(x) * 1.2, -1),
  maxq = round(max(q) * 1.2, -1),
  optmet = "nlopt"
)
}
\arguments{
\item{x, y}{Both \code{numeric} vectors of time (\eqn{t}; min) and response (\eqn{e_t}; e.g. yield), respectively. Must be of equal length.
Missing data is not allowed.}

\item{c0}{The maximum possible yield \eqn{e_0}. Must have \strong{units identical to those of \code{y}}.}

\item{q}{An \strong{optional} \code{numeric} vector of Solvent-Material (S/M) ratio \eqn{q} to be included in the final output.}

\item{f}{An \strong{explicit} value of the fraction of easily-desorbed extract \eqn{F} (between 0 and 1).
When provided (\code{NA} by default), iterative curve fitting of this parameter is not carried out (\code{est_f} is ignored).}

\item{k1_0, k2_0}{\strong{Mandatory} initial estimates for first-order rate constants \eqn{k_1} and \eqn{k_2}.}

\item{f0}{\strong{Optional} initial estimate for \eqn{F}.}

\item{est_f}{A \code{logical} indicating whether \eqn{F} should be estimated via iterative curve fitting (\code{TRUE} by default).}

\item{modpts}{The number of points to model between 0 and \code{maxq} (or \code{maxt}). Defaults to \code{100}.}

\item{aggreg}{A string specifying how the "best" results of non-linear optimization (\code{nlsm}) should be chosen.
One of \code{"aard"} (default; value with minimum Average Absolute Relative Deviation is chosen) or \code{"mean"}
(the arithmetic mean of all results is taken).}

\item{maxt, maxq}{Maximum x-axis value (solvent expended, kg/kg insoluble solid for \code{maxq}, time for \code{maxt}) to use
for model predictions. Defaults to 120\% of the maximum experimental value.}

\item{optmet}{The method of iterative curve fitting to use for estimating \eqn{k_1}, \eqn{k_2}, and (optionally) \eqn{F}.
One of \code{"nlopt"} (Non-Linear Optimization via \code{\link[nloptr]{nloptr}}; default) or \code{"nlrob"} (Robust Fitting
via \code{\link[robustbase]{nlrob}}).}
}
\value{
A \code{list} with the following elements:
\enumerate{
\item \strong{$ordt}: A \code{data.frame} containing the original input data including the extraction time (\code{x}),
\strong{optionally} the S/M ratio (\code{q}), the actual response (\code{y}, usually yield) and associated model
predictions (\code{"pred_y"}), as well as the actual and predicted fractional yield (\code{"cc0"} and \code{"pred_cc0"})
defined as the ratio of yield at time \eqn{e_t} and the maximum possible yield \eqn{e_0}.
\item \strong{$mdt}: A \code{data.frame} of modeled data including the model type (\code{$model}), the extraction time
(\code{"x"}), the \strong{optional} S/M ratio (\code{"q"}), the response (\code{"y"}), and the fraction of maximum
attainable yield (\code{"cc0"}).
\item \strong{$mod_pars}: A named \code{numeric} vector of input and estimated model parameters including the first-order
rate constants \eqn{k_1} and \eqn{k_2} (\code{c("k1","k2")}), the fraction of easily desorbed solute \eqn{F} (\code{"f"}),
and the maximum possible yield of extractable material \eqn{e_0} (\code{"c0"}).
\item \strong{$fit_pars}: A \code{character} vector of which parameters in \code{$mod_pars} were iteratively fit to the model.
\item \strong{$resid}: A named \code{numeric} vector of various error values for the model, including Average Absolute Relative Deviation
AARD (\code{"aard"}), round mean squared error RMSE (\code{"rmse"}), and the R2 value (\code{"r2"}).
}
}
\description{
Generates and summarizes a two-site kinetic desorption (TWS) model. Part of the \code{\link{ktsmod}} workflow.
}
\seealso{
\code{\link{ktsmod}}, \code{\link{tws_eq}}, \code{\link{moderr}}
}
