% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{tws_eq}
\alias{tws_eq}
\title{Equation: Two-site kinetic desorption model}
\usage{
tws_eq(t, k1, k2, f)
}
\arguments{
\item{t}{A \code{numeric} vector of extraction time (usually in minutes) at which yield or another response was recorded.}

\item{k1, k2}{First-order rate constants \eqn{k_1} and \eqn{k_2} describing the "fast" and "slow" component
of target compound desorption, respectively.}

\item{f}{The fraction of total extractable analyte that desorbs at a faster rate (i.e. is easily accessible for the solvent).}
}
\value{
A \code{numeric} vector of model predictions.
}
\description{
Derives two-site kinetic desorption (TWS) model predictions. Part of the \code{\link{ktsmod}} workflow.
}
\seealso{
\code{\link{tws_cmp}}, \code{\link{ktsmod}}
}
\keyword{internal}
