% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kts.R
\name{tws_err}
\alias{tws_err}
\title{RMSE: Two-site kinetic desorption model}
\usage{
tws_err(pars, input, tm, f = NA)
}
\arguments{
\item{pars}{Initial estimates of first-order rate constants \eqn{k_1} and \eqn{k_2} as well as an \strong{optional}
estimate for fraction \eqn{F} of easily-desorbed extract.}

\item{input}{Vector of \strong{actual} response values. Must be equal in length to \code{tm}.}

\item{tm}{Vector of extraction time. Must be equal in length to the response vector \code{input}.}

\item{f}{An explicit value of \eqn{F}. If provided (\code{NA} by default), any estimate included in \code{pars} is ignored.}
}
\value{
The RMSE as a single \code{numeric} value \eqn{RMSE = \sum{\sqrt{(x_{model}-x_{observed})^2}/m}}.
}
\description{
Calculate the Round Mean Squared Error (RMSE) of the two-site kinetic desorption (TWS) model.
Part of the \code{\link{ktsmod}} workflow. \strong{This function manual is hidden from the Index}.
}
\seealso{
\code{\link{tws_eq}}, \code{\link{tws_cmp}}, \code{\link{ktsmod}}
}
\keyword{internal}
