% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_doe.R
\name{uncode_levels}
\alias{uncode_levels}
\title{Get uncoded factor levels of an experimental design}
\usage{
uncode_levels(input, design, levels, factors, flims, fnames)
}
\arguments{
\item{input}{A \code{data.frame} containing the \strong{coded} input factors and their levels. Columns (factor names) must be named alphabetically.}

\item{design}{The type of design as a \code{character} value. Possible values are \code{"CCC"}, \code{"CCF"}, \code{"FFD"}, \code{"FrFD"}.}

\item{levels}{The number of factor levels as a single \code{numeric} value.}

\item{factors}{The number of factors as a single \code{numeric} value.}

\item{flims}{A \code{list} of vectors (each of length 2) containing the lower and upper limits of factors.
The number of list elements must match \code{factors}. Hard limits are specified by naming the corresponding element \code{"hard"}.}

\item{fnames}{A \code{character} vector of factor names. Must be of equal length to \code{factors}.}
}
\value{
Returns the input \code{data.frame} with added columns containing the \strong{uncoded} factor levels.
}
\description{
Part of the \strong{experimental design generation} workflow.
\strong{This function manual is hidden from the Index.}
}
\seealso{
\code{\link{doe_frfd}}, \code{\link{doe_ffd}}, \code{\link{doe_ccd}}, \code{\link{doe_bbd}}, \code{\link{doe_tm}}
}
\keyword{internal}
