% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bic.R
\name{bic_cmp}
\alias{bic_cmp}
\title{Derive the complete BIC model}
\usage{
bic_cmp(
  yield,
  q,
  n,
  a0,
  gam,
  porosity,
  qaver,
  Nm,
  ys,
  xu,
  r0,
  ksas0,
  kf0 = 0.001,
  qc0 = 400,
  modpts = 100,
  maxq = round(max(q) * 1.2, -1),
  nlsm = c("MM", "tau", "CM", "mtl"),
  aggreg = "aard",
  const_flow = TRUE
)
}
\arguments{
\item{yield}{A \code{numeric} vector of yields (g/g insoluble solid).}

\item{q}{A \code{numeric} vector of relative amounts of expended solvent (kg/kg insoluble solid).}

\item{n}{Period corresponding to the end of the CER (in \strong{number of experimental points}).}

\item{a0}{Specific surface area per unit volume of extraction bed (1/m).}

\item{gam}{Solvent to matrix ratio in the bed (\eqn{kg_{solvent}/kg_{insoluble solid}}).}

\item{porosity}{Bed porosity (dimensionless).}

\item{qaver}{Average solvent flow rate (kg/s).}

\item{Nm}{Mass of insoluble material (g).}

\item{ys}{Extract solubility in CO2 (g/g).}

\item{xu}{Weight fraction (concentration) in the untreated solid.}

\item{r0}{Initial estimate of the grinding efficiency (fraction of broken cells).}

\item{ksas0}{Initial estimate of the product of the solid phase mass transfer coefficient (\eqn{k_s}; 1/s)
and the specific area between intact and broken cells (\eqn{a_s}; 1/m), in 1/m/s.}

\item{kf0}{An initial estimate of the fluid phase mass transfer coefficient (1/s). Defaults to \code{0.001}.}

\item{qc0}{Initial estimate of solvent expended at the end of the CER (kg/kg CO2). Defaults to \code{400}.}

\item{modpts}{A single \code{numeric} value specifying how many points should be modelled throughout the OEC curve range (\code{100} by default).}

\item{maxq}{Maximum x-axis value (solvent expended, kg/kg insoluble solid) to use for model predictions.
Defaults to 120\% of the maximum \code{q} value.}

\item{nlsm}{A \code{character} vector of non-linear optimization methods to use for convergence of the model.
One or more of: \code{"MM"}, \code{"tau"}, \code{"CM"}, and/or \code{"mtl"} (\strong{see \code{\link[robustbase]{nlrob}} for details}).}

\item{aggreg}{A string specifying how the "best" results of non-linear optimization (\code{nlsm}) should be chosen.
One of \code{"aard"} (default; value with minimum Average Absolute Relative Deviation is chosen) or \code{"mean"}
(the arithmetic mean of all results is taken).}

\item{const_flow}{A \code{logical} specifying whether the flow rate is constant throughout the extraction.
If set to \code{TRUE}, the solvent-material ratio (S/M) is \strong{not} converted to extraction time (min).}
}
\value{
A named \code{list} containing the following elements:
\enumerate{
\item \strong{$ordt}: A \code{data.frame} containing the original input data including the temporal variable \code{$t},
the Solvent-Solute (S/M) ratio \code{$x}, the \strong{actual} fractional yield (\code{$y}; g/g insoluble solid), as well as yields
\strong{modeled} via the 3-period and 2-period models (\code{$y_cmp3} and \code{$y_cmp2}, respectively).
\item \strong{$mdt}: A \code{data.frame} of modeled data including the model type (\code{$model}),
the region of the extraction curve (\code{$period}), as well as S/M ratio \code{$x}, fractional yield \code{$y}, and time (\code{$t})
up to \code{maxq}.
\item \strong{$mod_pars}: The model parameters including the external mass transfer resistance (\eqn{\theta_e}; \code{"thetae"}),
the product of the fluid phase mass transfer coefficient \eqn{k_f} and specific surface area per unit volume of extraction bed \eqn{a_0}
(\eqn{k_fa_0}; \code{"kfa0"}), the relative amount of expended solvent (kg/kg insoluble solid) at the end of the CER (\eqn{q_m}; \code{"qm"}),
FER for the 3-period model (\eqn{q_n}; \code{"qn"}), and DC for the 2-period model (\eqn{q_{cmp}}; \code{"qs"}),
coefficient \eqn{\beta} (\code{"beta"}), initial fraction of solute in broken cells (\eqn{G}; \code{"G"}),
the fluid mass transfer coefficient (\eqn{k_f}; \code{"kf"}), grinding efficiency (\eqn{r}; \code{"r"}), as well as the product
of the solid phase mass transfer coefficient \eqn{k_s} and the specific area between intact and broken cells \eqn{a_s}
(\eqn{k_sa_s}; \code{"ksas"}). See \code{\link{show_pars}} for detailed description of these parameters.
\item \strong{$fit_pars}: A \code{character} vector of which parameters in \code{$mod_pars} were iteratively fit to the model.
\item \strong{$resid}: A named \code{numeric} vector of various error values for the model, including Average Absolute Relative Deviation
AARD (\code{"aard"}), round mean squared error RMSE (\code{"rmse"}), and the R2 value (\code{"r2"}).
}
}
\description{
Derives, summarises, and visualises the results of the complete Broken-and-Intact Cells (BIC) model of Sovova (2005).
Part of the \code{\link{bicmod}} workflow.
}
\details{
For a detailed description of this variant of the Broken and Intact Cells (BIC) model, refer to \code{\link{bicmod}}.
}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bicmod}}, \code{\link{show_pars}}, \code{\link{moderr}}
}
