% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicmod.R
\name{bicmod}
\alias{bicmod}
\title{Construct BIC models of varying complexity}
\usage{
bicmod(
  oec,
  oec_vars,
  pars,
  opt_est = "default",
  etoh_frac = 0,
  flowpar = rep(NA, 2),
  ro_co2 = NA,
  tmax = NA,
  qmax = NA,
  cumulative = FALSE,
  mass_flow = FALSE,
  draw = TRUE,
  aggreg = "aard",
  modtype = "all",
  units = "default",
  silent = FALSE
)
}
\arguments{
\item{oec}{A \code{data.frame} of OEC data. Must include all \code{oec_vars}.}

\item{oec_vars}{A \strong{named} \code{character} vector of column names included in \code{oec}.
These \strong{must} include the extraction time (\code{"x"}) and response (\code{"y"}) with appropriate \code{units}.
A third parameters indicating the solvent usage (\code{"slv"}) is \strong{optional} and may be provided as an
\strong{alternative} to an explicit flow rate normally provided in \code{pars["flow"]} - this is useful when the
flow rate is uneven throughout the extraction process.}

\item{pars}{A named \code{numeric} vector of input parameters for the model, divided into \strong{mandatory} and
\strong{optional} parameters. \strong{Mandatory} parameters include pressure (\code{"pres"}; bar),
temperature (\code{"temp"}; degC), mass of material loaded (\code{"mass_in"}; g), moisture content
(\code{"moisture"}; \%), diameter (\code{"D"}; m) and length (\code{"L"}; m) of the extraction vessel,
the fraction of ethanol co-solvent (\code{"etoh"}), the real (\code{"dr"}; g/L) and apparent (\code{"dp"}; g/L)
densities of the raw material, and the number of observations/OEC points corresponding to the end of the CER (\code{"n"}).
The \strong{optional} variables default to \code{NA} and include the flow rate of solvent (\code{"flow"}; units set in \code{units}),
and the maximum extractable material fraction (\code{"cu"}), which \strong{must be included unless} \code{modtype} is set to \code{"cu"}.}

\item{opt_est}{Either \code{"default"} or a \strong{named} \code{numeric} vector of initial parameter estimates
for iterative optimization. May include any of grinding efficiency \eqn{r} (\code{"r"}), the product of the solid phase
mass transfer coefficient (\eqn{k_s}) and the specific area between intact and broken cells (\eqn{a_s}) \eqn{k_sa_s} (\code{"ksas"}),
external mass transfer coefficient \eqn{\theta_e} (\code{"thetaf"}), extraction duration of FER \eqn{t_i} (\code{"ti"}),
the fluid phase mass transfer coefficient \eqn{k_f} (\code{"kf"}), and/or the relative amount of solvent expended
at the end of CER \eqn{q_m} (\code{"qc"}). An additional parameter \code{"c3"} is related to the maximum extractable material fraction
\eqn{c_u} and is \strong{only required} when \code{modtype} is set to \code{"cu"}.
Any values not specified will be set to the following defaults:
\code{c(r = 0.4, ksas = 10e-5, qc = 400, thetaf = 1, ti = 30, kf = 0.001, c3 = 0.15)}.}

\item{etoh_frac}{The single \code{numeric} fraction of ethanol co-solvent (between 0-0.99). Defaults to \code{0}.
Must be non-zero if \strong{flow rate} is not provided in \code{pars} and \code{pars["etoh"]} is \strong{non-zero}.}

\item{flowpar}{Either \code{NA} (default) or a \code{numeric} vector of length 2 providing temperature and pressure
at which flow rate of CO2 is measured.}

\item{ro_co2}{The supercritical CO2 density (in g/L). If not provided (\code{NA}; default),
it is calculated via the Bender Equation of State (see \code{\link{bendens}}).}

\item{tmax}{Maximum x-axis value (time, min) to use for model predictions \strong{exclusively} for \code{modtype = "ct"}.
Setting to \code{NA} defaults to 120\% of the maximum experimental value.}

\item{qmax}{Maximum x-axis value (solvent expended, kg/kg insoluble solid) to use for model predictions for all models except
\code{modtype = "ct"}. Setting to \code{NA} defaults to 120\% of the maximum experimental value.}

\item{cumulative}{A \code{logical} switch specifying whether the \strong{response} and/or \strong{solvent consumption}
values provided in \code{oec} are cumulative or not (defaults to \code{FALSE}).}

\item{mass_flow}{A \code{logical} indicating whether the flow rate provided in \code{pars} is \strong{mass}
or \strong{volumetric} (\code{FALSE}; default).}

\item{draw}{A \code{logical} switch. Should generated plots be plotted? Defaults to \code{TRUE}.}

\item{aggreg}{A string specifying how the "best" results of non-linear optimization (\code{nlsm}) should be chosen.
One of \code{"aard"} (default; value with minimum Average Absolute Relative Deviation is chosen) or \code{"mean"}
(the arithmetic mean of all results is taken).}

\item{modtype}{Specifies the type(s) of BIC models to generate. Either \code{"all"} (default) or one or more of:
\code{"sim"} (simplified), \code{"ct"} (characteristic times), \code{"cmp3"} (complete with 3 OEC regions),
and/or \code{"cmp2"} (complete with 2 OEC regions). A separate option \code{"cu"} is used when the
extractable material fraction \eqn{c_u} is to be estimated.}

\item{units}{A \strong{named} \code{character} vector of length 1 or 2 specifying the units of \code{"flow"}
(one of \code{"mL/min"}, the \strong{default} \code{"g/min"}, \code{"kg/h"}, \code{"L/h"}, or \code{"none"}) and/or
the response \code{"resp"} (one of \code{"g"}, the \strong{default} \code{"percent"}, \code{"permille"}, \code{"ppm"},
or \code{"ppb"}). Where not provided, default values are used.}

\item{silent}{Should console output be silenced? Defaults to \code{FALSE}).}
}
\value{
The output takes two forms. When \code{modtype} includes \code{"cu"} (for estimation of the maximum extractable fraction \eqn{c_u}),
the output is a \code{list} of the estimated \eqn{c_u} value (\code{$cu}) and the iteratively derived final estimate of the related constant
\strong{c3} (\code{$mod_coefs}).
Alternatively, when \code{modtype} includes any of \code{c("sim","ct","cmp3","cmp2")}, output is a \code{list} containing
the following elements:
\enumerate{
\item \strong{$data}: A \code{data.frame} of original \code{oec} input data with added values of Solvent-Material (S/M) ratio \code{$q},
as well as fractional yields \code{$e} (g/g insoluble solid) and \code{$estar} (g/g total dry solid used for \code{modtype == "ct"}).
\item \strong{$sim, $ct, and/or $cmp}: Each a \code{list} of results for the eponymous BIC model, with output structure outlined in
the documentation of \code{\link{bic_sm}}, \code{\link{bic_ct}}, and \code{\link{bic_cmp}}, respectively.
\item \strong{$plots}: A \code{list} of plot objects of class \code{"ggplot"}.
\item \strong{$input}: A named \code{numeric} vector of input parameters. In addition to those provided in argument \code{pars},
some calculated parameters are also included. These are the specific surface area per unit volume of extraction bed \eqn{a_0} (\code{"a0"};
1/m), the apparent porosity \eqn{\varepsilon} (\code{"porosity"}), the CO2 to insoluble solid ratio in the extraction bed \eqn{\gamma}
(\code{"gamma"}; kg/kg), the number of experimental points \eqn{m} (\code{"m"}), total \strong{dry} mass of material \eqn{N} (\code{"Ng"}; g),
mass of \strong{insoluble} material \eqn{N_m} (\code{"Nm"}; g), the ratio of solute to insoluble material \eqn{x_u} (\code{"xu"}; kg/kg),
and the apparent extract solubility \eqn{y_s} (\code{"ys"}; g/g).
\item \strong{$call}: The function call.
}
}
\description{
Constructs various Broken and Intact Cells (BIC) kinetic models from
overall extraction curves (OECs) of supercritical CO2, subcritical water, or any
other process. Consult \strong{References} and the \strong{See Also} section for further information.
\strong{This workflow is currently a Work in Progress (WIP)}.
}
\details{
This workflow derives and visualizes three different Broken and Intact Cells (BIC) kinetic models based on the work of
Sovova (2005, 2012, 2017) and previously utilized by Rizza (2014) and many other authors. These include the \strong{simplified},
\strong{characteristic times}, and \strong{complete} models - descriptions of associated parameters may be found in the descriptions of
\code{\link{bic_sm}}, \code{\link{bic_ct}}, and \code{\link{bic_cmp}}, respectively, and their linked functions.
}
\examples{
bic_res <- bicmod(oec = sfex[[2]][["data"]],
oec_vars = c(x = "Time_min", y = "Yield_g", slv = "Solvent_mL"),
pars = c(pres = 300,
         cu = 0.165,
         temp = 45,
         flow = NA,
         mass_in = 0.5125,
         moisture = 8.6,
         D = 0.015,
         L = 0.015,
         etoh = 0.5,
         dr = 1554,
         dp = 0.0004,
         n = 2),
opt_est = "default",
flowpar = c(1.01325, 25),
etoh_frac = 0.06, #Required when CO2 flow is not provided but 'etoh' is non-zero
ro_co2 = NA,
tmax = NA,
qmax = NA,
cumulative = FALSE,
mass_flow = FALSE,
draw = TRUE,
units = c(flow = "none", resp = "g"),
modtype = "all")

}
\references{
Rizza, C.S. (2014), \emph{Experiments and Modeling of Supercritical CO2 Extraction of Lipids from Microalgae}, MSc thesis, Universita Degli Studi Di Padova, Dipartimento Di Ingegneria Industriale.

Sovova, H. (2005), 'Mathematical model for supercritical fluid extraction of natural products and extraction curve evaluation', \emph{The Journal of Supercritical Fluids} \strong{33} (1), pp. 35-52, DOI: \url{https://doi.org/10.1016/j.supflu.2004.03.005}.

Sovova, H. (2012), 'Steps of supercritical fluid extraction of natural products and their characteristic times', \emph{The Journal of Supercritical Fluids} \strong{66}, pp. 73-79, DOI: \url{https://doi.org/10.1016/j.supflu.2011.11.004}.

Sovova, H. (2017), 'Broken-and-intact cell model for supercritical fluid extraction: Its origin and limits', \emph{The Journal of Supercritical Fluids} \strong{129}, pp. 3-8, DOI: \url{https://doi.org/10.1016/j.supflu.2017.02.014}.
}
\seealso{
\code{\link{bendens}}, \code{\link{etoh_dens}}, \code{\link{bic_sm}}, \code{\link{bic_ct}}, \code{\link{bic_cmp}}, \code{\link{show_pars}}, \code{\link{kin_plot}}, \code{\link{kin_splot}}, \code{\link{kin_export}}
}
