% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcom.R
\name{calcom}
\alias{calcom}
\title{Calculate COM for an SFE or SWE extraction process}
\usage{
calcom(
  input,
  invars = c(names(input)[1], names(input)[2]),
  pltlab = invars[1],
  cosol_loss = 0.05,
  gen,
  crm,
  auxpr = NA,
  auxfr = NA,
  cut,
  col,
  fci,
  taxrate = NA,
  flowpar = "auto",
  draw = TRUE,
  mass_flow = FALSE,
  comode = "sfe",
  use_coefs = FALSE,
  export = "none"
)
}
\arguments{
\item{input}{A named \code{character} vector of length 2, or a \code{data.frame} with 2 columns
containing the temporal (\strong{element 1}) and response (\strong{element 2}) components of the extraction process. When
a \code{data.frame} is provided, the input is treated like an extraction curve, and COM is calculated separately for each row.}

\item{invars}{A \code{character} vector containing names of the 2 input variables (temporal and response).
By default, the names provided in \code{input} are used.}

\item{pltlab}{The plot title. By default, \code{invars[1]} is used.}

\item{cosol_loss}{A \strong{single} \code{numeric} value between 0.001 and 0.99 denoting the \strong{fraction}
of co-solvent lost during post-processing (e.g. evaporation). Defaults to 0.05.}

\item{gen, crm, cut, col, fci}{\strong{Named} numeric vectors of input parameters required for \strong{general calculations} (\code{gen}),
\strong{Cost of Raw Materials} (\code{crm}), \strong{Cost of Labour} (\code{col}), \strong{Cost of Utilities} (\code{cut}),
and/or \strong{Fixed Costs} (\code{fci}). Some parameters are mandatory, while others are optional.
For an exhaustive list of available \strong{input} and \strong{output} parameters, use \code{show_pars("com")}.}

\item{auxpr, auxfr}{\strong{Identically named} \code{numeric} vectors containing the
\strong{price per unit mass} of any auxiliary materials used in the process
(e.g. for pre- or post-processing), and the \strong{fraction of said material
required relative to the mass of raw material} (>0).
Both default to \code{NA} (no auxiliary materials used).}

\item{taxrate}{The \code{fraction of gross profit} to be removed as tax to calculate \strong{net profit}.}

\item{flowpar}{The parameters of temperature and pressure at which flow rate is given.
Either \auto{"auto"} (default) or a \code{list} of length 2, where the first element specifies the pressure and temperature at which flow is
measured for the main solvent (CO2 or water depending on \code{comode}), while the second element specifies these values for the co-solvent (
either CO2 or ethanol when \code{comode} is set to \code{"swe"} or \code{"sfe"}, respectively). \strong{Note} that when \code{comode} is
\code{"sfe"}, only the temperature of ethanol is included in the second element, as pressure is disregarded in density calculations.
Used for conversion of volumetric flow to mass flow when \code{!mass_flow}.
If set to \code{"auto"}, uses a temperature and pressure of 10 degrees Celsius and 60 bar for CO2, and 25 degrees Celsius
for ethanol.}

\item{draw}{A \code{logical} switch specifying whether to plot the results.
Only works if \strong{an extraction curve is provided as } \code{input}.}

\item{mass_flow}{A \code{logical} specifying whether the \code{["flow"]} component of \code{gen} is provided in \strong{mass (g/min)}
or \strong{volumetric (mL/min)} units.}

\item{comode}{Which extraction process to calculate COM for?
One of: \code{"sfe"} (Supercritical Fluid Extraction) or \code{"swe"} (Subcritical Water Extraction).}

\item{use_coefs}{A \code{logical} denoting whether to use pre-defined coefficients to calculate COM (\code{TRUE}) from
CUT, CRM, FCI, and COL, or whether to sue a simple sum instead (\code{FALSE}, \strong{default}).}

\item{export}{An \strong{existing} folder path into which the results are exported as a .CSV file and accompanying graphics (if any).
Defaults to \code{"none"} (the results are not exported).}
}
\value{
A named \code{list} containing the input parameters (\code{$input}), the calculated output values (\code{$output})
its simplified form (\code{$simple_output}), extra parameters including FCI and COL (\code{$extra}),
and the function call (\code{$call}). Optionally, specific cost and payback time plots of the results are also included
when \code{draw} is \code{TRUE} and an extraction curve is provided as \code{input} (\code{$plots}).
Any plot points removed due to economic \strong{non-}viability are also noted in element \code{$rm_state}.
}
\description{
Calculates the Cost of Manufacturing (COM) and its accompanying components including Cost of Labour (COL),
Cost of Raw Materials (CRM), Cost of Ulitities (CUT) and Fixed Costs (FCI). See \strong{Details} for further information.
}
\details{
The calculation of Cost of Manufacturing (COM) is based on the formula provided by Turton et al. (1998)
when \code{use_coefs} is \code{TRUE}, and combined the Cost of Labour (\eqn{COL}), Cost of Utilities (\eqn{CUT}),
Cost of Raw Materials (\eqn{CRM}), and Fixed Costs (\eqn{FCI}). In the current implementation, the Cost of Waste Management
(\eqn{CWT}) normally included in this method has been omitted since neither supercritical CO2 and subcritical water extraction
do not produce significant amounts of waste.
\deqn{COM = 0.304 \times FCI + 2.73 \times COL + 1.23 \times (CRM + CUT)}
Alternatively (and by default), COM is calculated as a simple sum of \strong{monthly} CUT, CRM, COL, and FCI.
The various COM constituents are calculated as below.

The monthly COL is calculated using a simplified formula incorporating the average monthly wage (\eqn{W_{pay}}),
the number of daily work shifts (\eqn{W_{sh}}), and the number of personnel required (\eqn{W_{pers}}):
\deqn{COL = W_{sh} \times W_{pers} \times W_{pay}}

The monthly CRM encompasses a sum of various costs associated with raw materials, calculated in multiple steps. First,
the number of \strong{monthly} extraction cycles \eqn{N_{ex}} is calculated using \eqn{W_{sh}}, the number of work hours per shift
(\eqn{W_{hr}}), extraction time (\eqn{T_{ex}}), batch exchange time (\eqn{T_{aux}}), and the number of work days per month (\eqn{W_{days}}).
\deqn{N_{ex} = (W_{sh}\times W_{hr})/((T_{ex} + T_{aux})/60) \times W_{days}}

Any volumetric flow rates are converted to mass flow using CO2 (for SFE) or water (for SWE) densities at process conditions specified in
\code{crm} or \code{flowpar}. The equations incorporate fluid density \eqn{\rho} calculated by the Bender Equation (\code{\link{bendens}}) or
the IAPWS R6-95 formulation (\code{\link{h2o_dens}}).
\deqn{F_{mass} = \rho/1000 \times F_{vol}}
The density of main and co-solvent mixtures is calculated using the Linear Blend Rule (see \code{\link{etoh_dens}}).

Next, the main solvent usage (\eqn{SL_{main}}, water for SWE or a supercritical fluid like CO2 for SFE) and co-solvent usage (\eqn{SL_{aux}})
are calculated via the equations below. \strong{Currently only EtOH for SFE and CO2 for SWE are supported as co-solvents}.
The units are either \strong{L} or \strong{kg} depending on whether \code{mass_flow} is \code{TRUE}. For CO2, the equation
depends on whether a recovery tank pressure was specified - if it was, the system is assumed to be equipped with a CO2
recirculation/recovery system, incorporating the extractor volume \eqn{V_{ex}}. If not, the flow rate is simply converted
to kg/min and multiplied by extraction time (the same equation is used for SWE). In equations below, density \eqn{\rho} is assumed to be in
units of \strong{g/L}.
\deqn{SL = \rho/1000\times V_{ex} \text{ for CO2 as a main (SFE) or a co-solvent (SWE) with a recovery system}}
\deqn{SL = F_{mass}/1000 \times T_{ex} \times 0.05 \text{ for EtOH co-solvent (SFE) assuming a 5\% loss per batch}}
\deqn{SL = \rho_{CO2}/1000 \times (F_{aux}/(F_{aux} + F_{main})) \times V_{ex} \text{ for CO2 as a co-solvent for SWE}}
\deqn{SL = F_{mass}/1000 \times T_{ex} \text{ for all solvents where a recovery system is not present}}

Finally, the requirement (\eqn{REQ}, kg), \eqn{N_{ex}}, and the price \eqn{PR} of \strong{each} material \eqn{i} (including solvents)
are combined as follows and the results summed to obtain the monthly CRM:
\deqn{CRM = \sum{REQ_i \times N_{ex} \times PR_i}}

Similarly, all main power requirements (\eqn{PWRM_i}, in \strong{kWh}) are multiplied by \eqn{N_{ex}}, \eqn{T_{ex}}
(in \strong{hr}), and cost per kWh (\eqn{PR_{kWh}}) to obtain main power requirement \eqn{PW_{main}}:
\deqn{PW_{main} = PWRM_i \times N_{ex} \times T_{ex} \times PR_{kWh}}

Power requirements for \strong{drying}, \strong{comminution}, and \strong{evaporation}
(termed \eqn{PW_{aux}} when summed) are calculated
by multiplying the respective power requirement \eqn{PWRA_i} with the monthly consumption of the
associated material(s) \eqn{REQ_i} divided by the respective processing capacity of required equipment
\eqn{CAP_i}:
\deqn{PW_{aux} = \sum{PWRA_i \times (REQ_i/CAP_i)}}

The main and auxiliary power usage is finally summed to obtain \eqn{CUT}:
\deqn{CUT = PW_{main} + PW_{aux}}

The FCI is either used as-provided in the input data or incorporates the \strong{depreciation fraction} that is deducted yearly
from the CAPEX.

Once COM is calculated, the Specific Cost \eqn{SC} (USD/kg) of extract production is calculated by dividing monthly \eqn{COM} by the monthly
yield of extract \eqn{Y_{month}} (kg). Sales Volume \eqn{SV} (USD/month) is then obtained by multiplying \eqn{Y_{month}} by sales price
\eqn{SP_{kg}} (USD/kg). The manufacturing cost (\eqn{SC \times Y_{month}}) is then subtracted from \eqn{SV} to obtain the Gross Profit
(\eqn{GPr}, USD/month), which may then be corrected for taxes to get Net Profit (\eqn{NPr}).
\deqn{SC = COM/Y_{month}}
\deqn{SV = Y_{month} \times SP_{kg}}
\deqn{GPr = SV - SC \times Y_{month}}
\deqn{NPr = GPr - taxes}

Finally, the Profit Margin (\eqn{MRG}, \%) and Payback Period (\eqn{PBK}, yr) are obtained as follows:
\deqn{MRG = NPr/SV \times 100}
\deqn{PBK = CAPEX/(NPr \times 12)}
}
\examples{
#Calculate COM for a single process
comin <- c(time = 180, yield = 7)

#Calculate COM throughout an extraction curve and plot the results
comin <- data.frame(time = c(5,9,11,15,19,21,23,25,30,40,60,80,100,120,140,160,180),
yield = c(0.80, 2.18, 2.79, 3.40, 3.86, 4.17, 4.47, 4.63, 4.93,
          5.24, 5.70, 6.00, 6.31, 6.46, 6.77, 6.92, 7.00))

comres <- calcom(input = comin,
gen = c(volex = 30, load = 9, pres = 300, temp = 40, flow = 3000,
        extime = 30, csol_flow = 300, dilfac = 2, pr_sale = 90),
crm = c(bh = 155, id = 15.5, pr_mat = 0.5, pr_msol = 0.8, pr_csol = 2,
recp = 60, rect = 10, sept = 55),
cut = c(pw_main = 25, pr_kwh = 0.13, pw_dry = 2, pw_com = 1, pw_evap = 13.3,
cap_dry = 4.16, cap_com = 20, cap_evap = 60),
col = c(oper = 1, whr = 8, shifts = 3, wage = 940, wdays = 24),
fci = c(capex = 150000, maint = 250, other = 1210),
auxpr = c(oil = 1.52),
auxfr = c(oil = 1),
pltlab = "Time (min)")

}
\references{
Turton, R., Bailie, R.C., Whiting, W.B., Shaeiwitz, J.A. (1998), \emph{Analysis, Synthesis and Design of Chemical Process, PTR}, Prentice Hall, Upper Saddle River, NJ, USA.
}
\seealso{
\code{\link{show_pars}}, \code{\link{bendens}}, \code{\link{h2o_dens}}, \code{\link{com_export}}
}
