% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_desir.R
\name{doe_desir}
\alias{doe_desir}
\title{Combine models with different responses into a Desirability Function}
\usage{
doe_desir(
  mods,
  dsrng,
  frng = "default",
  obj = rep("max", length(mods)),
  dtype = "coded",
  wts = lapply(obj, function(x) if (x == "trg") rep(1, 2) else 1),
  spts = c(100, 10),
  modbase = "final",
  optmet = "nlopt",
  kmed = NA,
  export = "none",
  silent = FALSE
)
}
\arguments{
\item{mods}{A \code{list} of DoE model results output from function \code{\link{doe_analyze}}.}

\item{dsrng}{A named \code{list} of \code{numeric} vectors of length 2 or 3 specifying lower and upper response limits as well as
(\strong{optionally}) the target for each model in \code{mods}. The target can \strong{only} be specified if the
corresponding element of \code{obj} is set to \code{"trg"}. The list elements must have names identical to those of responses.}

\item{frng}{An optional \strong{named} \code{list} of \code{numeric} vectors of length 2 containing lower and upper limits of
all factors. Names must correspond to those of \strong{coded} factors (i.e. capital letters).
If set to the \code{"default"} value, uses the full range of each factor.}

\item{obj}{Either a single string or \code{character} vector denoting the objective of desirability function.
Each element must be one of: \code{"min"} (response minimization), \code{"max"} (maximization), or \code{"trg"}
(a specific response target). When a string is given, the value applies to all models in \code{mods}.}

\item{dtype}{A \code{character} string specifying whether \code{"coded"} (default) or \code{"uncoded"} factor values should be
used for processing. \strong{Note} that ranges in \code{frng} must be provided based on this option.}

\item{wts}{A \code{list} of desirability weights between \strong{0.1 and 10} where each element corresponds to a model
in \code{mods} and its associated \code{obj} and must either be a single \code{numeric} value or a vector of length
2 when the corresponding \code{obj} is set to \code{"trg"} (in such cases, the \strong{first and second} elements are weights
used for values \strong{lower and higher} than the target, respectively).}

\item{spts}{A \code{numeric} vector of length 2 providing the number of desirability optimization starting points
to be randomly generated (\strong{element 1}), and those to be retrieved from the original model \code{data.frame} (\strong{element 2}).
Defaults to \code{c(100,10)}.}

\item{modbase}{A \code{character} string specifying which model from each element of \code{mods} should be
used for processing. One of \code{"initial"} or \code{"final"} (default).}

\item{optmet}{optimization method. One of \code{"optim"} (uses the eponymous \code{\link[stats]{optim}} function) or
\code{"nlopt"} (uses the \strong{NLopt} library for non-linear optimization).}

\item{kmed}{The \strong{optional} number of clusters to use for Partitioning Around Medoids (PAM) clustering.
May be set to a specific integer, \code{"auto"} for automatic choice of appropriate cluster number, or \code{NA}
(\strong{default}, clustering \strong{not run}).}

\item{export}{An \strong{existing} folder path where a .CSV of the results should be exported.}

\item{silent}{A \code{logical} specifying whether useful updates in the console during processing should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
A named \code{list} containing the following elements:
\enumerate{
\item \strong{$factor_lims}: A \code{data.frame} summarizing the names, objectives, and lower/upper limits of all factors from \code{frng}.
\item \strong{$response_lims}: A \code{data.frame} with names, desirability function objectives, lower/upper limits, and weights of all
responses from \code{dsrng}.
\item \strong{$mod_sums}: A \code{data.frame} summarizing key performance metrics from models.
\item \strong{$orig_data}: The original input \code{data.frame} used to build the models, including coded and uncoded factor levels,
responses, as well as corresponding individual and overall desirability values.
\item \strong{$output_data}: The results of local desirability optimization from all starting points specified in \code{spts},
including individual and overall desirabilities.
\item \strong{$unique_solutions}: A \code{data.frame} of statistically unique solutions derived from \code{$output_data}.
\item \strong{$call}: The function call.
}
}
\description{
Calculates individual and overall desirability function from two or more models (linear or quadratic)
output from the \code{\link{doe_analyze}} function.
See \strong{Details} and \strong{References} for further information.
}
\details{
This function retrieves two or more responses from various DoE-derived linear and/or quadratic models analyzed by \code{\link{doe_analyze}}
and transforms each response into a desirability function using the approach of Derringer and Suich (1980) for simultaneous
optimization of several responses. The overall desirability is then calculated and optimized by non-linear techniques including
\code{\link[stats]{optim}} and the \strong{NLopt} library.

The process allows for calculation of one of three types of desirability functions including:
\enumerate{
\item \strong{Larger-The-Best (LTB)}: A maximization function.
\item \strong{Smaller-The-Best (STB)}: A minimization function.
\item \strong{Nominal-The-Best (NTB)}: A function targeting a specific optimal response value.
}

The response \eqn{\hat{Y}_i} as well as its minimum (\eqn{Y_{i\ast}}) and maximum (\eqn{Y_{i}^{\ast}}) acceptable values are incorporated
into the calculation of a desirability \eqn{d_i} where \eqn{0 \leq d_i \leq 1}. A weight \eqn{0.1 \leq r1 \leq 10} is also added
to modify the relative influence of the response value on desirability. When \eqn{r<1}, the desirability increases at a lower rate
above \eqn{Y_{i\ast}}, decreasing the relative importance of the response. The reverse is true when \eqn{r>1}, rapidly increasing
desirability above \eqn{Y_{i\ast}} and increasing importance of the response. The overall desirability equations for
\strong{one-sided transformations} is:

\deqn{d_i = 0 \text{ if } \hat{Y}_i \leq Y_{i\ast}}
\deqn{d_i = [(\hat{Y}_i - Y_{i\ast})/(Y_i^\ast - Y_{i\ast})]^{r1} \text{ if } Y_{i\ast} < \hat{Y}_i < Y_i^\ast}
\deqn{d_i = 1 \text{ if } \hat{Y}_i \geq Y_i^\ast}

Upper, lower, or both limits can be set for each response.

For two-sided transformations required by the \strong{NLB} approach, the value \eqn{\hat{Y}_i} has both a \strong{minimum and maximum}
constraint and the value chosen for \eqn{c_i} is based on maximum desirability provided within the constraints of \eqn{Y_{i\ast}}
and \eqn{Y_{i}^{\ast}}. The relative importance of each side of this target value is adjusted by factors \eqn{r1} and \eqn{r2}.
The associated equations are as follows:

\deqn{d_i = ((\hat{Y}_i - Y_{i\ast})/(c_i - Y_{i\ast}))^{r1} \text{ if } Y_{i\ast} \leq \hat{Y}_i \leq c_i}
\deqn{d_i = ((\hat{Y}_i - Y_i^\ast)/(c_i - Y_i^\ast))^{r2} \text{ if } c_i < \hat{Y}_i \leq Y_i^\ast}
\deqn{d_i = 0 \text{ if } \hat{Y}_i < Y_{i\ast} \text{ or } \hat{Y}_i > Y_i^\ast}

The \strong{overall} desirability \eqn{OD} may then be calculated by taking the \strong{geometric mean} of individual response desirability values:

\deqn{OD = (d_1 \times d_2 \times ... \times d_k)^{1/k}}

Both individual and overall desirability may be optimized using \strong{non-linear techniques}.
}
\examples{
#Calculate overall desirability among 3 responses
doe_lst1 <- load_internal("doe_lst1")

desires <- doe_desir(mods = doe_lst1,
                     dsrng = list(CarnosicAcid_mgg = c(0,150),
                     Carnosol_mgg = c(0,65), ExtYield = c(1,7)),
                     frng = list(B = c(40,60), A = c(10,30), C = c(1,3)),
                     obj = c("max", "max", "max"),
                     dtype = "uncoded",
                     wts = rep(1,3),
                     spts = c(100,10),
                     modbase = "final",
                     optmet = "nlopt",
                     kmed = "auto",
                     export = "none",
                     silent = FALSE)

}
\references{
Cardoso, R.P., da Motta Reis, J.S., Silva, D.E.W., de Barros, J.G.M., Sampaio, N.A.S. (2023), 'How to perform a simultaneous optimization with several response variables', \emph{Management and Administrative Professional Review} \strong{14} (1), pp. 564-578, DOI: \url{http://dx.doi.org/10.7769/gesec.v14i1.1536}.

Cojocaru, C., Khayet, M., Zakrzewska-Trznadel, G., Jaworska, A. (2009), 'Modeling and multi-response optimization of pervaporation of organic aqueous solutions using desirability function approach', \emph{Journal of Hazardous Materials} \strong{167}, pp. 52-63, DOI: \url{http://dx.doi.org/10.1016/j.jhazmat.2008.12.078}.

Derringer, G., Suich, R. (1980), 'Simultaneous Optimization of Several Response Variables', \emph{Journal of Quality Technology} \strong{12} (4), pp. 214-219, DOI: \url{https://doi.org/10.1080/00224065.1980.11980968}.
}
\seealso{
\code{\link{doe_analyze}}, \code{\link{get_meq}}
}
